% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{active}
\alias{active}
\alias{active<-}
\alias{active<-.ggheatmap}
\alias{active<-.ggAnnotationList}
\title{Determine the context of subsequent manipulations}
\usage{
active(x)

active(x) <- value

\method{active}{ggheatmap}(x) <- value

\method{active}{ggAnnotationList}(x) <- value
}
\arguments{
\item{x}{A \link[=ggheat]{ggheatmap} or \code{ggAnnotationList}object.}

\item{value}{What should get activated? Possible values are \code{"top"},
\code{"left"}, \code{"bottom"}, and \code{"right"}.}
}
\value{
\itemize{
\item \code{active}: A string of current active context.
}

\itemize{
\item \verb{active<-}: The same with \link{activate}, but won't check the arguments.
}
}
\description{
Determine the context of subsequent manipulations
}
