% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-.R
\docType{data}
\name{align_discrete}
\alias{align_discrete}
\alias{AlignDiscrete}
\title{Create a New \code{align} Object}
\usage{
align_discrete(
  align,
  data,
  ...,
  params = list(),
  plot = NULL,
  size = NULL,
  schemes = NULL,
  no_axes = NULL,
  active = NULL,
  check.param = TRUE,
  call = caller_call()
)
}
\arguments{
\item{align}{An \code{AlignDiscrete} object.}

\item{data}{A flexible input that specifies the data to be used
\itemize{
\item \code{NULL}: No data is set.
\item \code{\link[ggplot2:waiver]{waiver()}}: Try to use the layout matrix.
\item A \code{function} (including purrr-like lambda syntax) that is applied to the
layout matrix. If you want to transform the final plot data, please use
\code{\link[=scheme_data]{scheme_data()}}.
\item A \code{matrix}, \verb{data frame}, or atomic vector.
}}

\item{...}{Additional fields passed to the \code{align} object.}

\item{params}{A list of parameters for \code{align}.}

\item{plot}{A ggplot object.}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit}}.}

\item{schemes}{Options for \code{schemes}:
\itemize{
\item \code{NULL}: Used when \code{align} do not add a plot.
\item \code{\link[ggplot2:waiver]{waiver()}}: Try to infer \code{schemes} based on \code{data}.
}}

\item{no_axes}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Logical; if \code{TRUE},
removes axes elements for the alignment axis using \code{\link[=theme_no_axes]{theme_no_axes()}}. By
default, will use the option-
\code{"ggalign.align_no_axes"}.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}

\item{check.param}{Logical; if \code{TRUE}, checks parameters and provides
warnings as necessary.}

\item{call}{The \code{call} used to construct the \code{Align} object, for reporting
messages.}
}
\value{
A new \code{ggalign_plot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

An \code{AlignDiscrete} object interacts with the \code{Layout} object to reorder or
split observations and, in some cases, add plot components to the \code{Layout}.
}
\details{
Each of the \verb{Align*} objects is just a \code{\link[ggplot2:ggproto]{ggproto()}}
object, descended from the top-level \code{AlignDiscrete}, and each implements
various methods and fields.

To create a new type of \verb{Align*} object, you typically will want to
override one or more of the following:
\itemize{
\item \code{setup_params}: Prepare parameter or check parameters used by this plot.
\item \code{setup_data}: Prepare data used by this plot.
\item \code{compute}: A method used to compute statistics.
\item \code{align}: A method used to group observations into panel or reorder
observations.
\item \code{draw}: A method used to draw the plot. Must return a \code{ggplot} object.
}
}
\section{Discrete Axis Alignment}{

It is important to note that we consider rows as observations, meaning
\code{vec_size(data)}/\code{NROW(data)} must match the number of observations along the
axis used for alignment (x-axis for a vertical stack layout, y-axis for a
horizontal stack layout).
}

\examples{
align_dendro()
}
\keyword{datasets}
\keyword{internal}
