% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-gg.R
\name{align_gg}
\alias{align_gg}
\alias{ggalign}
\title{Create ggplot object in the layout}
\usage{
align_gg(
  mapping = aes(),
  data = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)

ggalign(
  mapping = aes(),
  data = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)
}
\arguments{
\item{mapping}{Additional default list of aesthetic mappings to use for plot.}

\item{data}{A matrix, a data frame, or even a simple vector that will be
converted into a one-column matrix. If the \code{data} argument is set to \code{NULL},
the \verb{align_*} will use the \code{layout} data. Additionally, the \code{data} argument
can also accept a function (purrr-like lambda is also okay), which will be
applied with the \code{layout} data,

It is important to note that all \verb{align_*} functions consider the \code{rows} as
the observations. It means the \code{NROW(data)} must return the same number with
the parallel \code{layout} axis.
\itemize{
\item \code{layout_heatmap}: for column annotation, the \code{layout} data will be
transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix). This is necessary because column annotation uses
heatmap columns as observations, but we need rows.
\item \code{layout_stack}: the \code{layout} data will be used as it is since we place all
plots along a single axis.
}}

\item{size}{Plot size, can be a \link[grid:unit]{unit} object.}

\item{labels}{Labels for axis parallelly with the layout, the default will
use the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the layout
axis). Note the labels will be reordered based on the layout.
\item An expression vector (must be the same length as layout axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()}: if \code{labels} is \code{NULL}, then \code{labels_nudge} will be \code{NULL},
otherwise it will inherit from the \code{layout} object.
\item A numeric.
}}

\item{set_context}{A single boolean value indicates whether to set the active
context to current \code{Align} object. If \code{TRUE}, all subsequent ggplot elements
will be added into this \code{Align} object.}

\item{order}{An single integer for the layout order.}

\item{name}{A string of the object name.}
}
\value{
A \code{AlignGG} object.
}
\description{
\code{ggalign} is just an alias of \code{align_gg}.
}
\section{ggplot2 details}{

\code{align_gg} initializes a \code{ggplot} data and \code{mapping}.

The internal will always use a default mapping of \code{aes(y = .data$.y)} or
\code{aes(x = .data$.x)}.

For ggplot usage, matrix (including a simple vector) data is converted into a
long-format data frame, similar to the process utilized in \code{ggheatmap}. But
note that the long-format data frame does not contain \code{.row_panel} or
\code{.column_panel} column, as \code{align_gg} can only have one facet axis. In the
case where the input data is already a data frame, three additional
columns-(\code{.row_names}, \code{.row_index}, and \code{.panel})—are added to the data
frame.

The data in the underlying \code{ggplot} object contains following columns:
\itemize{
\item \code{.panel}: the panel for current annotation
\item \code{.row_names} and \code{.row_index}: the row names and row index of the original
matrix or data frame.
\item \code{.column_names} and \code{.column_index}: the row and column index of the
original matrix (only applicable if \code{data} is a \code{matrix}).
\item \code{.x} or \code{.y}: the \code{x} or \code{y} coordinates
\item \code{value}: the actual matrix value  (only applicable if \code{data} is a
\code{matrix}).
}
}

\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    ggalign() +
    geom_point(aes(y = value))
}
