% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-ggplot.R
\name{align_ggplot}
\alias{align_ggplot}
\alias{gplot}
\title{Add ggplot to layout}
\usage{
align_ggplot(
  plot,
  size = NULL,
  action = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  ...
)

gplot(
  plot,
  size = NULL,
  action = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  ...
)
}
\arguments{
\item{plot}{A \code{\link[ggplot2:ggplot]{ggplot}} object, you can use \code{\link[=ggwrap]{ggwrap()}} to
convert any graphic into ggplot.}

\item{size}{Plot size, can be an \link[grid:unit]{unit} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current plot. If \code{TRUE}, all subsequent ggplot elements will be
added into this plot.}

\item{order}{An single integer for the plot area order.}

\item{name}{A string of the plot name. Used to switch the active context in
\code{\link[=hmanno]{hmanno()}} or \code{\link[=stack_active]{stack_active()}}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{AlignGgplot} object.
}
\description{
Add ggplot to layout
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    gplot(ggplot(mtcars, aes(wt, mpg))) +
    geom_point()
}
