% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-link.R, R/align-range.R
\name{align_line}
\alias{align_line}
\alias{align_range}
\title{Add a plot to annotate selected observations}
\usage{
align_line(
  data = waiver(),
  mapping = aes(),
  lines = waiver(),
  position = waiver(),
  size = NULL,
  active = NULL
)

align_range(
  data = waiver(),
  mapping = aes(),
  ranges = waiver(),
  position = waiver(),
  size = NULL,
  active = NULL
)
}
\arguments{
\item{data}{The following options can be used:
\itemize{
\item \code{NULL}: No data is set.
\item \code{\link[ggplot2:waiver]{waiver()}}: Inherits the data from the layout matrix.
\item A \code{function} (including purrr-like lambda syntax): Applied to the layout
matrix to transform the data before use. To transform the final plot
data, please use \code{\link[=scheme_data]{scheme_data()}}.
\item A \code{matrix}, \code{data.frame}, or atomic vector.
}}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{lines}{A list of observations. Each element of the list will be
represented by a facet panel.}

\item{position}{Which side the link should be added to? A string containing
one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}. For a horizontal \code{\link[=stack_layout]{stack_layout()}},
only \code{l} (left) and \code{r} (right) can be used. For a vertical
\code{\link[=stack_layout]{stack_layout()}}, only \code{b} (bottom) and \code{t} (top) are available.
\itemize{
\item Link lines can be customized using the \code{plot.ggalign_lines} theme element
with \code{\link[=element_line]{element_line()}}.
\item Link ranges can be customized using the \code{plot.ggalign_ranges} theme element
with \code{\link[=element_polygon]{element_polygon()}}.
}}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit}}. Note that for \code{\link[=circle_layout]{circle_layout()}}, all size values will
be interpreted as relative sizes, as this layout type adjusts based on the
available space in the circular arrangement.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}

\item{ranges}{A list of observation ranges. Each range will be represented
by a facet panel.}
}
\description{
\itemize{
\item \code{align_line}: Annotate a list of spread observations. Observations will be
connected to the panel by a line.
\item \code{align_range}: Annotate a list of ranges of observations. Observation
ranges will be connected to the panel by a polygon.
}
}
\section{ggplot2 specification}{

\code{align_ranges} initializes a ggplot.

The data in the underlying \code{ggplot} object will contain following columns:
\itemize{
\item \code{.panel}: the panel for the aligned axis. It means \code{x-axis} for vertical
stack layout (including top and bottom annotation), \code{y-axis} for
horizontal stack layout (including left and right annotation).
\item \code{.names} (\code{\link[vctrs:vec_names]{vec_names()}}) and \code{.index}
(\code{\link[vctrs:vec_size]{vec_size()}}/\code{\link[=NROW]{NROW()}}): a character names (only
applicable when names exists) and an integer of index of the original
data.
\item \code{.row_names} and \code{.row_index}: the row names and an integer of
row index of the original matrix (only applicable if \code{data} is a
\code{matrix}).
\item \code{.column_names} and \code{.column_index}: the column names and column index of
the original matrix (only applicable if \code{data} is a \code{matrix}).
\item \code{value}: the actual value (only applicable if \code{data} is a \code{matrix} or
atomic vector).
}

\code{matrix} input will be automatically melted into a long foramted data frame.

Atomic vector will be put in the \code{value} column of the data frame.

In the case where the input data is already a data frame, following columns
(\code{.panel}, \code{.index}, \code{.names}) are added to the data frame.
}

