% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-panel.R
\name{align_panel}
\alias{align_panel}
\alias{ggpanel}
\title{Create ggplot object with layout panel data}
\usage{
align_panel(
  mapping = aes(),
  size = NULL,
  action = NULL,
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated()
)

ggpanel(
  mapping = aes(),
  size = NULL,
  action = NULL,
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated()
)
}
\arguments{
\item{mapping}{Additional default list of aesthetic mappings to use for plot.}

\item{size}{Plot size, can be an \link[grid:unit]{unit} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot.}

\item{limits}{A boolean value indicates whether to set the layout limtis for
the plot.}

\item{facet}{A boolean value indicates whether to set the layout facet for
the plot. If this is \code{FALSE}, \code{limits} will always be \code{FALSE} too.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current plot. If \code{TRUE}, all subsequent ggplot elements will be
added into this plot.}

\item{order}{An single integer for the plot area order.}

\item{name}{A string of the plot name. Used to switch the active context in
\code{\link[=hmanno]{hmanno()}} or \code{\link[=stack_active]{stack_active()}}.}

\item{free_guides}{Override the \code{guides} collection behavior specified in the
heatmap layout for the annotation stack layout. Only used when \code{position} is
a string.}

\item{free_spaces}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{plot_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{theme}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_labs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}
}
\value{
A \code{AlignPanel} object.
}
\description{
This is similar with \code{ggalign()} function, but it will always use the layout
panel data. \code{ggpanel} is just an alias of \code{align_panel}.
}
\section{ggplot2 specification}{

\code{align_panel} initializes a ggplot \code{data} and \code{mapping}.

\code{align_panel()} always applies a default mapping for the axis of the data
index in the layout. This mapping is \code{aes(y = .data$.y)} for horizontal stack
layout (including left and right heatmap annotation) and \code{aes(x = .data$.x)}
for vertical stack layout (including top and bottom heatmap annotation).

The data in the underlying \code{ggplot} object contains following columns:
\itemize{
\item \code{.panel}: the panel for current layout axis. It means \code{x-axis} for
vertical stack layout, \code{y-axis} for horizontal stack layout.
\item \code{.index}: the index of the original layout data.
\item \code{.x} or \code{.y}: the \code{x} or \code{y} coordinates
}
}

\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    ggalign() +
    geom_point(aes(y = value))
}
