% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-phylo.R
\name{align_phylo}
\alias{align_phylo}
\title{Plot Phylogenetics tree}
\usage{
align_phylo(
  phylo,
  ...,
  ladderize = TRUE,
  no_axes = NULL,
  active = NULL,
  size = NULL
)
}
\arguments{
\item{phylo}{A \code{\link[ape:as.phylo]{phylo}} object.}

\item{...}{<\link[rlang:dyn-dots]{dyn-dots}> Additional arguments passed to
\code{\link[ggplot2:geom_segment]{geom_segment()}}.}

\item{ladderize}{A boolean value indicates whether to ladderize the tree. See
\code{\link[ape:ladderize]{ladderize()}}.}

\item{no_axes}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Logical; if \code{TRUE},
removes axes elements for the alignment axis using \code{\link[=theme_no_axes]{theme_no_axes()}}. By
default, will use the option-
\code{"ggalign.align_no_axes"}.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit()}}. Note that for \code{\link[=circle_layout]{circle_layout()}}, all size values
will be interpreted as relative sizes, as this layout type adjusts based on
the available space in the circular arrangement.}
}
\description{
Plot Phylogenetics tree
}
