% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-ranges.R
\name{align_ranges}
\alias{align_ranges}
\title{Add a plot to annotate a series of ranges of observations}
\usage{
align_ranges(
  data = waiver(),
  mapping = aes(),
  ranges = waiver(),
  position = waiver(),
  size = NULL,
  active = NULL
)
}
\arguments{
\item{data}{A flexible input that specifies the data to be used
\itemize{
\item \code{NULL}: No data is set.
\item \code{\link[ggplot2:waiver]{waiver()}}: Uses the layout matrix.
\item A \code{function} (including purrr-like lambda syntax) that is applied to the
layout matrix, and must return a matrix. If you want to transform the final
plot data, please use \code{\link[=scheme_data]{scheme_data()}}.
\item A \code{matrix}, \verb{data frame}, or atomic vector.
}}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{ranges}{A list of ranges to be annotated. Each range will be
represented by a facet panal.}

\item{position}{Which side the link should be added to? A string containing
one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}. For a horizontal \code{\link[=stack_layout]{stack_layout()}},
only \code{l} (left) and \code{r} (right) can be used. For a vertical
\code{\link[=stack_layout]{stack_layout()}}, only \code{b} (bottom) and \code{t} (top) are available.}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit}}.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}
}
\description{
Add a plot to annotate a series of ranges of observations
}
\section{ggplot2 specification}{

\code{align_ranges} initializes a ggplot.

The data in the underlying \code{ggplot} object will contain following columns:
\itemize{
\item \code{.panel}: the panel for the aligned axis. It means \code{x-axis} for vertical
stack layout (including top and bottom annotation), \code{y-axis} for
horizontal stack layout (including left and right annotation).
\item \code{.names} (\code{\link[vctrs:vec_names]{vec_names()}}) and \code{.index}
(\code{\link[vctrs:vec_size]{vec_size()}}/\code{\link[=NROW]{NROW()}}): a character names (only
applicable when names exists) and an integer of index of the original
data.
\item \code{.row_names} and \code{.row_index}: the row names and an integer of
row index of the original matrix (only applicable if \code{data} is a
\code{matrix}).
\item \code{.column_names} and \code{.column_index}: the column names and column index of
the original matrix (only applicable if \code{data} is a \code{matrix}).
\item \code{value}: the actual value (only applicable if \code{data} is a \code{matrix} or
atomic vector).
}

\code{matrix} input will be automatically melted into a long foramted data frame.

Atomic vector will be put in the \code{value} column of the data frame.

In the case where the input data is already a data frame, following columns
(\code{.panel}, \code{.index}, \code{.names}) are added to the data frame.
}

