% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-title.R
\name{align_title}
\alias{align_title}
\title{Draw Heatmap rows/columns titles}
\usage{
align_title(
  titles = NULL,
  ...,
  mapping = aes(),
  size = unit(1, "cm"),
  set_context = TRUE,
  order = NULL,
  name = NULL
)
}
\arguments{
\item{titles}{A character of titles, must be of length with the same number
of heatmap rows/columns panels.}

\item{...}{Additional arguments passed to \link[ggplot2:geom_text]{geom_text}.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{size}{Plot size, can be a \link[grid:unit]{unit} object.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current \code{Align} object. If \code{TRUE}, all subsequent ggplot elements
will be added into this \code{Align} object.}

\item{order}{An single integer for the layout order.}

\item{name}{A string of the object name.}
}
\value{
A new \code{Align} object.
}
\description{
Draw Heatmap rows/columns titles
}
\examples{
small_mat <- matrix(rnorm(81), nrow = 9)
ggheatmap(small_mat) +
    heatmap_active("top") +
    align_group(
        sample(letters[1:4], ncol(small_mat), replace = TRUE)
    ) +
    align_title()
}
