% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-wrap.R
\name{align_wrap}
\alias{align_wrap}
\alias{ggwrap}
\title{Wrap Arbitrary Graphics to layout}
\usage{
align_wrap(
  plot,
  ...,
  size = NULL,
  action = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)

ggwrap(
  plot,
  ...,
  size = NULL,
  action = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)
}
\arguments{
\item{plot}{A \link[ggplot2:ggplot]{ggplot} object or any graphic that can be
converted into a \link[ggplot2:ggplot]{ggplot} using \code{\link[=wrap]{wrap()}}.}

\item{...}{Additional arguments passed on to \code{\link[=wrap]{wrap()}}, only used if \code{plot} is
not a \link[ggplot2:ggplot]{ggplot} object.}

\item{size}{Plot size, can be an \link[grid:unit]{unit} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current plot. If \code{TRUE}, all subsequent ggplot elements will be
added into this plot.}

\item{order}{An single integer for the plot area order.}

\item{name}{A string of the plot name. Used to switch the active context in
\code{\link[=hmanno]{hmanno()}} or \code{\link[=stack_active]{stack_active()}}.}
}
\value{
A \code{AlignWrap} object.
}
\description{
Wrap Arbitrary Graphics to layout
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    ggwrap(ggplot(mtcars, aes(wt, mpg))) +
    geom_point()
}
