% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{alignwrap}
\alias{alignwrap}
\title{Wrap Arbitrary Graphics for Alignment}
\usage{
alignwrap(plot, ..., align = NULL, clip = TRUE)
}
\arguments{
\item{plot}{Any graphic that can be converted into a \link[grid:grid.grob]{grob} using
\code{\link[=wrap]{wrap()}}.}

\item{...}{Additional arguments passed to the \code{\link[=wrap]{wrap()}} method.}

\item{align}{A string specifying the area to place the plot: \code{"full"} for the
full area, \code{"plot"} for the full plotting area (including the axis label), or
\code{"panel"} for only the actual area where data is drawn.}

\item{clip}{A single boolean value indicating whether the grob should be
clipped if they expand outside their designated area.}
}
\value{
A \code{wrapped_plot} object that can be directly placed into
\code{\link[=align_plots]{align_plots()}}.
}
\description{
The \code{alignwrap()} function allows non-ggplot2 elements to be converted into a
compliant representation for use with \code{\link[=align_plots]{align_plots()}}. This is useful for
adding any graphics that can be converted into a \link[grid:grid.grob]{grob} with the
\code{\link[=wrap]{wrap()}} method.
}
