% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-.R
\name{anno}
\alias{anno}
\alias{show,anno-method}
\title{Create a heatmap annotation object}
\usage{
anno(
  Class,
  ...,
  data = NULL,
  position = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = NULL,
  order = NULL,
  name = NULL,
  call = caller_call()
)

\S4method{show}{anno}(object)
}
\arguments{
\item{Class}{Sub-class of \code{anno}.}

\item{...}{Additional components of the sub-class.}

\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{labels}{Labels for axis parallelly with heatmap, the default will use
the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the heatmap
axis)
\item An expression vector (must be the same length as heatmap axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. If \code{waiver()},
will inherit from the heatmap. If \code{NULL}, no breaks, in this way labels will
be removed too.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{call}{The \code{call} used to construct the scale for reporting messages.}

\item{object}{A \code{anno} object.}
}
\value{
A new \code{Class} object.
}
\description{
Create a heatmap annotation object
}
