% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-patchwork.R,
%   R/alignpatch-plot_grid.R
\name{as_alignpatches.patchwork}
\alias{as_alignpatches.patchwork}
\alias{as_alignpatches}
\title{Convert into an \code{alignpatches} object}
\usage{
\method{as_alignpatches}{patchwork}(x)

as_alignpatches(x)
}
\arguments{
\item{x}{An object that implements the method.}
}
\description{
Convert into an \code{alignpatches} object
}
