% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-grob.R
\name{as_grob.ggalign::alignpatches}
\alias{as_grob.ggalign::alignpatches}
\title{Convert Object into a Grob}
\usage{
\method{as_grob}{`ggalign::alignpatches`}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Not used currently.}
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
Convert Object into a Grob
}
\seealso{
\code{\link[=align_plots]{align_plots()}}

Other as_grob: 
\code{\link{as_grob.Heatmap}()},
\code{\link{as_grob.formula}()},
\code{\link{as_grob.ggplot}()},
\code{\link{as_grob.grob}()},
\code{\link{as_grob.patch}()},
\code{\link{as_grob.patch_ggplot}()},
\code{\link{as_grob.patchwork}()},
\code{\link{as_grob.pheatmap}()},
\code{\link{as_grob.recordedplot}()},
\code{\link{as_grob.trellis}()}
}
\concept{as_grob}
