% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-grob.R
\name{as_grob.patch_ggplot}
\alias{as_grob.patch_ggplot}
\title{Convert Object into a Grob}
\usage{
\method{as_grob}{patch_ggplot}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Not used currently.}
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
Convert Object into a Grob
}
\seealso{
\itemize{
\item \code{\link[=patch_titles]{patch_titles()}}
\item \code{\link[=inset]{inset()}}
\item \code{\link[=ggwrap]{ggwrap()}}
}

Other as_grob: 
\code{\link{as_grob.Heatmap}()},
\code{\link{as_grob.formula}()},
\code{\link{as_grob.ggalign::alignpatches}()},
\code{\link{as_grob.ggplot}()},
\code{\link{as_grob.grob}()},
\code{\link{as_grob.patch}()},
\code{\link{as_grob.patchwork}()},
\code{\link{as_grob.pheatmap}()},
\code{\link{as_grob.recordedplot}()},
\code{\link{as_grob.trellis}()}
}
\concept{as_grob}
