% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-grob.R
\name{as_grob.recordedplot}
\alias{as_grob.recordedplot}
\title{Convert Object into a Grob}
\usage{
\method{as_grob}{recordedplot}(x, ..., device = NULL)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{device}{
    A function that opens a graphics device for \code{grid.echo()} to
    work on.  By default this is an off-screen, in-memory device
    based on the \code{pdf} device.  This default device may not be
    satisfactory when using custom fonts.
  }
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
Convert Object into a Grob
}
\section{\code{as_grob} method collections}{

\itemize{
\item \code{\link[=as_grob.grob]{as_grob.grob()}}
\item \code{\link[=as_grob.gList]{as_grob.gList()}}
\item \code{\link[=as_grob.patch_ggplot]{as_grob.patch_ggplot()}}
\item \code{\link[=as_grob.ggplot]{as_grob.ggplot()}}
\item \code{\link[=as_grob.ggalign::alignpatches]{as_grob.ggalign::alignpatches()}}
\item \code{\link[=as_grob.patchwork]{as_grob.patchwork()}}
\item \code{\link[=as_grob.patch]{as_grob.patch()}}
\item \code{\link[=as_grob.formula]{as_grob.formula()}}
\item \code{\link[=as_grob.recordedplot]{as_grob.recordedplot()}}
\item \code{\link[=as_grob.trellis]{as_grob.trellis()}}
\item \code{\link[=as_grob.Heatmap]{as_grob.Heatmap()}}
\item \code{\link[=as_grob.pheatmap]{as_grob.pheatmap()}}
}
}

\seealso{
\code{\link[grDevices:recordplot]{recordPlot()}}

Other as_grob: 
\code{\link{as_grob.Heatmap}()},
\code{\link{as_grob.formula}()},
\code{\link{as_grob.gList}()},
\code{\link{as_grob.ggalign::alignpatches}()},
\code{\link{as_grob.ggplot}()},
\code{\link{as_grob.grob}()},
\code{\link{as_grob.patch}()},
\code{\link{as_grob.patch_ggplot}()},
\code{\link{as_grob.patchwork}()},
\code{\link{as_grob.pheatmap}()},
\code{\link{as_grob.trellis}()}
}
\concept{as_grob}
