% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{as_patch}
\alias{as_patch}
\alias{patch_gtable}
\alias{patch_align}
\title{Extends plots to be aligned with \code{plot_grid}}
\usage{
as_patch(x)

patch_gtable(patch)

patch_align(gt, guides)
}
\value{
\itemize{
\item \code{as_patch}: Any objects implements \code{patch_table} method.
}

\itemize{
\item \code{patch_gtable}: A \verb{gtable_*} object
}

\itemize{
\item \code{patch_align}: A \verb{align_*} object
}
}
\description{
Plots implement \code{as_patch} method must implement following methods to control
the layout details.
\itemize{
\item \code{as_patch}: Convert a plot object into a \code{patch} to be aligned, the \code{patch}
must implement proper \code{patch_table} method. \code{ggalign} has implement
\code{patch_table} method for \link[ggplot2:ggplot]{ggplot},
\link[grid:grid.grob]{grob}, and \link[=plot_grid]{alignpatches} objects.
\item \code{patch_table}: Convert the plot into a \link[gtable:gtable]{gtable}.
\item \code{patch_align}: Build a standard \link[gtable:gtable]{gtable} object.
}
}
