% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft-boxes.R
\name{BoxOrientation}
\alias{BoxOrientation}
\title{box_orientation Class}
\format{
An S7 class with the following properties:
\itemize{
\item axis: A character string indicating the alignment axis of the box, either
"x" or "y".
\item is_linear: A logical value indicating whether the box should be arranged
in a linear fashion.
\item direction: A character string specifying the direction of the box layout.
Possible values are "horizontal" or "vertical". The default is \code{NA},
indicating no direction specified.
\item is_horizontal: A logical getter indicating whether the direction is
horizontal. Returns \code{TRUE} if \code{direction} is set to "horizontal",
otherwise \code{FALSE}.
\item is_vertical: A logical getter indicating whether the direction is
vertical. Returns \code{TRUE} if \code{direction} is set to "vertical", otherwise
\code{FALSE}.
\item position: A character string that determines the position of the box.
Valid values come from the \code{.TLBR} set (Top, Left, Bottom, Right). The
default is \code{NA}, indicating no position.
\item is_annotation: A logical getter that checks if the box is an annotation.
Returns \code{TRUE} if \code{position} is \code{NA} (i.e., no position is specified).
\item inherit_quad_matrix: A logical value indicating whether the box inherits
a quad matrix layout. The default is \code{FALSE}.
}
}
\usage{
BoxOrientation(
  axis = character(0),
  is_linear = logical(0),
  direction = NA_character_,
  position = NA_character_,
  inherit_quad_matrix = FALSE
)
}
\description{
The \strong{\code{box_orientation}} class defines the intrinsic context for aligning
CraftBox objects. It is used to specify layout properties like alignment,
direction, and positioning of boxes within a layout. These properties are
essential for controlling the placement and alignment of the CraftBox
elements.
}
