% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_matrix}
\alias{build_matrix}
\title{Build matrix for heatmap and anno_cluster}
\usage{
build_matrix(matrix, arg = rlang::caller_arg(matrix))
}
\arguments{
\item{matrix}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}
}
\description{
Build matrix for heatmap and anno_cluster
}
