% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-methods.R
\name{build_patchwork}
\alias{build_patchwork}
\title{Build \code{Layout} object for rendering.}
\usage{
build_patchwork(layout)
}
\arguments{
\item{layout}{A \code{\link[=layout_heatmap]{layout_heatmap()}} or \code{\link[=layout_stack]{layout_stack()}} object.}
}
\value{
A \code{patchwork} object or \code{NULL} if no plots.
}
\description{
Build \code{Layout} object for rendering.
}
\examples{
build_patchwork(ggheatmap(matrix(rnorm(100L), nrow = 10L)))
}
