% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_tibble}
\alias{build_tibble}
\title{Build data.frame for annotation.}
\usage{
build_tibble(data, arg = rlang::caller_arg(data))
}
\arguments{
\item{data}{A data.frame, if it is a simple vector, it will be converted to a
one-column data.frame with "V1" as column name. Matrix will also be coerced
into data.frame.}
}
\description{
Build data.frame for annotation.
}
