% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{$,ggheatmap-method}
\alias{$,ggheatmap-method}
\title{Subset a \code{ggheatmap} object}
\usage{
\S4method{$}{ggheatmap}(x, name)
}
\arguments{
\item{x}{A ggheatmap object}

\item{name}{A string of slot name in \code{ggheatmap} object.}
}
\description{
Used by \link[ggplot2:ggplot_build]{ggplot_build}
}
\keyword{internal}
