% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{$,HeatmapLayout-method}
\alias{$,HeatmapLayout-method}
\title{Subset a \code{HeatmapLayout} object}
\usage{
\S4method{$}{HeatmapLayout}(x, name)
}
\arguments{
\item{x}{A \code{HeatmapLayout} object}

\item{name}{A string of slot name in \code{HeatmapLayout} object.}
}
\description{
Used by \link[ggplot2:ggplot_build]{ggplot_build} and \link[ggplot2:ggsave]{ggsave}
}
\keyword{internal}
