% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{$,Layout-method}
\alias{$,Layout-method}
\title{Subset a \code{Layout} object}
\usage{
\S4method{$}{Layout}(x, name)
}
\arguments{
\item{x}{A \code{Layout} object}

\item{name}{A string of slot name in \code{Layout} object.}
}
\description{
Used by \link[ggplot2:ggplot_build]{ggplot_build} and \link[ggplot2:ggsave]{ggsave}
}
\keyword{internal}
