% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{$,LayoutProto-method}
\alias{$,LayoutProto-method}
\title{Subset a \code{Layout} object}
\usage{
\S4method{$}{LayoutProto}(x, name)
}
\arguments{
\item{x}{A \code{Layout} object}

\item{name}{A string of slot name in \code{Layout} object.}
}
\value{
The slot value.
}
\description{
Used by \code{\link[ggplot2:build_ggplot]{ggplot_build}} and
\code{\link[ggplot2:ggsave]{ggsave}}
}
\keyword{internal}
