% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{$,LayoutStack-method}
\alias{$,LayoutStack-method}
\title{Subset a \code{LayoutStack} object}
\usage{
\S4method{$}{LayoutStack}(x, name)
}
\arguments{
\item{x}{A \code{LayoutStack} object}

\item{name}{A string of slot name in \code{LayoutStack} object.}
}
\description{
Used by \link[ggplot2:ggplot_build]{ggplot_build} and \link[ggplot2:ggsave]{ggsave}
}
\keyword{internal}
