% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{$,StackLayout-method}
\alias{$,StackLayout-method}
\title{Subset a \code{StackLayout} object}
\usage{
\S4method{$}{StackLayout}(x, name)
}
\arguments{
\item{x}{A \code{StackLayout} object}

\item{name}{A string of slot name in \code{StackLayout} object.}
}
\description{
Used by \link[ggplot2:ggplot_build]{ggplot_build} and \link[ggplot2:ggsave]{ggsave}
}
\keyword{internal}
