% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-circle-.R
\name{circle_discrete}
\alias{circle_discrete}
\title{Arrange Plots in a Circular Layout}
\usage{
circle_discrete(data = NULL, ..., radial = NULL, theme = NULL)
}
\arguments{
\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to
convert data to a matrix.}

\item{...}{Additional arguments passed to \code{\link[=fortify_matrix]{fortify_matrix()}}.}

\item{radial}{A \code{\link[ggplot2:coord_polar]{coord_radial()}} object that defines
the global parameters for \code{coord_radial} across all plots in the layout.
The parameters \code{start}, \code{end}, \code{direction}, and \code{expand} will be inherited
and applied uniformly to all plots within the layout. The parameters
\code{theta} and \code{r.axis.inside} will always be ignored and will be set to
\code{"x"} and \code{TRUE}, respectively, for all plots.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} object used to customize various
elements of the layout, including \code{guides}, \code{title}, \code{subtitle}, \code{caption},
\code{margins}, \code{panel.border}, and \code{background}. By default, the theme will
inherit from the parent \code{layout}. It also controls the panel spacing for all
plots in the layout.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{circle_discrete} function arranges plots by aligning discrete variables
in a circular layout.
}
\examples{
set.seed(123)
circle_discrete(matrix(rnorm(56), nrow = 7L),
    radial = coord_radial(inner.radius = 0.1)
) +
    ggalign() +
    geom_tile(aes(y = .column_index, fill = value)) +
    scale_fill_viridis_c() +
    align_dendro(aes(color = branch), k = 3L) +
    scale_color_brewer(palette = "Dark2")

}
