% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\title{Plot context settings}
\usage{
context(order = waiver(), active = waiver(), name = waiver())
}
\arguments{
\item{order}{An integer specifying the order of the plot area within the
layout.}

\item{active}{A logical (\code{TRUE}/\code{FALSE}) indicating whether to set the
active context to the current plot area when added to a layout. If \code{TRUE},
any subsequent \code{ggplot} elements will be applied to this active plot area.}

\item{name}{A string specifying the plot's name, useful for switching active
contexts through the \code{what} argument in functions like
\code{\link[=quad_anno]{quad_anno()}}/\code{\link[=stack_switch]{stack_switch()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} These settings control the arrangement
and behavior of individual plot areas within the layout. By default, the
active context is set only for functions that add plot areas. This approach
allows users to seamlessly add other \code{ggplot2} elements—such as \code{geoms},
\code{stats}, \code{scales}, or \code{themes}—to the current plot area. The default ordering
of plot areas is from top to bottom or from left to right, depending on
layout orientation. However, users can customize this order using the \code{order}
argument.
}
