% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-align.R
\name{continuous_limits}
\alias{continuous_limits}
\title{Set continuous limits for the layout}
\usage{
continuous_limits(..., x = waiver(), y = waiver())
}
\arguments{
\item{...}{A list of two numeric values, specifying the left/lower limit and
the right/upper limit of the scale.}

\item{x, y}{Same as \code{...}, but specifically for \code{quad_layout()}.}
}
\description{
To align continuous axes, it is important to keep the limits consistent
across all plots in the layout. You can set the limits by passing a function
directly to the \code{limits} or \code{xlim}/\code{ylim} argument, using \code{...} only.
Alternatively, you can add a \code{continuous_limits()} object to the layout. For
the \code{quad_layout()} function, you must specify \code{x}/\code{y} arguments. For other
layouts, you should pass the limits using \code{...} directly.
}
