% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-cross-.R
\name{cross_align}
\alias{cross_align}
\alias{cross_layout}
\alias{cross_alignv}
\alias{cross_alignh}
\title{Arrange Plots Crosswise Horizontally or Vertically}
\usage{
cross_align(data = NULL, direction, ..., theme = NULL, sizes = NA)

cross_alignv(data = NULL, ...)

cross_alignh(data = NULL, ...)
}
\arguments{
\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout:
\itemize{
\item For \code{stack_free}, \code{\link[=fortify_data_frame]{fortify_data_frame()}} will be used to convert data to
a data frame.
\item For \code{stack_align}, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to convert data to a
matrix.
}}

\item{direction}{A string indicating the direction of the stack layout,
either \code{"horizontal"} or \code{"vertical"}.}

\item{...}{Additional arguments passed to \code{\link[=fortify_data_frame]{fortify_data_frame()}} or
\code{\link[=fortify_matrix]{fortify_matrix()}}.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} object used to customize various
elements of the plot, including \code{guides}, \code{title}, \code{subtitle}, \code{caption},
\code{margins}, \code{patch.title}, \code{panel.border}, and \code{background}. By default, the
theme will inherit from the parent \code{layout}. When aligning observations, it
also controls the panel spacing for all plots in the layout.}

\item{sizes}{A numeric value or a \code{\link[grid:unit]{unit}} object. If used for
the \code{\link[=quad_layout]{quad_layout()}} annotation, it must have a length of \code{1}.  If used in
the layout with a nested \code{\link[=quad_layout]{quad_layout()}}, it should have a length of \code{3},
indicating the relative heights (for \code{direction = "horizontal"}) or widths
(for \code{direction = "vertical"}) to be applied.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{cross_align} function aligns observations, and allow different layout
ordering index in a single layout. \code{cross_layout} is an alias for
\code{cross_align}.

Two aliases are provided for convenience:
\itemize{
\item \code{cross_alignv}: A special case of \code{cross_align} that sets \code{direction = "vertical"}.
\item \code{cross_alignh}: A special case of \code{cross_align} that sets \code{direction = "horizontal"}.
}
}
\seealso{
\code{\link[=ggcross]{ggcross()}}
}
