% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-gg.R
\name{cross_gg}
\alias{cross_gg}
\alias{ggcross}
\title{Connect two layout crosswise}
\usage{
cross_gg(mapping = aes(), size = NULL, no_axes = NULL, active = NULL)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit}}.}

\item{no_axes}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Logical; if \code{TRUE},
removes axes elements for the alignment axis using \code{\link[=theme_no_axes]{theme_no_axes()}}. By
default, will use the option-
\code{"ggalign.align_no_axes"}.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}
}
\description{
\code{cross_gg} resets the layout ordering index of a \code{\link[=cross_align]{cross_align()}}. This
allows you to add other \verb{align_*} objects to define a new layout ordering
index. Any objects added after \code{cross_gg} will use this updated layout
ordering index. This feature is particularly useful for creating \code{tanglegram}
visualizations. \code{ggcross()} is an alias of \code{cross_gg()}.
}
\section{ggplot2 specification}{

\code{ggcross()} initializes a ggplot \code{data} and \code{mapping}.

\code{ggcross()} always applies a default mapping for the axis of the data index
in the layout. This mapping is \code{aes(y = .data$y)} for horizontal stack layout
(including left and right annotation) and \code{aes(x = .data$x)} for vertical
stack layout (including top and bottom annotation).

The data in the underlying \code{ggplot} object will contain following columns:
\itemize{
\item \code{.panel}: the panel for the aligned axis. It means \code{x-axis} for vertical
stack layout (including top and bottom annotation), \code{y-axis} for
horizontal stack layout (including left and right annotation).
\item \code{.index}: an integer of the data index.
\item \code{.names}: a character of data labels.
\item \code{.hand}: a factor indicates the index groups.
\item \code{.x} or \code{.y}: the \code{x} or \code{y} coordinates.
}

It is recommended to use \code{.x}/\code{.y} as the \code{x}/\code{y} mapping.
}

