% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-align-cross.R
\name{cross_link}
\alias{cross_link}
\title{Connect two layout crosswise}
\usage{
cross_link(mapping = aes(), size = NULL, no_axes = NULL, active = NULL)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{size}{The relative size of the plot, can be specified as a
\code{\link[grid:unit]{unit}}.}

\item{no_axes}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Logical; if \code{TRUE},
removes axes elements for the alignment axis using \code{\link[=theme_no_axes]{theme_no_axes()}}. By
default, will controled by the option-
\code{"ggalign.align_no_axes"}.}

\item{active}{A \code{\link[=active]{active()}} object that defines the context settings when
added to a layout.}
}
\description{
\code{cross_link} resets the layout ordering index of a \code{\link[=stack_layout]{stack_layout()}}. This
allows you to add other \verb{align_*} objects to define a new layout ordering
index. Any objects added after \code{cross_link} will use this updated layout
ordering index. This feature is particularly useful for creating \code{tanglegram}
visualizations.
}
