% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-none.R
\name{cross_none}
\alias{cross_none}
\title{Add a plot to connect selected observations}
\usage{
cross_none(data = waiver(), inherit_panel = NULL, inherit_nobs = NULL)
}
\arguments{
\item{data}{The dataset to use for the layout. By default,
\code{\link[=fortify_matrix]{fortify_matrix()}} will convert the data to a matrix. This argument
allows you to change the layout data. If not specified, the original data
will be used.}

\item{inherit_panel}{A boolean value indicating whether to inherit the
panel group information.}

\item{inherit_nobs}{A boolean value indicating whether to inherit the
number of observations.}
}
\description{
Add a plot to connect selected observations
}
