% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram_data}
\alias{dendrogram_data}
\title{Dengrogram x and y coordinates}
\usage{
dendrogram_data(
  tree,
  center = FALSE,
  type = "rectangle",
  leaf_pos = NULL,
  leaf_braches = NULL,
  branch_gap = NULL,
  root = NULL
)
}
\arguments{
\item{tree}{A \link[stats:hclust]{hclust} or a \link[stats:dendrogram]{dendrogram}
object.}

\item{center}{A boolean value. if \code{TRUE}, nodes are plotted centered with
respect to the leaves in the branch. Otherwise (default), plot them in the
middle of all direct child nodes.}

\item{type}{A string indicates the plot type, \code{"rectangle"} or \code{"triangle"}.}

\item{leaf_pos}{The x-coordinates of the leaf node. Must be the same length
of the number of observations in \code{tree}.}

\item{leaf_braches}{Branches of the leaf node. Must be the same length of the
number of observations in \code{tree}. Usually come from \link[stats:cutree]{cutree}.}

\item{root}{A length one string or numeric indicates the root branch.}
}
\value{
A list of 2 data.frame. One for node coordinates, another for edge
coordinates.
}
\description{
Dengrogram x and y coordinates
}
