% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-magick.R
\name{.raster_magick}
\alias{.raster_magick}
\title{Rasterize the input object}
\usage{
.raster_magick(x, magick = NULL, ...)
}
\arguments{
\item{x}{An object to rasterize, can be a \code{\link[grid:grid.grob]{grob()}},
\code{\link[ggplot2:layer]{layer()}}, \code{\link[ggplot2:ggplot]{ggplot()}}, or a list of such
objects.}

\item{magick}{A function (purrr-style formula is accepted) that takes an
\code{\link[magick:editing]{image_read()}} object as input and returns an object
compatible with \code{\link[grDevices:as.raster]{as.raster()}}. You can use any of
the \verb{image_*()} functions from the \strong{magick} package to process the raster
image.}

\item{...}{Not used currently.}
}
\value{
An object with the same class of the input.
}
\description{
An internal function designed to implement the functionality of
\code{raster_magick()}. It assumes the input arguments are valid and does not
perform any additional checks.
}
\keyword{internal}
