% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helper.R
\name{element_polygon}
\alias{element_polygon}
\title{Theme Polygon elements}
\usage{
element_polygon(
  fill = NULL,
  colour = NULL,
  linewidth = NULL,
  linetype = NULL,
  alpha = NULL,
  lineend = NULL,
  linejoin = NULL,
  linemitre = NULL,
  color = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{fill}{Fill colour. \code{fill_alpha()} can be used to set the transparency
of the fill.}

\item{colour, color}{Line/border colour. Color is an alias for colour.
\code{alpha()} can be used to set the transparency of the colour.}

\item{linewidth}{Line size in \code{mm}.}

\item{linetype}{Line type for lines. An integer (0:8), a name (blank, solid,
dashed, dotted, dotdash, longdash, twodash), or a string with an even number
(up to eight) of hexadecimal digits which give the lengths in consecutive
positions in the string.}

\item{alpha}{A transparency value between 0 (transparent) and 1 (opaque),
parallel to \code{fill}.}

\item{lineend}{Line end style, one of \code{"round"}, \code{"butt"} or \code{"square"}.}

\item{linejoin}{Line join style, one of \code{"round"}, \code{"mitre"} or \code{"bevel"}.}

\item{linemitre}{Line mitre limit (number greater than 1).}

\item{inherit.blank}{Should this element inherit the existence of an
\code{element_blank} among its parents? If \code{TRUE} the existence of
a blank element among its parents will cause this element to be blank as
well. If \code{FALSE} any blank parent element will be ignored when
calculating final element state.}
}
\value{
A \code{element_polygon} object
}
\description{
Draw polygon.
}
\seealso{
\code{\link[ggplot2:element]{element_rect}}
}
