% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data_frame.R
\name{fortify_data_frame.default}
\alias{fortify_data_frame.default}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{default}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
By default, it calls \code{\link[ggplot2:fortify]{fortify()}} to build the
data frame.
}
\seealso{
Other \code{\link[=fortify_data_frame]{fortify_data_frame()}} methods: 
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.matrix}()},
\code{\link{fortify_data_frame.numeric}()}
}
\concept{\code{\link[=fortify_data_frame]{fortify_data_frame()}} methods}
