% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data_frame.R
\name{fortify_data_frame.matrix}
\alias{fortify_data_frame.matrix}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{matrix}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
When \code{data} is a matrix, it will automatically be transformed into a
long-form data frame, where each row represents a unique combination of
matrix indices and their corresponding values. The resulting data frame will
contain the following columns:
\itemize{
\item \code{.row_names} and \code{.row_index}: the row names and an integer representing
the row index of the original matrix.
\item \code{.column_names} and \code{.column_index}: the column names and column index of
the original matrix.
\item \code{value}: the actual value.
}
}
\seealso{
Other fortify_data_frame methods: 
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.complex}()},
\code{\link{fortify_data_frame.default}()},
\code{\link{fortify_data_frame.logical}()},
\code{\link{fortify_data_frame.numeric}()}
}
\concept{fortify_data_frame methods}
