% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-phylo.R
\name{fortify_data_frame.phylo}
\alias{fortify_data_frame.phylo}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{phylo}(
  data,
  ...,
  type = "rectangle",
  center = FALSE,
  tree_type = NULL,
  tip_pos = NULL
)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}

\item{type}{A string indicates the plot type, \code{"rectangle"} or \code{"triangle"}.}

\item{center}{A boolean value. if \code{TRUE}, nodes are plotted centered with
respect to all leaves/tips in the branch. Otherwise (default), plot them in
the middle of the direct child nodes.}

\item{tree_type}{A single string, one of
\code{"phylogram"} or \code{"cladogram"}, indicating the type of tree.
Usually, you don't need to modify this.}

\item{tip_pos}{The x-coordinates of the tip. Must be the same length
of the number of tips in \code{tree}.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts various objects to a data frame.
}
\details{
A \verb{data frame} with the node coordinates:
\itemize{
\item \code{.index}: the original index in the tree for the the tip/node.
\item \code{label}: the tip/node label text.
\item \code{x} and \code{y}: x-axis and y-axis coordinates for the tip/node.
\item \code{tip}: A logical value indicates whether current node is a tip.
}
}
\section{ggalign attributes}{

\code{edge}: A \verb{data frame} for edge coordinates:
\itemize{
\item \code{x} and \code{y}: x-axis and y-axis coordinates for the start node of the edge.
\item \code{xend} and \code{yend}: the x-axis and y-axis coordinates of the terminal node
for edge.
}
}

\seealso{
Other fortify_data_frame methods: 
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.complex}()},
\code{\link{fortify_data_frame.logical}()},
\code{\link{fortify_data_frame.matrix}()},
\code{\link{fortify_data_frame.numeric}()}
}
