% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-data.R
\name{fortify_heatmap.matrix}
\alias{fortify_heatmap.matrix}
\title{Build data for the heatmap layout}
\usage{
\method{fortify_heatmap}{matrix}(data, ...)
}
\arguments{
\item{data}{Any objects to be plot with \code{\link[=ggheatmap]{ggheatmap()}}.}

\item{...}{Not used currently.}
}
\value{
A matrix used by \code{\link[=ggheatmap]{ggheatmap()}}.
}
\description{
Build data for the heatmap layout
}
