% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-data.R
\name{fortify_heatmap}
\alias{fortify_heatmap}
\title{Build data for the heatmap layout}
\usage{
fortify_heatmap(data, ...)
}
\arguments{
\item{data}{Any objects to be plot with \code{\link[=ggheatmap]{ggheatmap()}}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Build data for the heatmap layout
}
