% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix.default}
\alias{fortify_matrix.default}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{default}(data, ..., data_arg = NULL, call = NULL)
}
\arguments{
\item{data}{An object to be converted into a matrix.}

\item{...}{These dots are for future extensions and must be empty.}

\item{data_arg}{The argument name for \code{data}. Developers can use it to
improve messages. Not used by the user.}

\item{call}{The execution environment where \code{data} and other arguments for
the method are collected, e.g., \code{\link[rlang:stack]{caller_env()}}.
Developers can use it to improve messages. Not used by the user.}
}
\value{
A matrix.
}
\description{
By default, it calls \code{\link[=as.matrix]{as.matrix()}} to build a matrix.
}
\seealso{
Other fortify_matrix methods: 
\code{\link{fortify_matrix.GISTIC}()},
\code{\link{fortify_matrix.MAF}()},
\code{\link{fortify_matrix.list_upset}()},
\code{\link{fortify_matrix.matrix}()},
\code{\link{fortify_matrix.matrix_upset}()},
\code{\link{fortify_matrix.phylo}()}
}
\concept{fortify_matrix methods}
