% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix.matrix}
\alias{fortify_matrix.matrix}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{matrix}(data, ...)
}
\arguments{
\item{data}{Any objects to be converted to a matrix.}

\item{...}{Not used currently.}
}
\value{
A matrix containing the converted data.
}
\description{
This function converts various objects into a matrix format.
}
