% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-data.R
\name{fortify_stack}
\alias{fortify_stack}
\title{Build data for the stack layout}
\usage{
fortify_stack(data, ...)
}
\arguments{
\item{data}{Any objects to be plot with \code{\link[=ggstack]{ggstack()}}.}

\item{...}{Arguments passed to methods.}
}
\value{
A matrix or data.frame used by \code{\link[=ggstack]{ggstack()}}.
}
\description{
Build data for the stack layout
}
