% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-upset.R
\name{fortify_upset.list}
\alias{fortify_upset.list}
\title{Build a Matrix for UpSet plot}
\usage{
\method{fortify_upset}{list}(data, mode = "distinct", ...)
}
\arguments{
\item{data}{A list of sets.}

\item{mode}{A string of \code{"distinct"}, \code{"intersect"}, or \code{"union"}
indicates the mode to define the set intersections. Check
\url{https://jokergoo.github.io/ComplexHeatmap-reference/book/upset-plot.html#upset-mode}
for details.}

\item{...}{Not used.}
}
\value{
A matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts various objects into a matrix format suitable for
creating an UpSet plot. It always returns a matrix for a \code{horizontal} UpSet
plot.
}
\section{ggalign attributes}{

\itemize{
\item \code{intersection_sizes}: An integer vector indicating the size of each
intersection.
\item \code{set_sizes}: An integer vector indicating the size of each set.
}
}

\seealso{
Other fortify_upset methods: 
\code{\link{fortify_upset.matrix}()}
}
\concept{fortify_upset methods}
