% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-upset.R
\name{fortify_upset}
\alias{fortify_upset}
\title{Build a Matrix for UpSet plot}
\usage{
fortify_upset(data, mode = "distinct", ...)
}
\arguments{
\item{data}{An object to be converted to a matrix for UpSet  plot.}

\item{mode}{A string of \code{"distinct"}, \code{"intersect"}, or \code{"union"}
indicates the mode to define the set intersections. Check
\url{https://jokergoo.github.io/ComplexHeatmap-reference/book/upset-plot.html#upset-mode}
for details.}

\item{...}{Arguments passed to methods.}
}
\value{
A matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function converts various objects into a matrix format suitable for
creating an UpSet plot. It always returns a matrix for a \code{horizontal} UpSet
plot.
}
\seealso{
\itemize{
\item \code{\link[=fortify_upset.list]{fortify_upset.list()}}
\item \code{\link[=fortify_upset.matrix]{fortify_upset.matrix()}}
}
}
