% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-free-border.R
\name{free_border}
\alias{free_border}
\title{Free panel border from alignment}
\usage{
free_border(plot, borders = c("t", "l", "b", "r"))
}
\arguments{
\item{plot}{A \link[ggplot2:ggplot]{ggplot} or \link[=plot_grid]{alignpatches} object.}

\item{borders}{Which border shouldn't be aligned? Allowed values: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}
}
\description{
\link{plot_grid} will try to align every element of a plot. If we want to
compose plots without aligning the panel borders (but still align the
panels themselves), we can wrap the plot with \code{free_border}.
}
