% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-free-lab.R
\name{free_lab}
\alias{free_lab}
\title{Free the axis titles from alignment}
\usage{
free_lab(plot, labs = c("t", "l", "b", "r"))
}
\arguments{
\item{labs}{Axis labs to be free.}
}
\value{
A modified version of \code{plot} with a \code{free_lab} class
}
\description{
Free the axis titles from alignment
}
