% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-free-size.R
\name{free_size}
\alias{free_size}
\title{Removing the sizes when aligning}
\usage{
free_size(plot, ...)
}
\arguments{
\item{plot}{A \link[ggplot2:ggplot]{ggplot} or \link[=plot_grid]{alignpatches} object.}

\item{...}{What ggplot2 elements to remove?}
}
\description{
Removing the sizes when aligning
}
