% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_attr}
\alias{ggalign_attr}
\title{Get a field data from the \code{ggalign} attribute}
\usage{
ggalign_attr(x, field)
}
\arguments{
\item{x}{Input data for the function used to transform the layout data.}

\item{field}{A single string indicating which data to use. Typically, this
list of data is attached by the \code{\link[=fortify_heatmap]{fortify_heatmap()}} function (see the
\verb{ggalign attributes} section in the documentation). Check
\code{\link[=fortify_heatmap.MAF]{fortify_heatmap.MAF()}} for examples.}
}
\value{
The requested data field or \code{NULL} if not found.
}
\description{
When rendering the layout with \code{\link[=heatmap_layout]{heatmap_layout()}} or \code{\link[=stack_layout]{stack_layout()}} object, a special attribute is kept
in the data called \code{"ggalign"}, which can be used to pass additional
information for the input data. This function helps extract data from that
attribute. This is particularly useful in the \code{data} function for
transforming the parent layout data.
}
