% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_build}
\alias{ggalign_build}
\title{Build \code{Layout} object for rendering.}
\usage{
ggalign_build(layout)
}
\arguments{
\item{layout}{A \code{Layout} object.}
}
\value{
A \code{patchwork} object.
}
\description{
Build \code{Layout} object for rendering.
}
\examples{
ggalign_build(ggheatmap(matrix(rnorm(100L), nrow = 10L)))
}
