% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_data_set}
\alias{ggalign_data_set}
\title{Attach supplementary data and levels for ggalign}
\usage{
ggalign_data_set(.data, ..., .lvls = NULL)
}
\arguments{
\item{.data}{Input data for the layout.}

\item{...}{A list of data to be attached.}

\item{.lvls}{A character vector representing the attached levels.}
}
\description{
Attach supplementary data and levels for ggalign
}
\note{
Used by developers in \code{\link[=fortify_matrix]{fortify_matrix()}}, \code{\link[=fortify_data_frame]{fortify_data_frame()}},
and other related methods.
}
\seealso{
\code{\link[=ggalign_attr]{ggalign_attr()}}/\code{\link[=ggalign_lvls]{ggalign_lvls()}}
}
