% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalign_gtable}
\alias{ggalign_gtable}
\title{Convert a Built Object into a gtable}
\usage{
ggalign_gtable(x)
}
\arguments{
\item{x}{The object to be converted into a gtable.}
}
\value{
A \code{\link[gtable:gtable]{gtable}} object.
}
\description{
This generic function takes a built graphical object and converts it into a
\code{\link[gtable:gtable]{gtable}}.
}
