% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalign_inherit}
\alias{ggalign_inherit}
\title{Inherit Properties from Another Object}
\usage{
ggalign_inherit(x, object, ...)
}
\arguments{
\item{x}{The object that will inherit properties.}

\item{object}{The object from which \code{x} will inherit properties.}

\item{...}{Arguments passed to methods.}
}
\value{
The \code{x} object with properties inherited from \code{object}.
}
\description{
This generic function allows one object to inherit properties or
configurations from another object. It is used when an object needs to adopt
the state of another object without directly modifying it.
}
