% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalign_init}
\alias{ggalign_init}
\title{Initialize the Default Properties of an Object}
\usage{
ggalign_init(x)
}
\arguments{
\item{x}{An object that needs initialization.}
}
\value{
The initialized object, ready for use.
}
\description{
This helper function is used to initialize an object before use by setting
its properties or internal elements to their default values. It is
particularly useful when the object's internal defaults differ from those
required for proper usage, such as when internal defaults indicate values not
set by the user. The function ensures that the object is properly configured
for subsequent operations.
}
