% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_lvls_set}
\alias{ggalign_lvls_set}
\alias{ggalign_lvls_get}
\title{Set or Get the Attached Levels}
\usage{
ggalign_lvls_set(x, lvls)

ggalign_lvls_get(x)
}
\arguments{
\item{x}{Input data for the layout.}

\item{lvls}{A character vector representing the attached levels.}
}
\description{
\itemize{
\item \code{ggalign_lvls_set}: Attaches levels to the input, enabling the restoration
of the \code{value} column when transformed from a matrix to a data frame.
\item \code{ggalign_lvls_get}: Extracts previously attached levels during the
transformation process.
}
}
\details{
Attributes attached to the data are especially useful when the input data is
transformed in ways that limit access to the complete dataset. For example,
\code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}} might filter mutation data while adding attributes
that retain important context, such as the total number of observations, for
detailed or aggregated analyses. Additionally, it stores the levels of
\code{Variant_Classification} for further usage.
}
\seealso{
\code{\link[=ggalign_lvls]{ggalign_lvls()}}
}
