% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/layout-.R
\name{ggalign_stat}
\alias{ggalign_stat}
\alias{ggalign_stat.ggalign::QuadLayout}
\alias{ggalign_stat.ggalign::StackLayout}
\title{Get the statistics from the layout}
\usage{
ggalign_stat(x, ...)

\method{ggalign_stat}{`ggalign::QuadLayout`}(x, position, ...)

\method{ggalign_stat}{`ggalign::StackLayout`}(x, what, ...)
}
\arguments{
\item{x}{A \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object.}

\item{...}{Arguments passed to methods.}

\item{position}{A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.}

\item{what}{A single number or string of the plot elements in the stack
layout.}
}
\value{
The statistics
}
\description{
Get the statistics from the layout
}
