% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalign_update}
\alias{ggalign_update}
\title{Update the Properties of an Object}
\usage{
ggalign_update(x, object, ...)
}
\arguments{
\item{x}{The object whose properties are being updated.}

\item{object}{The object from which the properties will be applied to \code{x}.}
}
\description{
This generic function allows for updating the properties or state of an
object by applying changes from another object. It ensures that the input
object is modified according to the specifications of the provided \code{object}.
}
