% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno-.R
\name{gganno}
\alias{gganno}
\alias{gganno_top}
\alias{gganno_bottom}
\alias{gganno_left}
\alias{gganno_right}
\title{Heatmap annotation of \code{gganno}}
\usage{
gganno(
  mapping = aes(),
  data = NULL,
  position = NULL,
  size = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)

gganno_top(...)

gganno_bottom(...)

gganno_left(...)

gganno_right(...)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{ "bottom"}, and \code{"right"}. If \code{NULL}, the active context
of the \link[=ggheat]{ggheatmap} will be used.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{...}{Additional arguments passed to \code{gganno}.}
}
\value{
A \code{gganno} object.
}
\description{
Heatmap annotation of \code{gganno}
}
