% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{ggdendrogram}
\alias{ggdendrogram}
\title{Dendrogram plot}
\usage{
ggdendrogram(
  tree,
  mapping = NULL,
  ...,
  center = FALSE,
  type = "rectangle",
  cutree_k = NULL,
  cutree_height = NULL,
  leaf_braches = NULL,
  branch_gap = NULL,
  root = NULL,
  leaf_label = TRUE,
  leaf_guide = waiver()
)
}
\arguments{
\item{tree}{A \link[stats:hclust]{hclust} or a \link[stats:dendrogram]{dendrogram}
object.}

\item{mapping}{Additional aesthetic mappings to be added into
\link[ggplot2:geom_segment]{geom_segment}.}

\item{...}{Additional arguments passed to
\link[ggplot2:geom_segment]{geom_segment}}

\item{center}{A boolean value. if \code{TRUE}, nodes are plotted centered with
respect to the leaves in the branch. Otherwise (default), plot them in the
middle of all direct child nodes.}

\item{type}{A string indicates the plot type, \code{"rectangle"} or \code{"triangle"}.}

\item{leaf_braches}{Branches of the leaf node. Must be the same length of the
number of observations in \code{tree}. Usually come from \link[stats:cutree]{cutree}.}

\item{root}{A length one string or numeric indicates the root branch.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Dendrogram plot
}
\examples{
hc <- hclust(dist(USArrests), "ave")
ggdendrogram(hc)
ggdendrogram(hc, leaf_label = FALSE)
ggdendrogram(hc, aes(color = branch),
    leaf_braches = cutree(hc, 4L), type = "t"
)
ggdendrogram(hc, aes(color = branch),
    leaf_braches = cutree(hc, 4L), type = "r"
)
}
