% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{ggheat_build}
\alias{ggheat_build}
\title{Build ggheatmap for rendering.}
\usage{
ggheat_build(x, ...)
}
\arguments{
\item{x}{A \code{ggheatmap} object.}

\item{...}{Not used currently.}
}
\value{
A \code{patchwork} object.
}
\description{
Build ggheatmap for rendering.
}
