% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-add.R
\name{ggheatmap-add}
\alias{ggheatmap-add}
\alias{+.ggheatmap}
\alias{+,ggheatmap,ANY-method}
\title{Add components to \code{ggheatmap}}
\usage{
\S4method{+}{ggheatmap,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=ggheat]{ggheatmap} object.}

\item{e2}{An object to be added to the plot, including
\link[ggplot2:gg-add]{gg} elements, \link{gganno} object, or \link{htanno}
object.}
}
\value{
A modified \code{ggheatmap} object.
}
\description{
Add components to \code{ggheatmap}
}
\examples{
ggheat(matrix(rnorm(81), nrow = 9)) +
    gganno(position = "top") +
    geom_point(aes(y = value))
}
\seealso{
ggheatmap_add
}
