% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-and_add.R
\name{ggheatmap-and}
\alias{ggheatmap-and}
\alias{&.ggheatmap}
\alias{&,LayoutHeatmap,ANY-method}
\title{Add components to heatmap and heatmap annotations}
\usage{
\S4method{&}{LayoutHeatmap,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=ggheat]{ggheatmap} object.}

\item{e2}{An object to be added to the plot.}
}
\value{
A modified \code{ggheatmap} object.
}
\description{
Add components to heatmap and heatmap annotations
}
\seealso{
ggheatmap_and_add
}
