% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap-add.R
\name{ggheatmap_add}
\alias{ggheatmap_add}
\title{Add custom objects to \code{ggheatmap}}
\usage{
ggheatmap_add(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{A \code{ggheatmap} object}

\item{object_name}{The name of the object to add}
}
\value{
A modified \code{ggheatmap} object.
}
\description{
Add custom objects to \code{ggheatmap}
}
\examples{
ggheatmap_add(
    geom_point(aes(y = value)),
    ggheat(matrix(rnorm(81), nrow = 9)),
    deparse(quote(geom_point(aes(y = value))))
)
}
