% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheatmap-and.R
\name{ggheatmap_and_add}
\alias{ggheatmap_and_add}
\title{Add custom objects to \code{ggheatmap}}
\usage{
ggheatmap_and_add(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{A \code{ggheatmap} object}

\item{object_name}{The name of the object to add}
}
\value{
A modified \code{ggheatmap} object.
}
\description{
Add custom objects to \code{ggheatmap}
}
