% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-add.R
\name{heatmap-add}
\alias{heatmap-add}
\alias{+.ggheatmap}
\alias{+.LayoutHeatmap}
\alias{+,LayoutHeatmap,ANY-method}
\title{Add components to \code{LayoutHeatmap}}
\usage{
\S4method{+}{LayoutHeatmap,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=layout_heatmap]{LayoutHeatmap} object.}

\item{e2}{An object to be added to the plot, including
\link[ggplot2:gg-add]{gg} elements, \link{gganno} object, or \link{htanno}
object.}
}
\value{
A modified \code{LayoutHeatmap} object.
}
\description{
Add components to \code{LayoutHeatmap}
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("t") +
    gganno() +
    geom_point(aes(y = value))
}
\seealso{
layout_heatmap_add
}
