% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-and_add.R
\name{heatmap-and}
\alias{heatmap-and}
\alias{&.LayoutHeatmap}
\alias{&.ggheatmap}
\alias{&,LayoutHeatmap,ANY-method}
\title{Add components to heatmap and heatmap annotations}
\usage{
\S4method{&}{LayoutHeatmap,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=layout_heatmap]{LayoutHeatmap} object.}

\item{e2}{An object to be added to the plot.}
}
\value{
A modified \code{LayoutHeatmap} object.
}
\description{
Add components to heatmap and heatmap annotations
}
\seealso{
\link{layout_heatmap_and_add}
}
