% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{hmanno}
\alias{hmanno}
\title{Determine the active context of heatmap layout}
\usage{
hmanno(
  position = NULL,
  size = NULL,
  what = waiver(),
  width = NULL,
  height = NULL,
  guides = NA,
  free_labs = NA,
  free_sizes = NA,
  plot_data = NA
)
}
\arguments{
\item{position}{Which heatmap annotation should get activated? Possible
values are follows:
\itemize{
\item A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.
\item \code{NULL}: means set the active context into the \code{heatmap} itself.
}}

\item{size}{A \link[grid:unit]{unit} object to set the total size of the heatmap
annotation. This will only be used if \code{what} is a string of \code{"top"},
\code{"left"}, \code{"bottom"}, or \code{"right"}.
\itemize{
\item If position is \code{"top"} or \code{"bottom"}, \code{size} set the total height of the
annotation.
\item If position is \code{"left"} or \code{"right"}, \code{size} set the total width of the
annotation.
}}

\item{what}{What should get activated for the stack layout? Possible values
are follows:
\itemize{
\item A single number or string of the plot elements in the stack layout.
\item \code{NULL}: means set the active context into the last \code{Align} object. In
this way, we can add other \code{Align} objects into the \code{StackLayout}.
}}

\item{width, height}{Heatmap body width/height, can be a \link[grid:unit]{unit}
object.}

\item{guides}{A single boolean value indicates whether to collect the guides,
or you can specify the string of the guide position to collect. Allowed
strings are: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}

\item{free_labs}{A boolean value or a character of the axis position (\code{"t"},
\code{"l"}, \code{"b"}, \code{"r"}) indicates which axis title should be free from
alignment. By default, all axis title won't be aligned.}

\item{free_sizes}{A character specifies the ggplot elements which won't
count space sizes when alignment.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is to not modify the data. Use \code{NULL}, if you don't want to modify
anything.

Used to modify the data after layout has been created, but before the data is
handled of to the ggplot2 for rendering. Use this hook if the you needs
change the default data for all \code{geoms}.}
}
\value{
A \code{active} object which can be added into
\link[=layout_heatmap]{HeatmapLayout}.
}
\description{
Determine the active context of heatmap layout
}
