% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-.R
\docType{data}
\name{htanno}
\alias{htanno}
\alias{HtannoProto}
\title{Create a ggheatmap annotation object}
\usage{
htanno(
  htanno_class,
  params,
  position = NULL,
  mapping = NULL,
  labels = NULL,
  labels_nudge = waiver(),
  size = NULL,
  data = NULL,
  set_context = NULL,
  order = NULL,
  name = NULL,
  check.param = TRUE,
  call = caller_call()
)
}
\arguments{
\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of \code{ggheatmap} will be used.}

\item{labels}{Labels for axis parallelly with heatmap, the default will use
the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the heatmap
axis)
\item An expression vector (must be the same length as heatmap axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()}: if \code{labels} is \code{NULL}, then \code{labels_nudge} will be \code{NULL},
otherwise it will inherit from the heatmap.
\item A numeric.
}}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{data}{A matrix, a data frame, or even a simple vector that will be
converted into a one-column matrix. If the \code{data} argument is set to \code{NULL},
the function will use the heatmap matrix. Additionally, the \code{data} argument
can also accept a function (purrr-like lambda is also okay), which will be
applied with the heatmap matrix.

It is important to note that all annotations consider the \code{rows} as the
observations. It means the \code{NROW} function must return the same number with
the heatmap parallel axis. So for column annotation, the heatmap matrix will
be transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix).}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{check.param}{A single boolean value indicates whether to check the
supplied parameters and warn.}

\item{call}{The \code{call} used to construct the scale for reporting messages.}
}
\value{
A new \code{Class} object.
}
\description{
Create a ggheatmap annotation object
}
\section{HtannoProto}{

Each of the \verb{Htanno*} objects is just a \link[ggplot2:ggproto]{ggproto()} object,
descended from the top-level \code{HtannoProto}, and each implements various
methods and fields.

To create a new type of \verb{Htanno*} object, you typically will want to
override one or more of the following:
\itemize{
\item \code{setup_params}: Prepare parameter or check parameters used by this
annotation.
\item \code{setup_data}: Prepare data used by this annotation.
\item \code{compute}: A method used to compute statistics.
\item \code{layout}: A method used to group heamap rows/columns into panels or
reorder heamtap rows/columns.
\item \code{draw}: A method used to draw the plot. Must return a \code{ggplot} object.
}
}

\keyword{datasets}
\keyword{internal}
