% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-gg.R
\name{htanno_gg}
\alias{htanno_gg}
\alias{gganno}
\alias{gganno_top}
\alias{gganno_bottom}
\alias{gganno_left}
\alias{gganno_right}
\title{Heatmap annotation with \code{HtannoGG}}
\usage{
htanno_gg(
  mapping = aes(),
  data = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  position = NULL
)

gganno(
  mapping = aes(),
  data = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  position = NULL
)

gganno_top(...)

gganno_bottom(...)

gganno_left(...)

gganno_right(...)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{data}{A matrix, a data frame, or even a simple vector that will be
converted into a one-column matrix. If the \code{data} argument is set to \code{NULL},
the function will use the heatmap matrix. Additionally, the \code{data} argument
can also accept a function (purrr-like lambda is also okay), which will be
applied with the heatmap matrix.

It is important to note that all annotations consider the \code{rows} as the
observations. It means the \code{NROW} function must return the same number with
the heatmap parallel axis. So for column annotation, the heatmap matrix will
be transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix).}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{labels}{Labels for axis parallelly with heatmap, the default will use
the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the heatmap
axis)
\item An expression vector (must be the same length as heatmap axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()}: if \code{labels} is \code{NULL}, then \code{labels_nudge} will be \code{NULL},
otherwise it will inherit from the heatmap.
\item A numeric.
}}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of \code{ggheatmap} will be used.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
A \code{HtannoGG} object.
}
\description{
Heatmap annotation with \code{HtannoGG}
}
\examples{
ggheat(matrix(rnorm(81), nrow = 9)) +
    gganno(position = "top") +
    geom_point(aes(y = value))
}
