% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-group.R
\name{htanno_group}
\alias{htanno_group}
\title{Group Heatmap rows/columns}
\usage{
htanno_group(group, set_context = NULL, name = NULL, position = NULL)
}
\arguments{
\item{group}{A character define the heatmap groups, this will split the
heatmap into different panels.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{name}{A string of the annotation name.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of \code{ggheatmap} will be used.}
}
\value{
A new \code{Class} object.
}
\description{
Group Heatmap rows/columns
}
\examples{
small_mat <- matrix(rnorm(81), nrow = 9)
ggheat(small_mat) +
    htanno_group(
        sample(letters[1:4], ncol(small_mat), replace = TRUE),
        position = "top"
    )
}
