% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-reorder.R
\name{htanno_reorder}
\alias{htanno_reorder}
\title{Reorder Heatmap rows/columns}
\usage{
htanno_reorder(
  fun = rowMeans,
  ...,
  strict = TRUE,
  decreasing = FALSE,
  data = NULL,
  set_context = NULL,
  name = NULL,
  position = NULL
)
}
\arguments{
\item{fun}{A summary function. It should take a data and return the weights
for the heatmap rows/columns.}

\item{...}{Additional arguments passed to \code{fun}.}

\item{strict}{A boolean value indicates whether the order should be strict.
If previous groups exist, and strict is \code{FALSE}, this will reorder the
heatmap rows/columns in each group.}

\item{decreasing}{A boolean value. Should the sort order be increasing or
decreasing?}

\item{data}{A matrix, a data frame, or even a simple vector that will be
converted into a one-column matrix. If the \code{data} argument is set to \code{NULL},
the function will use the heatmap matrix. Additionally, the \code{data} argument
can also accept a function (purrr-like lambda is also okay), which will be
applied with the heatmap matrix.

It is important to note that all annotations consider the \code{rows} as the
observations. It means the \code{NROW} function must return the same number with
the heatmap parallel axis. So for column annotation, the heatmap matrix will
be transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix).}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{name}{A string of the annotation name.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of \code{ggheatmap} will be used.}
}
\value{
A new \code{Class} object.
}
\description{
Reorder Heatmap rows/columns
}
\examples{
ggheat(matrix(rnorm(81), nrow = 9)) +
    htanno_reorder(position = "top")
}
