% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-title.R
\name{htanno_title}
\alias{htanno_title}
\title{Draw Heatmap rows/columns titles}
\usage{
htanno_title(
  titles = NULL,
  ...,
  mapping = aes(),
  size = unit(1, "cm"),
  set_context = TRUE,
  order = NULL,
  name = NULL,
  position = NULL
)
}
\arguments{
\item{titles}{A character of titles, must be of length with the same number
of heatmap rows/columns panels.}

\item{...}{Additional arguments passed to \link[ggplot2:geom_text]{geom_text}.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of \code{ggheatmap} will be used.}
}
\value{
A new \code{Class} object.
}
\description{
Draw Heatmap rows/columns titles
}
\examples{
small_mat <- matrix(rnorm(81), nrow = 9)
ggheat(small_mat) +
    htanno_group(
        sample(letters[1:4], ncol(small_mat), replace = TRUE),
        position = "top"
    ) +
    htanno_title()
}
