% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{is.ggheatmap}
\alias{is.ggheatmap}
\title{Reports whether x is a ggheatmap object}
\usage{
is.ggheatmap(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A boolean value
}
\description{
Reports whether x is a ggheatmap object
}
\examples{
is.ggheatmap(ggheat(1:10))
}
