% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{is.ggstack}
\alias{is.ggstack}
\title{Reports whether \code{x} is a \code{StackLayout} object}
\usage{
is.ggstack(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A boolean value
}
\description{
Reports whether \code{x} is a \code{StackLayout} object
}
\examples{
is.ggstack(ggstack(1:10))
}
