% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{is_ggheatmap}
\alias{is_ggheatmap}
\title{Reports whether \code{x} is a \code{\link[=heatmap_layout]{heatmap_layout()}} object}
\usage{
is_ggheatmap(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A boolean value
}
\description{
Reports whether \code{x} is a \code{\link[=heatmap_layout]{heatmap_layout()}} object
}
\examples{
is_ggheatmap(ggheatmap(1:10))
}
