% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-operator.R
\name{layout-add}
\alias{layout-add}
\alias{+.Layout}
\alias{+.HeatmapLayout}
\alias{+.ggheatmap}
\alias{+.StackLayout}
\alias{+.ggstack}
\alias{+.QuadLayout}
\alias{+,Layout,ANY-method}
\title{Add components to \code{Layout}}
\usage{
\S4method{+}{Layout,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object.}

\item{e2}{An object to be added to the plot, including \link[ggplot2:gg-add]{gg}
elements or \link{align} object.}
}
\value{
A modified \code{Layout} object.
}
\description{
Add components to \code{Layout}
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    anno_top() +
    ggalign() +
    geom_point(aes(y = value))
}
