% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{layout-and}
\alias{layout-and}
\alias{&.Layout}
\alias{&.HeatmapLayout}
\alias{&.ggheatmap}
\alias{&.StackLayout}
\alias{&.ggstack}
\alias{&,Layout,ANY-method}
\title{Add components to all plots}
\usage{
\S4method{&}{Layout,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \code{\link[=layout_heatmap]{layout_heatmap()}} or \code{\link[=layout_stack]{layout_stack()}} object.}

\item{e2}{An object to be added to the plot.}
}
\value{
A modified \code{Layout} object.
}
\description{
Add components to all plots
}
