% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{Layout-subset}
\alias{Layout-subset}
\title{Subset a \code{Layout} object}
\arguments{
\item{x}{A \code{Layout} object}

\item{name}{A string of property name in \code{Layout} object.}
}
\value{
The property value.
}
\description{
Used by \code{\link[ggplot2:ggplot_build]{ggplot_build}} and
\code{\link[ggplot2:ggsave]{ggsave}}
}
\keyword{internal}
