% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-align_plots.R
\name{layout_annotation}
\alias{layout_annotation}
\title{Annotate the whole layout}
\usage{
layout_annotation(title = waiver(), subtitle = waiver(), caption = waiver())
}
\arguments{
\item{title}{The text for the title.}

\item{subtitle}{The text for the subtitle for the plot which will be
displayed below the title.}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}
}
\value{
A \code{layout_annotation} object to be added into \code{\link[=heatmap_layout]{heatmap_layout()}} or \code{\link[=stack_layout]{stack_layout()}} object.
}
\description{
Annotate the whole layout
}
