% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-align.R
\name{layout_expand}
\alias{layout_expand}
\title{Set Expansion for the Layout}
\usage{
layout_expand(..., x = waiver(), y = waiver())
}
\arguments{
\item{...}{A list of range expansion constants, used to add padding around
the data to ensure they are placed some distance away from the axes. Use the
convenience function \code{\link[ggplot2:expansion]{expansion()}} to generate the
values.}

\item{x, y}{Same as \code{...}, but specifically for \code{quad_layout()}.}
}
\description{
To align axes, it is important to keep the expansion consistent across all
plots in the layout. You can add a \code{layout_expand} object to the layout. For
the \code{quad_layout()} function, you must specify \code{x} and \code{y} arguments. For
other layouts, you can pass the expansion values using \code{...} directly.
}
\keyword{internal}
