% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-add.R
\name{layout_heatmap_add}
\alias{layout_heatmap_add}
\title{Add custom objects to \code{LayoutHeatmap}}
\usage{
layout_heatmap_add(object, heatmap, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{heatmap}{A \code{LayoutHeatmap} object}

\item{object_name}{The name of the object to add}
}
\value{
A modified \code{LayoutHeatmap} object.
}
\description{
Add custom objects to \code{LayoutHeatmap}
}
\examples{
layout_heatmap_add(
    geom_point(aes(y = value)),
    ggheatmap(matrix(rnorm(81), nrow = 9))
)
}
