% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-and_add.R
\name{layout_heatmap_and_add}
\alias{layout_heatmap_and_add}
\title{Add custom objects to heatmap and heatmap annotations}
\usage{
layout_heatmap_and_add(object, heatmap, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{heatmap}{A \link[=layout_heatmap]{LayoutHeatmap} object}

\item{object_name}{The name of the object to add}
}
\value{
A modified \code{LayoutHeatmap} object.
}
\description{
Add custom objects to heatmap and heatmap annotations
}
