% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{layout_stack}
\alias{layout_stack}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
layout_stack(
  data,
  direction = NULL,
  rel_sizes = NULL,
  guides = "collect",
  plot_data = NULL
)

ggstack(
  data,
  direction = NULL,
  rel_sizes = NULL,
  guides = "collect",
  plot_data = NULL
)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{rel_sizes}{A numeric or \link[grid:unit]{unit} object of length \code{3}
indicates the relative widths (\code{direction = "horizontal"}) / heights
(\code{direction = "vertical"}).}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. Used to modify the data
after layout preparation has been applied, but before the data is handled of
to the ggplot2 for rendering. The default is to not modify the data. Use this
hook if the you needs change the default data for all \code{geoms}.}
}
\value{
A \code{LayoutStack} object.
}
\description{
\code{ggstack} is an alias of \code{layout_stack}.
}
