% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-and_add.R
\name{layout_stack_and_add}
\alias{layout_stack_and_add}
\title{Add custom objects to all stack plots}
\usage{
layout_stack_and_add(object, stack, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{stack}{A \link[=layout_stack]{Layoutstack} object}

\item{object_name}{The name of the object to add}
}
\value{
A modified \code{LayoutStack} object.
}
\description{
Add custom objects to all stack plots
}
