% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{LayoutContext}
\alias{LayoutContext}
\title{LayoutContext class - represents the context for layout properties.
Context will be inherited by child elements, such as child layouts or child
plots.}
\usage{
LayoutContext(theme = theme(), schemes = Schemes())
}
\description{
LayoutContext class - represents the context for layout properties.
Context will be inherited by child elements, such as child layouts or child
plots.
}
