% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{LayoutControl}
\alias{LayoutControl}
\title{LayoutControl class - integrates properties that cannot be inherited.
Whenever a new property is added to this class, we must ensure
\code{LayoutControlUnion} is updated. This ensures that the \code{ggalign_update}
method is automatically implemented for the new property.}
\usage{
LayoutControl(titles = layout_title())
}
\description{
LayoutControl class - integrates properties that cannot be inherited.
Whenever a new property is added to this class, we must ensure
\code{LayoutControlUnion} is updated. This ensures that the \code{ggalign_update}
method is automatically implemented for the new property.
}
