% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link}
\alias{link}
\alias{link_range}
\title{Helper function to create a pair of links}
\usage{
link(link1 = NULL, link2 = NULL)

link_range(point1, point2)
}
\arguments{
\item{link1, link2}{An integer or character index, or a \code{link_range()} object
to define the linked observations. For integer indices, you can wrap them
with \code{\link[=I]{I()}} to indicate the order based on the layout. You can also use
\code{waiver()} to inherit the values from the opposite link argument.}

\item{point1, point2}{A single integer or character index, defining the lower
and higher bounds of the range. For integer indices, wrap them with \code{\link[=I]{I()}}
to indicate the ordered index by the layout.}
}
\description{
\itemize{
\item \code{link}: Helper function to create a pair of links.
\item \code{link_range}: Helper function to target a range of observations.
}
}
