% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_line}
\alias{link_line}
\title{Link the observations with a line}
\usage{
link_line(..., element = NULL)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dyn-dots}> A list of formulas, where each side
of the formula should be an \code{integer} or \code{character} index of the original
data, or a \code{range_link()} object defining the linked observations. Use
\code{NULL} to indicate no link on that side. You can also combine these by
wrapping them into a single \code{list()}. If only the left-hand side of the
formula exists, you can input it directly. For integer indices, wrap them
with \code{\link[=I]{I()}} to use the ordering from the layout. You can also use
\code{\link[=waiver]{waiver()}} to inherit values from the opposite link argument.}

\item{element}{A \code{\link[ggplot2:element]{element_line()}} object.}
}
\description{
Link the observations with a line
}
