% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link_tetragon}
\alias{link_tetragon}
\title{Link the paired observations with a quadrilateral}
\usage{
link_tetragon(..., .element = NULL)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dyn-dots}> A list of formulas, where each side
of the formula should be an \code{integer} or \code{character} index of the original
data, or a \code{range_link()} object defining the linked observations. Use
\code{NULL} to indicate no link on that side. You can also combine these by
wrapping them into a single \code{list()}. If only the left-hand side of the
formula exists, you can input it directly. For integer indices, wrap them
with \code{\link[=I]{I()}} to use the ordering from the layout. You can also use
\code{\link[ggplot2:waiver]{waiver()}} to inherit values from the other group.}

\item{.element}{A \code{\link[=element_polygon]{element_polygon()}} object. Vectorized fields will be
recycled to match the total number of groups, or you can wrap the element
with \code{\link[=I]{I()}} to recycle to match the drawing groups. The drawing groups
are usually the same as the defined groups, but they will differ when the
defined group of observations is separated and cannot be linked with a
single quadrilateral. In such cases, the number of drawing groups will be
larger than the number of defined groups.}
}
\description{
Link the paired observations with a quadrilateral
}
