% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_pdraw}
\alias{mark_pdraw}
\title{Define the links to connect the marked observations}
\usage{
mark_pdraw(.draw, ..., .link1 = NULL, .link2 = NULL)
}
\arguments{
\item{.draw}{A function used to draw the links. The function must return a
\code{\link[grid:grid.grob]{grob()}} object. If the function does not return a valid
\code{grob}, nothing will be drawn. The input data for the function contains a
list, where each item is a list of two data frames: one for the panel side
coordinates (\code{"panel"}) and one for the marked observations coordinates
(\code{"link"}).}

\item{...}{<\link[rlang:dyn-dots]{dyn-dots}> A list of \code{link()} objects used to
define the linked observations for a panel. Each element of the list will
define a plot panel.}

\item{.link1, .link2}{Same as \code{link1} and \code{link2}. But you can provide a
list of them directly.}
}
\description{
A base version of \code{\link{mark_draw}}, designed for performance optimization.  This
function is used to build other \verb{mark_*} functions that manage the drawing of
links between marked observations.
}
