% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_tetragon}
\alias{mark_tetragon}
\title{Link the observations and the panel with a quadrilateral}
\usage{
mark_tetragon(..., element = NULL)
}
\arguments{
\item{...}{Additional arguments passed on to \code{\link[=mark_draw]{mark_draw()}}.}

\item{element}{A \code{\link[=element_polygon]{element_polygon()}} object.}
}
\description{
Link the observations and the panel with a quadrilateral
}
