% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-.R
\name{new_anno}
\alias{new_anno}
\title{Create a anno object.}
\usage{
new_anno(
  Class,
  ...,
  data = NULL,
  position = NULL,
  size = unit(10, "mm"),
  active = NULL,
  name = NULL,
  order = NULL
)
}
\arguments{
\item{Class}{Sub-class of \code{anno}.}

\item{...}{Additional components of the sub-class.}

\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{ "bottom"}, and \code{"right"}.}

\item{size}{Annotation size, must be a \link[grid:unit]{unit} object.}

\item{active}{A logical value of length \code{1} or \code{2}.}

\item{name}{A string of the annotation name.}

\item{order}{Annotation order, must be an single integer.}
}
\value{
A new \code{Class} object.
}
\description{
Create a anno object.
}
