% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-.R
\name{new_gganno}
\alias{new_gganno}
\title{Used by both gganno and htanno ------------------------------}
\usage{
new_gganno(
  Class,
  ...,
  data = NULL,
  position = NULL,
  size = unit(10, "mm"),
  order = NULL
)
}
\description{
Used by both gganno and htanno ------------------------------
}
