% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-.R
\name{new_htanno}
\alias{new_htanno}
\title{Build \code{htanno} object}
\usage{
new_htanno(
  htanno,
  data = NULL,
  position = NULL,
  params = list(),
  size = unit(10, "mm"),
  active = NULL,
  name = NULL,
  order = NULL,
  check.param = TRUE
)
}
\arguments{
\item{htanno}{A \code{HtannoProto} object}
}
\description{
\code{htanno} is a special annotation, which can act with the main heatmap.
Especially control the order of the main heatmap or split the heatmap into
different slices.
}
