% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{new_tune}
\alias{new_tune}
\alias{tune_data}
\title{Change the shape of the input object}
\usage{
new_tune(x, ..., class = character())

tune_data(x)
}
\arguments{
\item{x}{An R object.}

\item{...}{Additional attributes passed to \code{\link[=structure]{structure()}}.}

\item{class}{A character vector specifying the class name to be added.}
}
\description{
\itemize{
\item \code{new_tune}: Creates a new object by wrapping it in a scalar list with
the specified attributes and class.
\item \code{tune_data}: Retrieves the original input data.
}
}
