% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helper.R
\name{no_expansion}
\alias{no_expansion}
\title{Remove scale expansion}
\usage{
no_expansion(borders = "tlbr")
}
\arguments{
\item{borders}{Which border should be removed? A string containing
one or more of \code{"t"}, \code{"l"}, \code{"b"}, \code{"r"}, \code{"x"}, and \code{"y"}.}
}
\value{
An object which can be added to ggplot.
}
\description{
Remove scale expansion
}
