% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft-align-order2.R,
%   R/layout-heatmap-oncoplot.R
\name{order2}
\alias{order2}
\alias{order2.hclust}
\alias{order2.dendrogram}
\alias{order2.ser_permutation_vector}
\alias{order2.ser_permutation}
\alias{order2.phylo}
\alias{order2.memo_weights}
\title{Ordering Permutation}
\usage{
order2(x)

\method{order2}{hclust}(x)

\method{order2}{dendrogram}(x)

\method{order2}{ser_permutation_vector}(x)

\method{order2}{ser_permutation}(x)

\method{order2}{phylo}(x)

\method{order2}{memo_weights}(x)
}
\arguments{
\item{x}{Any objects can be extracting ordering.}
}
\value{
An integer vector unless any of the inputs has \code{2^31} or more
elements, when it is a double vector.
}
\description{
\code{order2} returns a permutation which rearranges its first argument into
ascending order.
}
\examples{
order2(hclust2(matrix(rnorm(100L), nrow = 10L)))
}
