% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch.alignpatches}
\alias{patch.alignpatches}
\title{Convert Object into a Grob}
\usage{
\method{patch}{alignpatches}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Not used currently.}
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
The \code{patch()} function is used by \code{\link[=ggwrap]{ggwrap()}} and \code{\link[=inset]{inset()}} to convert
objects into a \code{\link[grid:grid.grob]{grob}}.
}
\seealso{
\code{\link[=align_plots]{alignpatches}}

Other patch methods: 
\code{\link{patch.Heatmap}()},
\code{\link{patch.formula}()},
\code{\link{patch.ggplot}()},
\code{\link{patch.grob}()},
\code{\link{patch.patch}()},
\code{\link{patch.patch_ggplot}()},
\code{\link{patch.patchwork}()},
\code{\link{patch.pheatmap}()},
\code{\link{patch.recordedplot}()},
\code{\link{patch.trellis}()}
}
\concept{patch methods}
