% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{patch.formula}
\alias{patch.formula}
\title{Convert Object into a Grob}
\usage{
\method{patch}{formula}(x, ..., device = NULL)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{
  Arguments passed on to \code{\link[graphics:par]{graphics::par}}
  \describe{
    \item{\code{}}{}
  }}

\item{device}{
    A function that opens a graphics device for \code{grid.echo()} to
    work on.  By default this is an off-screen, in-memory device
    based on the \code{pdf} device.  This default device may not be
    satisfactory when using custom fonts.
  }
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} to convert objects into a
\link[grid:grid.grob]{grob}, enabling their alignment within \code{\link[=align_plots]{align_plots()}}.
}
