% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch.patch}
\alias{patch.patch}
\title{Convert Object into a Grob}
\usage{
\method{patch}{patch}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Not used currently.}
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=ggwrap]{ggwrap()}} and \code{\link[=inset]{inset()}} to convert objects
into a \link[grid:grid.grob]{grob}.
}
\seealso{
\link[patchwork:patchGrob]{patch}
}
