% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch.patch_ggplot}
\alias{patch.patch_ggplot}
\title{Convert Object into a Grob}
\usage{
\method{patch}{patch_ggplot}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Not used currently.}
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} and \code{\link[=inset]{inset()}} to convert objects
into a \link[grid:grid.grob]{grob}.
}
\seealso{
\itemize{
\item \link{patch_titles}
\item \link{inset}
\item \link{wrap}
}
}
