% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{patch}
\alias{patch}
\title{Get Patch representation}
\usage{
patch(x)
}
\arguments{
\item{x}{Any objects has a \link{Patch} representation}
}
\value{
A \code{ggalign::Patch}
object.
}
\description{
\code{ggalign::Patch} represents the
layout manager for a single subplot within a composite plot. The \code{Patch}
object provides the interface for aligning the subplot, managing panel sizes,
and handling guide legends.
}
\examples{
patch(ggplot())
}
\seealso{
\code{\link{Patch}}
}
\keyword{internal}
