% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-alignpatches.R
\name{patch_options}
\alias{patch_options}
\title{Options passed to the Patch \code{gtable} method}
\usage{
patch_options(theme = NULL, guides = NULL, tagger = NULL)
}
\description{
This class defines the options that can be passed to the \code{gtable} method of a
\code{Patch} object.  It includes:
}
\details{
\itemize{
\item \code{theme}: The theme to be applied, which can be either \code{NULL} or a ggplot2
\link[ggplot2:theme]{theme} object.
\item \code{guides}: The guides for the plot, which can be \code{NULL} or a character
vector.
\item \code{tagger}: Either \code{NULL} (no tagging) or a \code{LayoutTagger} object that
provides a \verb{$tag_table} method (accepting the \code{gtable} and \code{theme})
used to add tag.
}
}
\keyword{internal}
