% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action.R
\name{plot_action}
\alias{plot_action}
\title{Plot Action Specifications in the Layout}
\usage{
plot_action(
  data = NA,
  theme = NA,
  guides = NA,
  free_spaces = NA,
  free_labs = NA,
  inherit = NA
)
}
\arguments{
\item{data}{A function to transform the plot data before rendering. Whether
this function is applied after the parent layout's action data depends on
the \code{inherit} argument. Defaults to \code{\link[ggplot2:waiver]{waiver()}}, which
directly inherits from the parent layout. If no parent layout is specified,
the default is \code{NULL}, meaning the data won't be modified.

Use this hook to modify the data for all \code{geoms} after the layout is created
but before rendering by \code{ggplot2}. The returned data must be a data frame.}

\item{theme}{Default plot theme, one of:
\itemize{
\item \code{NULL}: will inherit from the parent layout directly.
\item \code{\link[ggplot2:theme]{theme()}}: will be added with the parent layout theme.
If you want to override the parent layout theme, set \code{complete=TRUE}.
}

\strong{Note:} Axis titles and labels that are parallel to the layout axis will
always be removed by default. For vertical stack layouts, this refers to the
x-axis; for horizontal stack layouts, this refers to the y-axis. To display
these axis titles or labels, you must manually add the appropriate
\link[ggplot2:theme]{theme()} elements for the parallel axis.}

\item{guides}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicating
which side of guide legends should be collected. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent layout, all guides will be collected. If \code{NULL}, no guides will be
collected.}

\item{free_spaces}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}
indicating which border spaces should be removed. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent, the default is \code{NULL}, meaning no spaces are removed.

Usually you want to apply this with the whole layout, instead of individual
plots.}

\item{free_labs}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicating
which axis titles should be free from alignment. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent layout, no axis titles will be aligned. If \code{NULL}, all axis titles
will be aligned.}

\item{inherit}{A single boolean value indicating whether to apply the parent
action \code{data} first and then apply the specified action \code{data}. Defaults to
\code{FALSE}.}
}
\value{
A \code{plot_action} object.
}
\description{
The \code{plot_action()} function defines the behavior of plots within a layout.
It can be used in the \code{action} argument of layout functions like \code{hmanno()}
or \code{stack_active()} to set global actions for all plots in the layout.
Additionally, \code{plot_action()} can be applied directly to specific plots
through the \code{action} argument in the \verb{align_*()} functions, or it can be
added directly to a plot.
}
\examples{
# used in the layout, define the default action for all plots in the layout
ggheatmap(matrix(rnorm(100L), nrow = 10),
    action = plot_action(
        theme = theme(plot.background = element_rect(fill = "red"))
    )
)

# You can also add it for a single plot
ggheatmap(matrix(rnorm(100L), nrow = 10),
    action = plot_action(
        theme = theme(plot.background = element_rect(fill = "red"))
    )
) + plot_action( # here, we modify the plot action for the heatmap body
    theme = theme(plot.background = element_rect(fill = "blue"))
)

}
