% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme-.R
\name{plot_add_scheme}
\alias{plot_add_scheme}
\title{Apply a scheme to a plot}
\usage{
plot_add_scheme(plot, scheme, ...)
}
\arguments{
\item{plot}{A plot object, typically a ggplot.}

\item{scheme}{A \code{\link{Scheme}} or \code{\link{Schemes}} object to apply.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
The modified \code{plot} object.
}
\description{
\code{plot_add_scheme()} is a generic used to apply a \code{\link{Scheme}} (or a \code{\link{Schemes}}
container) to a plot object. This allows schemes to update or modify
the plot's appearance or behavior according to their configuration.
}
\details{
By default When a \code{\link{Schemes}} object is passed, each individual \code{\link{Scheme}} is
applied in sequence via its respective \code{plot_add_scheme()} method.
}
