% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{plot_context}
\alias{plot_context}
\title{Plot context related actions}
\usage{
plot_context(order = waiver(), active = waiver(), name = waiver())
}
\arguments{
\item{order}{An integer number specifying the order of the plot area in the
layout.}

\item{active}{A single boolean value; if \code{TRUE}, sets the active context to
the current plot when added to a layout, so all subsequent \code{ggplot}
elements are added to this plot.}

\item{name}{A single string specifying the plot name, used to switch active
contexts in \code{what} argument of \code{\link[=quad_anno]{quad_anno()}}/\code{\link[=stack_switch]{stack_switch()}}.}
}
\description{
Plot context related actions
}
