% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-plot_grid.R
\name{plot_grid}
\alias{plot_grid}
\title{Arrange multiple plots into a grid}
\usage{
plot_grid(
  ...,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  widths = NULL,
  heights = NULL,
  guides = NULL,
  design = NULL
)
}
\description{
Arrange multiple plots into a grid
}
