% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craftbox-.R
\name{+,CraftBox,ANY-method}
\alias{+,CraftBox,ANY-method}
\title{Add custom objects to ggalign plot}
\usage{
\S4method{+}{CraftBox,ANY}(e1, e2)
}
\description{
Add custom objects to ggalign plot
}
\keyword{internal}
