% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-.R
\name{+,ggalign_plot,ANY-method}
\alias{+,ggalign_plot,ANY-method}
\title{Add custom objects to ggalign plot}
\usage{
\S4method{+}{ggalign_plot,ANY}(e1, e2)
}
\description{
Add custom objects to ggalign plot
}
\keyword{internal}
