% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-quad-.R
\name{quad_free}
\alias{quad_free}
\alias{quad_alignh}
\alias{quad_alignv}
\alias{quad_alignb}
\title{Arrange Plots in the Quad-Side of a main plot}
\usage{
quad_free(
  data = NULL,
  mapping = aes(),
  ...,
  action = NULL,
  theme = NULL,
  context = NULL,
  width = NA,
  height = NA
)

quad_alignh(
  data = NULL,
  mapping = aes(),
  ...,
  action = NULL,
  theme = NULL,
  context = NULL,
  width = NA,
  height = NA
)

quad_alignv(
  data = NULL,
  mapping = aes(),
  ...,
  action = NULL,
  theme = NULL,
  context = NULL,
  width = NA,
  height = NA
)

quad_alignb(
  data = NULL,
  mapping = aes(),
  ...,
  action = NULL,
  theme = NULL,
  context = NULL,
  width = NA,
  height = NA
)
}
\arguments{
\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout. By default, it will try to inherit from parent layout.
\itemize{
\item For \code{quad_free}, the function uses \code{\link[=fortify_data_frame]{fortify_data_frame()}} to convert
the data into a data frame.
\item For all other functions, it employs \code{\link[=fortify_matrix]{fortify_matrix()}} to convert
the data into a matrix.
}}

\item{mapping}{Default list of aesthetic mappings to use for the main plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Additional arguments passed to \code{\link[=fortify_matrix]{fortify_matrix()}} or
\code{\link[=fortify_data_frame]{fortify_data_frame()}}.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used to define the default action
for the plots in the layout.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} used to render the \code{guides},
\code{title}, \code{subtitle}, \code{caption}, \code{margins}, \code{patch.title}, \code{panel.border}, and
\code{background}. If \code{NULL} (default), will inherit from the parent \code{layout}.}

\item{context}{A \code{\link[=plot_context]{plot_context()}} object that defines the action when
added to a layout.}

\item{width, height}{The relative width/height of the main plot, can be a
\code{\link[grid:unit]{unit}} object}
}
\description{
\itemize{
\item \code{quad_free}: Never align observations.
\item \code{quad_alignh}: Align observations in the horizontal direction.
\item \code{quad_alignv}: Align observations in the vertical direction.
\item \code{quad_alignb}: Align observations in both horizontal and vertical
directions.
}
}
\details{
These functions arrange plots around a main plot, allowing for flexible
alignment of observations in different directions.
}
