% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-quad-switch.R
\name{quad_init}
\alias{quad_init}
\title{Initialize Quad-Layout Annotation}
\usage{
quad_init(position, data = waiver(), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated, you can add \code{stack_layout()} directly.
}
\keyword{internal}
