% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_example}
\alias{read_example}
\title{Read Example Data}
\usage{
read_example(file = NULL)
}
\arguments{
\item{file}{A string representing the name of the example file to be read. If
\code{NULL}, the function will return a list of available example file names.}
}
\value{
If \code{file} is \code{NULL}, returns a character vector of available example
file names. Otherwise, returns the contents of the specified example file,
read as an R object.
}
\description{
This function reads example data from the file. If no file is specified, it
returns a list of available example files.
}
