% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-grid.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gpar}
\alias{unit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{gpar}}, \code{\link[grid]{unit}}}
}}

