% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme-.R
\name{scheme_inherit}
\alias{scheme_inherit}
\title{Inherit a scheme from a parent}
\usage{
scheme_inherit(e1, e2)
}
\arguments{
\item{e1}{The parent \code{scheme} object.}

\item{e2}{The child \code{scheme} object. Usually should have the same \code{key} as
\code{e1}.}
}
\value{
A new \code{scheme} object.
}
\description{
This generic is used by developers to define how one \code{scheme} object inherits
from another (typically the scheme defined in the layout). This is called
when adding a new \code{scheme} via inheritance.
}
