% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-.R
\name{show,anno-method}
\alias{show,anno-method}
\title{Show \code{anno} object}
\usage{
\S4method{show}{anno}(object)
}
\arguments{
\item{object}{A \code{anno} object.}
}
\description{
Show \code{anno} object
}
\keyword{internal}
