% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craftbox-.R
\name{show,CraftBox-method}
\alias{show,CraftBox-method}
\title{Show \code{CraftBox} information}
\usage{
\S4method{show}{CraftBox}(object)
}
\arguments{
\item{object}{A \code{CraftBox} object.}
}
\value{
The input invisiblely.
}
\description{
Show \code{CraftBox} information
}
\keyword{internal}
