% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-.R
\name{show,ggalign_plot-method}
\alias{show,ggalign_plot-method}
\title{Show \code{ggalign_plot} information}
\usage{
\S4method{show}{ggalign_plot}(object)
}
\arguments{
\item{object}{A \code{ggalign_plot} object.}
}
\value{
The input invisiblely.
}
\description{
Show \code{ggalign_plot} information
}
\keyword{internal}
