% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{show,LayoutHeatmap-method}
\alias{show,LayoutHeatmap-method}
\title{Show \code{LayoutHeatmap}}
\usage{
\S4method{show}{LayoutHeatmap}(object)
}
\arguments{
\item{object}{A \code{LayoutHeatmap} object.}
}
\description{
Show \code{LayoutHeatmap}
}
