% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{show,LayoutProto-method}
\alias{show,LayoutProto-method}
\title{Print Layout object}
\usage{
\S4method{show}{LayoutProto}(object)
}
\arguments{
\item{object}{A \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object.}
}
\value{
The input invisiblely.
}
\description{
Print Layout object
}
\keyword{internal}
