% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-add.R
\name{stack-add}
\alias{stack-add}
\alias{+.ggstack}
\alias{+.LayoutStack}
\alias{+,LayoutStack,ANY-method}
\title{Add components to \code{LayoutStack}}
\usage{
\S4method{+}{LayoutStack,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=layout_stack]{LayoutStack} object.}

\item{e2}{An object to be added to the plot, including
\link[ggplot2:gg-add]{gg} elements, \link{gganno} object, or \link{htanno}
object.}
}
\value{
A modified \code{LayoutStack} object.
}
\description{
Add components to \code{LayoutStack}
}
\examples{
layout_stack(matrix(rnorm(81), nrow = 9)) +
    gganno() +
    geom_point(aes(y = value))
}
\seealso{
layout_stack_add
}
