% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-and_add.R
\name{stack-and}
\alias{stack-and}
\alias{&.Layoutstack}
\alias{&.ggstack}
\alias{&,LayoutStack,ANY-method}
\title{Add components to all stack plots}
\usage{
\S4method{&}{LayoutStack,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=layout_stack]{Layoutstack} object.}

\item{e2}{An object to be added to the plot.}
}
\value{
A modified \code{LayoutStack} object.
}
\description{
Add components to all stack plots
}
\seealso{
\link{layout_stack_and_add}
}
