% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_continuous}
\alias{stack_continuous}
\alias{stack_free}
\alias{stack_continuousv}
\alias{stack_freev}
\alias{stack_continuoush}
\alias{stack_freeh}
\title{Arrange Plots Horizontally or Vertically by aligning continuous axis}
\usage{
stack_continuous(
  direction,
  data = NULL,
  ...,
  limits = NULL,
  theme = NULL,
  sizes = NA
)

stack_free(
  direction,
  data = NULL,
  ...,
  limits = NULL,
  theme = NULL,
  sizes = NA
)

stack_continuousv(data = NULL, ...)

stack_freev(data = NULL, ...)

stack_continuoush(data = NULL, ...)

stack_freeh(data = NULL, ...)
}
\arguments{
\item{direction}{A string indicating the direction of the stack layout,
either \code{"h"}(\code{horizontal}) or \code{"v"}(\code{vertical}).}

\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout, \code{\link[=fortify_data_frame]{fortify_data_frame()}} will be used to
convert data to a data frame. Note that if the data is a \code{matrix}, it will
be automatically converted to a long-formatted data frame, which differs
from \code{ggplot2}'s behavior.}

\item{...}{Additional arguments passed to \code{\link[=fortify_data_frame]{fortify_data_frame()}}.}

\item{limits}{A \code{\link[=continuous_limits]{continuous_limits()}} object specifying the left/lower
limit and the right/upper limit of the scale. Used to align the continuous
axis.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} object used to customize various
elements of the layout, including \code{guides}, \code{title}, \code{subtitle}, \code{caption},
\code{margins}, \code{panel.border}, and \code{background}. By default, the theme will
inherit from the parent \code{layout}. It also controls the panel spacing for all
plots in the layout.}

\item{sizes}{A numeric value or a \code{\link[grid:unit]{unit}} object. If used for
the \code{\link[=quad_layout]{quad_layout()}} annotation, it must have a length of \code{1}.  If used in
the layout with a nested \code{\link[=quad_layout]{quad_layout()}}, it should have a length of \code{3},
indicating the relative heights (for \code{direction = "horizontal"}) or widths
(for \code{direction = "vertical"}) to be applied.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{stack_continuous} function arranges plots by aligning continuous
variables. The alias \code{stack_free} is retained for historical reasons.
\itemize{
\item \code{stack_continuousv()}/\code{stack_freev}: A special case of \code{stack_free} that
sets \code{direction = "vertical"}.
\item \code{stack_continuoush()}/\code{stack_freeh}: A special case of \code{stack_free} that
sets \code{direction = "horizontal"}.
}
}
