% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_layout}
\alias{stack_layout}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
stack_layout(data, direction = NULL, sizes = NA, ...)

ggstack(data, direction = NULL, sizes = NA, ...)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{sizes}{A numeric or \link[grid:unit]{unit} object of length \code{3} indicates
the relative widths (\code{direction = "horizontal"}) / heights (\code{direction = "vertical"}).}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \code{StackLayout} object.
}
\description{
\code{ggstack} is an alias of \code{stack_layout}.
}
\examples{
ggstack(matrix(rnorm(100L), nrow = 10L)) + align_dendro()
}
