% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-.R
\name{summary.AlignProto}
\alias{summary.AlignProto}
\title{Summary the action of \code{AlignProto}}
\usage{
\method{summary}{AlignProto}(object, ...)
}
\arguments{
\item{object}{A \code{AlignProto} object}
}
\value{
A logical vector of length 2, indicating:
\itemize{
\item Whether the object reorders the observations.
\item Whether the object splits the observations into groups.
}
}
\description{
Summary the action of \code{AlignProto}
}
\keyword{internal}
