% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-.R
\name{summary.ggalign_plot}
\alias{summary.ggalign_plot}
\title{Summary the action of \code{ggalign_plot}}
\usage{
\method{summary}{ggalign_plot}(object, ...)
}
\arguments{
\item{object}{A \code{ggalign_plot} object}
}
\value{
A logical vector of length 3, indicating:
\itemize{
\item Whether the object adds a plot.
\item Whether the object reorders the observations.
\item Whether the object splits the observations into groups.
}
}
\description{
Summary the action of \code{ggalign_plot}
}
\keyword{internal}
