% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix-list.R
\name{tune.list}
\alias{tune.list}
\title{Convert the shape of a list for fortify method}
\usage{
\method{tune}{list}(data, shape = NULL)
}
\arguments{
\item{data}{A list}

\item{shape}{Not used currently.}
}
\description{
Convert the shape of a list for fortify method
}
\seealso{
\code{\link[=fortify_matrix.list_upset]{fortify_matrix.list_upset()}}

Other tune: 
\code{\link{tune.MAF}()},
\code{\link{tune.matrix}()}
}
\concept{tune}
