% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix-maftools.R
\name{tune.MAF}
\alias{tune.MAF}
\title{Convert the shape of a MAF for fortify method}
\usage{
\method{tune}{MAF}(data, shape = NULL)
}
\arguments{
\item{data}{A \code{\link[maftools:read.maf]{MAF}} object.}

\item{shape}{Not used currently.}
}
\description{
Convert the shape of a MAF for fortify method
}
\seealso{
\code{\link[=fortify_matrix.MAF_pathways]{fortify_matrix.MAF_pathways()}}

Other tune: 
\code{\link{tune.list}()},
\code{\link{tune.matrix}()}
}
\concept{tune}
