% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.R
\name{tune}
\alias{tune}
\title{Change the shape of the input object}
\usage{
tune(data, shape = NULL)
}
\arguments{
\item{data}{An R object.}

\item{shape}{Usually a string, specifying the new shape for the object. Refer
to the detailed method for allowed values.}
}
\description{
Change the shape of the input object
}
\details{
In most cases, \code{\link[=fortify_matrix]{fortify_matrix()}} or \code{\link[=fortify_data_frame]{fortify_data_frame()}} provide full
support for transforming objects. However, some objects may require two
completely different approaches to be fortified. The \code{tune} function acts as
a helper to create a new class tailored for these objects.
}
\seealso{
\itemize{
\item \code{\link[=tune.MAF]{tune.MAF()}}
}
}
