% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-quad-scope.R
\name{with_quad}
\alias{with_quad}
\title{Create ggplot object with layout panel data}
\usage{
with_quad(x, position = NULL, main = deprecated(), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated, please use \code{quad_scope()} instead.
}
\keyword{internal}
