% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{wrap}
\alias{wrap}
\title{Convert Object into a Grob for Wrapping}
\usage{
wrap(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{wrap()} function is used by \code{\link[=alignwrap]{alignwrap()}} to convert objects into a
\link[grid:grid.grob]{grob}, enabling their alignment within \code{\link[=plot_grid]{plot_grid()}}.
}
\keyword{internal}
