% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\title{Determine the context of subsequent manipulations}
\usage{
activate(x, what)
}
\arguments{
\item{x}{A \link[=ggheat]{ggheatmap} object.}
}
\value{
A object with the same class of \code{x}, whose active context will be
\code{set}.
}
\description{
Determine the context of subsequent manipulations
}
