% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\title{Determine the context of subsequent manipulations}
\usage{
activate(x, what)
}
\arguments{
\item{x}{A \code{\link[=layout_heatmap]{layout_heatmap()}} or \code{\link[=layout_stack]{layout_stack()}} object.}

\item{what}{What should get activated? See \code{position} of \code{\link[=hmanno]{hmanno()}} or \code{what}
of \code{\link[=stack_active]{stack_active()}}.}
}
\value{
A object with the same class of \code{x}, whose active context will be
\code{set}.
}
\description{
Determine the context of subsequent manipulations
}
