% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{deactivate}
\alias{activate.ggheatmap}
\alias{deactivate.ggheatmap}
\alias{active}
\alias{active.default}
\alias{active.ggheat}
\alias{active<-}
\alias{active<-.default}
\alias{active<-.ggheat}
\title{Determine the context of subsequent manipulations}
\usage{
activate(x, what)

deactivate(x)

\method{activate}{ggheatmap}(x, what)

\method{deactivate}{ggheatmap}(x)

active(x)

\method{active}{default}(x)

\method{active}{ggheat}(x)

active(x) <- value

\method{active}{default}(x) <- value

\method{active}{ggheat}(x) <- value
}
\arguments{
\item{x}{A \link[=ggheat]{ggheatmap} or \link[=gganno]{ggannotation} object.}

\item{what}{What should get activated?
\itemize{
\item \code{ggheatmap}: Possible values are "top", "left", "bottom", and "right".
\item \code{ggannotation}: A string of name in the annotation.
}}
}
\value{
\itemize{
\item \code{activate}: A object with the same class of \code{x}.
}

\itemize{
\item \code{active}: A string of current active context.
}

\itemize{
\item \verb{active<-}: The same with \code{activate}, but won't check the arguments.
}
}
\description{
Determine the context of subsequent manipulations
}
