% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate}
\alias{activate}
\alias{deactivate}
\alias{activate.ggheatmap}
\alias{deactivate.ggheatmap}
\title{Activate the context of subsequent manipulations}
\usage{
activate(x, what)

deactivate(x)

\method{activate}{ggheatmap}(x, what)

\method{deactivate}{ggheatmap}(x)
}
\arguments{
\item{x}{A \link[=ggheat]{ggheatmap} object.}

\item{what}{What should get activated? Possible values are \code{"top"}, \code{"left"},
\code{ "bottom"}, and \code{"right"}.}
}
\value{
A object with the same class of \code{x}, whose active context will be set
or unset.
}
\description{
Activate the context of subsequent manipulations
}
