% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-.R
\docType{data}
\name{align}
\alias{align}
\alias{Align}
\title{Create a New \code{Align} Object}
\usage{
align(
  align_class,
  params,
  data,
  size = NULL,
  action = NULL,
  action_data = NA,
  limits = TRUE,
  facet = TRUE,
  context = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated(),
  check.param = TRUE,
  call = caller_call()
)
}
\arguments{
\item{data}{Options for \code{data}:
\itemize{
\item A matrix, data frame, or atomic vector.
\item \code{\link[=waiver]{waiver()}}: Uses the \verb{layout matrix}.
\item \code{NULL}: No data is set.
\item A \code{function} (including purrr-like \code{lambda} syntax), which will be applied
to the \verb{layout matrix}.
}}

\item{size}{Plot size, specified as a \code{\link[grid:unit]{unit}} object.}

\item{action}{Options for \code{action}:
\itemize{
\item \code{NULL}: For \verb{align_*()} functions that do not add a plot.
\item \code{\link[=waiver]{waiver()}}: Uses the default \code{\link[=plot_action]{plot_action()}}.
\item A \code{\link[=plot_action]{plot_action()}} object to define specific plot actions.
}}

\item{action_data}{Default action data, inferred from the \code{data} argument by
default.}

\item{limits}{Logical; determines if layout limits are set for the plot.}

\item{facet}{Logical; determines if layout facets are set for the plot. If
\code{FALSE}, \code{limits} will also be set to \code{FALSE}.}

\item{context}{A \code{\link[=plot_context]{plot_context()}} object that defines the action when
added to a layout.}

\item{free_guides}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{free_spaces}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{plot_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{theme}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_labs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{check.param}{Logical; if \code{TRUE}, checks the supplied parameters and
provides warnings as necessary.}

\item{call}{The \code{call} used to construct the \code{Align} object, for reporting
messages.}
}
\value{
A new \code{Align} object.
}
\description{
An \code{Align} object interacts with the \code{layout} object to reorder or split
observations and, in some cases, add plot components to the \code{layout}.
}
\details{
Each of the \verb{Align*} objects is just a \code{\link[ggplot2:ggproto]{ggproto()}}
object, descended from the top-level \code{Align}, and each implements various
methods and fields.

To create a new type of \verb{Align*} object, you typically will want to
override one or more of the following:
\itemize{
\item \code{setup_params}: Prepare parameter or check parameters used by this plot.
\item \code{setup_data}: Prepare data used by this plot.
\item \code{compute}: A method used to compute statistics.
\item \code{layout}: A method used to group observations into panel or reorder
observations.
\item \code{draw}: A method used to draw the plot. Must return a \code{ggplot} object.
}
}
\section{Aligned Axis}{

It is important to note that we consider rows as observations, meaning
\code{NROW(data)} must match the number of observations along the specified layout
axis (x-axis for a vertical stack layout, y-axis for a horizontal stack
layout).
\itemize{
\item \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}}: for column annotation, the
\verb{layout matrix} will be transposed before using (If data is a function, it
will be applied with the transposed matrix). This is necessary because
column annotation uses columns as observations, but we need rows.
\item \code{\link[=stack_layout]{stack_layout()}}: the \verb{layout matrix} will be used as it is since we
place all plots along a single axis.
}
}

\examples{
align_gg()
align_dendro()

}
\keyword{datasets}
\keyword{internal}
