% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-.R
\docType{data}
\name{align}
\alias{align}
\alias{Align}
\title{Create a new \code{Align} object}
\usage{
align(
  align_class,
  params,
  data,
  size = NULL,
  action = NULL,
  action_data = NA,
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated(),
  check.param = TRUE,
  call = caller_call()
)
}
\arguments{
\item{data}{A matrix, data frame, or a simple vector. If an atomic vector is
provided, it will be converted into a one-column matrix. When \code{data = NULL},
the internal \code{layout} data will be used by default. Additionally, \code{data} can
be a function (including purrr-like lambdas), which will be applied to the
\code{layout} data.

It is important to note that we consider the \code{rows} as the observations. It
means the \code{NROW(data)} must return the same number with the specific \code{layout}
axis (meaning the x-axis for vertical stack layout, or y-axis for horizontal
stack layout).
\itemize{
\item \code{heatmap_layout()}: for column annotation, the \code{layout} data will be
transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix). This is necessary because column annotation uses
heatmap columns as observations, but we need rows.
\item \code{stack_layout()}: the \code{layout} data will be used as it is since we place
all plots along a single axis.
}}

\item{size}{Plot size, can be an \link[grid:unit]{unit} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot.}

\item{action_data}{Default action data, which by default tries to infer
from the \code{data} argument.}

\item{limits}{A boolean value indicates whether to set the layout limtis for
the plot.}

\item{facet}{A boolean value indicates whether to set the layout facet for
the plot. If this is \code{FALSE}, \code{limits} will always be \code{FALSE} too.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current plot. If \code{TRUE}, all subsequent ggplot elements will be
added into this plot.}

\item{order}{An single integer for the plot area order.}

\item{name}{A string of the plot name. Used to switch the active context in
\code{\link[=hmanno]{hmanno()}} or \code{\link[=stack_active]{stack_active()}}.}

\item{free_guides}{Override the \code{guides} collection behavior specified in the
heatmap layout for the annotation stack layout. Only used when \code{position} is
a string.}

\item{free_spaces}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{plot_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{theme}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_labs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{check.param}{A single boolean value indicates whether to check the
supplied parameters and warn.}

\item{call}{The \code{call} used to construct the \code{Align} for reporting messages.}
}
\value{
A new \code{Align} object.
}
\description{
\code{Align} object will act with the \code{layout} object, reorder or split the
observations, some of them can also add plot components into the \code{layout}
object.
}
\section{Align}{

Each of the \verb{Align*} objects is just a \link[ggplot2:ggproto]{ggproto()} object,
descended from the top-level \code{Align}, and each implements various
methods and fields.

To create a new type of \verb{Align*} object, you typically will want to
override one or more of the following:
\itemize{
\item \code{setup_params}: Prepare parameter or check parameters used by this
annotation.
\item \code{setup_data}: Prepare data used by this annotation.
\item \code{compute}: A method used to compute statistics.
\item \code{layout}: A method used to group heamap rows/columns into panel or
reorder heamtap rows/columns.
\item \code{draw}: A method used to draw the plot. Must return a \code{ggplot} object.
}
}

\examples{
align_gg()
align_dendro()
}
\keyword{datasets}
\keyword{internal}
