% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-gg.R
\name{align_gg}
\alias{align_gg}
\alias{ggalign}
\title{Create ggplot object with a customized data}
\usage{
align_gg(
  data = waiver(),
  mapping = aes(),
  size = NULL,
  action = NULL,
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated()
)

ggalign(
  data = waiver(),
  mapping = aes(),
  size = NULL,
  action = NULL,
  limits = TRUE,
  facet = TRUE,
  set_context = TRUE,
  order = NULL,
  name = NULL,
  free_guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated()
)
}
\arguments{
\item{data}{A matrix, data frame, or a simple vector. If an atomic vector is
provided, it will be converted into a one-column matrix. When \code{data = NULL},
the internal \code{layout} data will be used by default. Additionally, \code{data} can
be a function (including purrr-like lambdas), which will be applied to the
\code{layout} data.

It is important to note that we consider the \code{rows} as the observations. It
means the \code{NROW(data)} must return the same number with the specific \code{layout}
axis (meaning the x-axis for vertical stack layout, or y-axis for horizontal
stack layout).
\itemize{
\item \code{heatmap_layout()}: for column annotation, the \code{layout} data will be
transposed before using (If data is a \code{function}, it will be applied with
the transposed matrix). This is necessary because column annotation uses
heatmap columns as observations, but we need rows.
\item \code{stack_layout()}: the \code{layout} data will be used as it is since we place
all plots along a single axis.
}}

\item{mapping}{Additional default list of aesthetic mappings to use for plot.}

\item{size}{Plot size, can be an \link[grid:unit]{unit} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot.}

\item{limits}{A boolean value indicates whether to set the layout limtis for
the plot.}

\item{facet}{A boolean value indicates whether to set the layout facet for
the plot. If this is \code{FALSE}, \code{limits} will always be \code{FALSE} too.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current plot. If \code{TRUE}, all subsequent ggplot elements will be
added into this plot.}

\item{order}{An single integer for the plot area order.}

\item{name}{A string of the plot name. Used to switch the active context in
\code{\link[=hmanno]{hmanno()}} or \code{\link[=stack_active]{stack_active()}}.}

\item{free_guides}{Override the \code{guides} collection behavior specified in the
heatmap layout for the annotation stack layout. Only used when \code{position} is
a string.}

\item{free_spaces}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{plot_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{theme}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_labs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}
}
\value{
A \code{AlignGG} object.
}
\description{
\code{ggalign} is just an alias of \code{align_gg}.
}
\section{ggplot2 specification}{

\code{align_gg} initializes a ggplot \code{data} and \code{mapping}.

\code{align_gg()} always applies a default mapping for the axis of the data index
in the layout. This mapping is \code{aes(y = .data$.y)} for horizontal stack
layout (including left and right heatmap annotation) and \code{aes(x = .data$.x)}
for vertical stack layout (including top and bottom heatmap annotation).

For ggplot usage, matrix (including a simple vector) data is converted into a
long-format data frame. The data in the underlying \code{ggplot} object will
contain following columns:
\itemize{
\item \code{.panel}: the panel for current layout axis. It means \code{x-axis} for
vertical stack layout, \code{y-axis} for horizontal stack layout.
\item \code{.index}: the index of the original layout data (only applicable if \code{data}
is \code{NULL}).
\item \code{.x} or \code{.y}: the \code{x} or \code{y} coordinates
\item \code{.row_names} and \code{.row_index}: A factor of the row names and an integer of
row index of the original matrix or data frame.
\item \code{.column_names} and \code{.column_index}: the column names and column index of
the original matrix (only applicable if \code{data} is a \code{matrix}).
\item \code{value}: the actual matrix value  (only applicable if \code{data} is a
\code{matrix}).
}

In the case where the input data is already a data frame, three additional
columns (\code{.row_names}, \code{.row_index}, and \code{.panel}) are added to the data
frame.

If the data is inherit from \code{\link[=heatmap_layout]{heatmap_layout()}}, an additional column will be
added.
\itemize{
\item \code{.extra_panel}: the panel information for column (left or right
annotation) or row (top or bottom annotation).
}
}

\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    ggalign() +
    geom_point(aes(y = value))

# if data is `NULL`, a three column data frame
# will be created (`.panel`, `.index`, `.x`/`.y`)
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("t", size = 0.5) +
    align_dendro(k = 3L) +
    ggalign(data = NULL, size = 0.2) +
    geom_tile(aes(y = 1L, fill = .panel))

}
