% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align-group.R
\name{align_group}
\alias{align_group}
\title{Group and align layout based on categorical or factor levels.}
\usage{
align_group(
  group,
  context = NULL,
  set_context = deprecated(),
  name = deprecated()
)
}
\arguments{
\item{group}{A character define the groups of the observations.}

\item{context}{A \code{\link[=context]{context()}} object that defines the active context when
added to a layout.}

\item{set_context}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{context}
argument instead.}

\item{name}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{context} argument
instead.}
}
\value{
A \code{"AlignGroup"} object.
}
\description{
Group and align layout based on categorical or factor levels.
}
\examples{
set.seed(1L)
small_mat <- matrix(rnorm(81), nrow = 9)
ggheatmap(small_mat) +
    anno_top() +
    align_group(sample(letters[1:4], ncol(small_mat), replace = TRUE))
}
