% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign.R
\name{align_panel}
\alias{align_panel}
\alias{ggpanel}
\title{Create ggplot object with layout panel data}
\usage{
align_panel(...)

ggpanel(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because we realised that it's a special case of
the \code{\link[=ggalign]{ggalign()}} function. Please use \code{ggalign(data = NULL)} instead.
}
\keyword{internal}
