% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{alignpatch}
\alias{alignpatch}
\title{Prepare plots to be aligned with \code{align_plots}}
\usage{
alignpatch(x)
}
\arguments{
\item{x}{A plot object to be prepared for alignment.}
}
\value{
A \code{ggalign::Patch} object.
}
\description{
Prepare plots to be aligned with \code{align_plots}
}
\details{
\code{ggalign} has implement \code{alignpatch} method for following objects:
\itemize{
\item \code{\link[ggplot2:ggplot]{ggplot}}
\item \code{\link[=align_plots]{alignpatches}}
\item \code{\link[=ggwrap]{wrapped_plot}}
\item \code{\link[patchwork:patchGrob]{patch}}
\item \code{\link[patchwork:wrap_elements]{wrapped_patch}}
\item \code{\link[patchwork:plot_spacer]{spacer}}
}
}
\examples{
alignpatch(ggplot())
}
\seealso{
\code{\link[=align_plots]{align_plots()}}
}
\keyword{internal}
