% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{alignpatch}
\alias{alignpatch}
\title{Get Patch representation}
\usage{
alignpatch(x)
}
\arguments{
\item{x}{Any objects has a Patch representation}
}
\value{
A \code{ggalign::Patch}
object.
}
\description{
Get Patch representation
}
\details{
\code{ggalign} has implement \code{alignpatch} method for following objects:
\itemize{
\item \code{\link[ggplot2:ggplot]{ggplot}}
\item \code{\link{alignpatches}}
\item \code{\link[=ggwrap]{wrapped_plot}}
\item \code{\link[patchwork:patchGrob]{patchwork::patchGrob}}
\item \code{\link[patchwork:wrap_elements]{patchwork::wrap_elements}}
\item \code{\link[patchwork:plot_spacer]{patchwork::plot_spacer}}
}
}
\examples{
alignpatch(ggplot())
}
\seealso{
\code{\link[=align_plots]{align_plots()}}
}
\keyword{internal}
