% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{alignpatch}
\alias{alignpatch}
\title{Prepare plots to be aligned with \code{align_plots}}
\usage{
alignpatch(x)
}
\arguments{
\item{x}{A plot object to be prepared for alignment.
\code{ggalign} has implement \code{alignpatch} method for following objects:
\itemize{
\item \link[ggplot2:ggplot]{ggplot}
\item \link[=align_plots]{alignpatches}
\item \link[=ggwrap]{wrapped_plot}
\item \link[patchwork:patchGrob]{patch}
\item \link[patchwork:wrap_elements]{wrapped_patch}
\item \link[patchwork:plot_spacer]{spacer}
}}
}
\value{
A \code{Patch} object.
}
\description{
Prepare plots to be aligned with \code{align_plots}
}
\examples{
alignpatch(ggplot())
}
\seealso{
\link{align_plots}
}
\keyword{internal}
