% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-.R
\name{anno}
\alias{anno}
\alias{show,anno-method}
\title{Create a heatmap annotation object}
\usage{
anno(
  Class,
  ...,
  data = NULL,
  position = NULL,
  size = NULL,
  set_context = NULL,
  order = NULL,
  name = NULL
)

\S4method{show}{anno}(object)
}
\arguments{
\item{Class}{Sub-class of \code{anno}.}

\item{...}{Additional components of the sub-class.}

\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{ "bottom"}, and \code{"right"}. If \code{NULL}, the active context
of the \link[=ggheat]{ggheatmap} will be used.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{object}{A \code{anno} object.}
}
\value{
A new \code{Class} object, ``.
}
\description{
Create a heatmap annotation object
}
