% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{build_patchwork}
\alias{build_patchwork}
\title{Build \code{Layout} object for rendering.}
\usage{
build_patchwork(layout)
}
\arguments{
\item{layout}{A \code{Layout} object.}
}
\value{
A \code{patchwork} object.
}
\description{
Build \code{Layout} object for rendering.
}
\examples{
build_patchwork(ggheatmap(matrix(rnorm(100L), nrow = 10L)))
}
