% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-circle-.R
\name{circle_layout}
\alias{circle_layout}
\title{Arrange Plots in a Circular Layout}
\usage{
circle_layout(data = NULL, ..., limits = waiver())
}
\arguments{
\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout:
\itemize{
\item If \code{limits} is not provided, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to get a
matrix.
\item If \code{limits} is specified, \code{\link[=fortify_data_frame]{fortify_data_frame()}} will be used to get a
data frame. Note that if the data is a \code{matrix}, it will be automatically
converted to a long-formatted data frame, which differs from \code{ggplot2}'s
behavior.
}}

\item{...}{Additional arguments passed to \code{\link[=fortify_data_frame]{fortify_data_frame()}}.}

\item{limits}{A \code{\link[=continuous_limits]{continuous_limits()}} object specifying the left/lower
limit and the right/upper limit of the scale. Used to align the continuous
axis.}
}
\value{
A \code{CircleLayout} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function integrates the functionalities of \code{circle_discrete()} and
\code{circle_continuous()} into a single interface.
}
\examples{
set.seed(123)
small_mat <- matrix(rnorm(56), nrow = 7L)

# circle_discrete
circle_layout(matrix(rnorm(56), nrow = 7L)) +
    ggalign() +
    geom_tile(aes(y = .column_index, fill = value)) +
    scale_fill_viridis_c() +
    align_dendro(aes(color = branch), k = 3) +
    circle_switch(coord_radial(inner.radius = 0.5))

# circle_continuous
circle_layout(mpg, limits = continuous_limits(c(3, 5))) +
    ggalign(mapping = aes(displ, hwy, colour = class)) +
    geom_point(size = 2) +
    ggalign(mapping = aes(displ, hwy, colour = class)) +
    geom_point(size = 2) &
    scale_color_brewer(palette = "Dark2") &
    theme_bw()
}
\seealso{
\itemize{
\item \code{\link[=circle_discrete]{circle_discrete()}}
\item \code{\link[=circle_continuous]{circle_continuous()}}
}
}
