% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-cross-.R
\name{cross_align}
\alias{cross_align}
\alias{cross_alignv}
\alias{cross_alignh}
\title{Arrange Plots Horizontally or Vertically}
\usage{
cross_align(data = NULL, direction, ..., theme = NULL, sizes = NA)

cross_alignv(data = NULL, ...)

cross_alignh(data = NULL, ...)
}
\arguments{
\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout:
\itemize{
\item For \code{stack_free}, \code{\link[=fortify_data_frame]{fortify_data_frame()}} will be used to convert data to
a data frame.
\item For \code{stack_align}, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to convert data to a
matrix.
}}

\item{direction}{A string indicating the direction of the stack layout,
either \code{"horizontal"} or \code{"vertical"}.}

\item{...}{Additional arguments passed to \code{\link[=fortify_data_frame]{fortify_data_frame()}} or
\code{\link[=fortify_matrix]{fortify_matrix()}}.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} used to render the \code{guides},
\code{title}, \code{subtitle}, \code{caption}, \code{margins}, \code{patch.title}, \code{panel.border}, and
\code{background}. If \code{NULL} (default), will inherit from the parent \code{layout}.}

\item{sizes}{A numeric or a \code{\link[grid:unit]{unit}} object of length \code{3}
indicating the relative heights (for \code{direction = "horizontal"}) or widths
(for \code{direction = "vertical"}). This is only used if you include a nested
\code{\link[=quad_layout]{quad_layout()}} in the layout.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{cross_align} function aligns observations, and allow different layourt
ordering index in a single layout.

Two aliases are provided for convenience:
\itemize{
\item \code{cross_alignv}: A special case of \code{cross_align} that sets \code{direction = "vertical"}.
\item \code{cross_alignh}: A special case of \code{cross_align} that sets \code{direction = "horizontal"}.
}
}
\seealso{
\code{\link[=cross_link]{cross_link()}}
}
