% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram_data}
\alias{dendrogram_data}
\title{Dengrogram x and y coordinates}
\usage{
dendrogram_data(
  tree,
  double_spanned_horizontal = FALSE,
  center = FALSE,
  type = "rectangle",
  leaf_pos = NULL,
  leaf_braches = NULL,
  branch_gap = NULL,
  root = NULL
)
}
\arguments{
\item{tree}{A \link[stats:hclust]{hclust} or a \link[stats:dendrogram]{dendrogram}
object.}

\item{double_spanned_horizontal}{A boolean value indicates whether double the
horizontal lines span across multiple panels. This is useful if you want to
do facet operation in \code{ggplot2}.}

\item{center}{A boolean value. if \code{TRUE}, nodes are plotted centered with
respect to the leaves in the branch. Otherwise (default), plot them in the
middle of all direct child nodes.}

\item{type}{A string indicates the plot type, \code{"rectangle"} or \code{"triangle"}.}

\item{leaf_pos}{The x-coordinates of the leaf node. Must be the same length
of the number of observations in \code{tree}.}

\item{leaf_braches}{Branches of the leaf node. Must be the same length of the
number of observations in \code{tree}. Usually come from \link[stats:cutree]{cutree}.}

\item{branch_gap}{A numeric value indicates the gap between different
branches. If a \link{unit} object is provided, it'll convert into \code{native} values.}

\item{root}{A length one string or numeric indicates the root branch.}
}
\value{
A list of 2 data.frame. One for node coordinates, another for edge
coordinates.
\code{node} and tree segments \code{edge} contains following columns:
\itemize{
\item \code{index}: the original index in the tree for the current node
\item \code{label}: node label text
\item \code{x} and \code{y}: x-axis and y-axis coordinates for current node or the start
node of the current edge.
\item \code{xend} and \code{yend}: the x-axis and y-axis coordinates of the terminal node
for current edge.
\item \code{branch}: which branch current node is. You can use this column to color
different groups.
\item \code{panel}: which panel current node is, if we split the plot into panels
using \link[ggplot2:facet_grid]{facet_grid}, this column will show
which panel current node or edge is from. Note: some nodes
may fall outside panels, so there are possible \code{NA} values in
this column. we also provide \code{.panel} column, which always
regard the right branch for the out-sided \code{node}.
\item \code{.panel}: See \code{panel}.
\item \code{leaf}: A logical value indicates whether current node is a leaf.
}
}
\description{
Dengrogram x and y coordinates
}
\examples{
dendrogram_data(hclust(dist(USArrests), "ave"))
}
