% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{.mark_draw}
\alias{.mark_draw}
\title{Define the links to connect the marked observations}
\usage{
.mark_draw(.draw, ..., .group1 = NULL, .group2 = NULL)
}
\arguments{
\item{.draw}{A function used to draw the links. The function must return a
\code{\link[grid:grid.grob]{grob()}} object. If the function does not return a valid
\code{grob}, nothing will be drawn. The input data for the function contains a
list, where each item is a list of two data frames: one for the panel side
coordinates (\code{"panel"}) and one for the marked observations coordinates
(\code{"link"}).}

\item{...}{A list of formula, each side of the formula should be an integer
or character index, or a \code{link_range()} object to define the linked
observations. For integer indices, you can wrap them with \code{\link[=I]{I()}} to
indicate the order based on the layout. You can also use \code{waiver()} to
inherit the values from the opposite link argument.}

\item{.group1, .group2}{A single boolean value indicating whether to use the
panel group information from the layout as the links. By default, will
guess from the layout.}
}
\description{
A base version of \code{\link{mark_draw}}, designed for performance optimization. This
function is used to build other \verb{mark_*} functions that manage the drawing of
links between marked observations.
}
