% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{.mark_draw}
\alias{.mark_draw}
\title{Define the links to connect the marked observations}
\usage{
.mark_draw(.draw, ..., .group1 = NULL, .group2 = NULL)
}
\arguments{
\item{.draw}{A function used to draw the links. The function must return a
\code{\link[grid:grid.grob]{grob()}} object. If the function does not return a valid
\code{grob}, nothing will be drawn. The input data for the function contains a
list, where each item is a list of two data frames: one for the panel side
coordinates (\code{"panel"}) and one for the marked observations coordinates
(\code{"link"}).}

\item{...}{A list of formulas, where each side of the formula should be an
\code{integer} or \code{character} index, or a \code{range_link()} object defining the
linked observations. Use \code{NULL} to indicate no link on that side. You can
also combine these by wrapping them into a single \code{list()}. If only the
left-hand side of the formula exists, you can input it directly. For
integer indices, wrap them with \code{\link[=I]{I()}} to preserve their order according
to the layout. You can also use \code{\link[=waiver]{waiver()}} to inherit values from the
opposite link argument.}

\item{.group1, .group2}{A single boolean value indicating whether to use the
panel group information from the layout as the links. By default, will
guess from the layout.}
}
\description{
A base version of \code{\link{mark_draw}}, designed for performance optimization. This
function is used to build other \verb{mark_*} functions that manage the drawing of
links between marked observations.
}
