% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data_frame.R
\name{fortify_data_frame.character}
\alias{fortify_data_frame.character}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{character}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
When \code{data} is an atomic vector, it'll be converted to a data frame with
following columns:
\itemize{
\item \code{.names}: the names for the vector (only applicable if names exist).
\item \code{value}: the actual value of the vector.
}
}
\seealso{
Other \code{\link[=fortify_data_frame]{fortify_data_frame()}} methods: 
\code{\link{fortify_data_frame.default}()},
\code{\link{fortify_data_frame.matrix}()},
\code{\link{fortify_data_frame.numeric}()}
}
\concept{\code{\link[=fortify_data_frame]{fortify_data_frame()}} methods}
