% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data-frame-.R
\name{fortify_data_frame.default}
\alias{fortify_data_frame.default}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{default}(data, ..., data_arg = NULL, call = NULL)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:fortify]{fortify()}}.}

\item{data_arg}{The argument name for \code{data}. Developers can use it to
improve messages. Not used by the user.}

\item{call}{The execution environment where \code{data} and other arguments for
the method are collected. Developers can use it to improve messages. Not
used by the user.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts various objects to a data frame.
}
\details{
By default, it calls \code{\link[ggplot2:fortify]{fortify()}} to build the
data frame.
}
\seealso{
Other \code{\link[=fortify_data_frame]{fortify_data_frame()}} methods:
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.dendrogram}()},
\code{\link{fortify_data_frame.matrix}()},
\code{\link{fortify_data_frame.phylo}()}
}
\concept{fortify_data_frame}
