% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data-frame.R
\name{fortify_data_frame.matrix}
\alias{fortify_data_frame.matrix}
\alias{fortify_data_frame.DelayedMatrix}
\alias{fortify_data_frame.Matrix}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{matrix}(data, ...)

\method{fortify_data_frame}{DelayedMatrix}(data, ...)

\method{fortify_data_frame}{Matrix}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts various objects to a data frame.
}
\details{
Matrix will be transformed into a long-form data frame, where each row
represents a unique combination of matrix indices and their corresponding
values. The resulting data frame will contain the following columns:
\itemize{
\item \code{.row_names} and \code{.row_index}: the row names (only applicable when names
exist) and an integer representing the row index of the original matrix.
\item \code{.column_names} and \code{.column_index}: the column names (only applicable
when names exist) and column index of the original matrix.
\item \code{value}: the actual value.
}
}
\seealso{
Other fortify_data_frame methods: 
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.complex}()},
\code{\link{fortify_data_frame.default}()},
\code{\link{fortify_data_frame.logical}()},
\code{\link{fortify_data_frame.numeric}()}
}
\concept{fortify_data_frame methods}
