% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data_frame.R
\name{fortify_data_frame.matrix}
\alias{fortify_data_frame.matrix}
\title{Build a data frame}
\usage{
\method{fortify_data_frame}{matrix}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
When \code{data} is a matrix, it will automatically be transformed into a
long-form data frame, where each row represents a unique combination of
matrix indices and their corresponding values.
}
\seealso{
Other fortify_data_frame methods: 
\code{\link{fortify_data_frame.character}()},
\code{\link{fortify_data_frame.default}()},
\code{\link{fortify_data_frame.numeric}()}
}
\concept{fortify_data_frame methods}
