% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-data_frame.R
\name{fortify_data_frame}
\alias{fortify_data_frame}
\title{Build a data frame}
\usage{
fortify_data_frame(data, ...)
}
\arguments{
\item{data}{Any objects to be converted to a data frame.}

\item{...}{Arguments passed to methods.}
}
\value{
A data frame.
}
\description{
This function converts various objects to a data frame. By default, it calls
\code{\link[ggplot2:fortify]{fortify()}} to perform the conversion.
}
\details{
When \code{data} is a matrix, it will automatically be transformed into a
long-form data frame, where each row represents a unique combination of
matrix indices and their corresponding values.
}
\seealso{
\code{\link[ggplot2:fortify]{fortify}}
}
