% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-data.R
\name{fortify_heatmap.MAF}
\alias{fortify_heatmap.MAF}
\title{Build data for the heatmap layout}
\usage{
\method{fortify_heatmap}{MAF}(
  data,
  ...,
  genes = NULL,
  n_top = NULL,
  remove_empty_samples = TRUE
)
}
\arguments{
\item{data}{Any objects to be plot with \code{\link[=ggheatmap]{ggheatmap()}}.}

\item{...}{Not used currently.}

\item{genes}{An atomic character defines the genes to draw.}

\item{n_top}{A single number indicates how many top genes to be drawn.}
}
\value{
A matrix used by \code{\link[=ggheatmap]{ggheatmap()}}.
}
\description{
Build data for the heatmap layout
}
