% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix.default}
\alias{fortify_matrix.default}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{default}(data, ...)
}
\arguments{
\item{data}{An object to be converted to a matrix.}

\item{...}{Arguments passed to methods.}
}
\value{
A matrix.
}
\description{
By default, it calls \code{\link[=as.matrix]{as.matrix()}} to build a matrix.
}
\seealso{
Other \code{\link[=fortify_matrix]{fortify_matrix()}} methods: 
\code{\link{fortify_matrix.MAF}()}
}
\concept{\code{\link[=fortify_matrix]{fortify_matrix()}} methods}
