% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix.GISTIC}
\alias{fortify_matrix.GISTIC}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{GISTIC}(
  data,
  ...,
  n_top = NULL,
  bands = NULL,
  ignored_bands = NULL,
  sample_anno = NULL,
  remove_empty_samples = TRUE
)
}
\arguments{
\item{data}{An object to be converted to a matrix.}

\item{...}{Not used currently.}

\item{n_top}{A single number indicates how many top bands to be drawn.}

\item{bands}{An atomic character defines the bands to draw.}

\item{ignored_bands}{An atomic character defines the bands to be ignored.}

\item{sample_anno}{A data frame of sample clinical features to be added.}

\item{remove_empty_samples}{A single boolean value indicating whether to drop
samples without any genomic alterations.}
}
\value{
A matrix.
}
\description{
This function converts various objects into a matrix format.
}
\section{ggalign attributes}{

\itemize{
\item \code{sample_anno}: sample clinical informations provided in \code{sample_anno}.
\item \code{sample_summary}: sample copy number summary informations. See
\code{data@cnv.summary} for details.
\item \code{cytoband_summary}: cytoband summary informations. See
\code{data@cytoband.summary} for details.
\item \code{gene_summary}: gene summary informations. See
\code{data@gene.summary} for details.
\item \code{summary}: A data frame of summary information. See \code{data@summary} for
details.
}
}

\seealso{
Other fortify_matrix methods: 
\code{\link{fortify_matrix.MAF}()},
\code{\link{fortify_matrix.default}()}
}
\concept{fortify_matrix methods}
