% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix.MAF}
\alias{fortify_matrix.MAF}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{MAF}(
  data,
  ...,
  genes = NULL,
  n_top = NULL,
  remove_empty_samples = TRUE,
  collapse_vars = TRUE,
  use_syn = TRUE
)
}
\arguments{
\item{data}{An object to be converted to a matrix.}

\item{...}{Not used currently.}

\item{genes}{An atomic character defines the genes to draw.}

\item{n_top}{A single number indicates how many top genes to be drawn.}

\item{remove_empty_samples}{A single boolean value indicating whether to drop
samples without any genomic alterations.}

\item{collapse_vars}{A single boolean value indicating whether to collapse
multiple alterations in the same sample and gene into a single value
\code{"Multi_Hit"}. Alternatively, you can provide a single string indicates the
collapsed values.}

\item{use_syn}{A single boolean value indicates whether to include synonymous
variants when Classifies SNPs into transitions and transversions.}
}
\value{
A matrix.
}
\description{
By default, it calls \code{\link[=as.matrix]{as.matrix()}} to build a matrix.
}
\section{ggalign attributes}{

\itemize{
\item \code{gene_summary}: gene summary informations. See
\code{maftools::getGeneSummary()} for details.
\item \code{sample_summary}: sample summary informations. See
\code{maftools::getSampleSummary()} for details.
\item \code{sample_anno}: sample clinical informations. See
\code{maftools::getClinicalData()} for details.
\item \code{n_genes}: Total of genes.
\item \code{n_samples}: Total of samples.
\item \code{titv}: A list of \code{data.frames} with Transitions and Transversions
summary. See \code{maftools::titv()} for details.
}
}

\seealso{
Other fortify_matrix methods: 
\code{\link{fortify_matrix.default}()}
}
\concept{fortify_matrix methods}
