% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-oncoplot.R
\name{fortify_matrix.MAF}
\alias{fortify_matrix.MAF}
\title{Build a Matrix}
\usage{
\method{fortify_matrix}{MAF}(
  data,
  ...,
  genes = NULL,
  n_top = NULL,
  remove_empty_samples = TRUE,
  collapse_vars = TRUE,
  use_syn = TRUE,
  data_arg = caller_arg(data),
  call = NULL
)
}
\arguments{
\item{data}{A \code{\link[maftools:read.maf]{MAF}} object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{genes}{An atomic character defines the genes to draw.}

\item{n_top}{A single number indicates how many top genes to be drawn.}

\item{remove_empty_samples}{A single boolean value indicating whether to drop
samples without any genomic alterations.}

\item{collapse_vars}{A single boolean value indicating whether to collapse
multiple alterations in the same sample and gene into a single value
\code{"Multi_Hit"}. Alternatively, you can provide a single string indicates the
collapsed values.}

\item{use_syn}{A single boolean value indicates whether to include synonymous
variants when Classifies SNPs into transitions and transversions.}

\item{data_arg}{The argument name for \code{data}. Developers can use it to
improve messages. Not used by the user.}

\item{call}{The execution environment where \code{data} and other arguments for
the method are collected, e.g., \code{\link[rlang:stack]{caller_env()}}.
Developers can use it to improve messages. Not used by the user.}
}
\value{
A matrix.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function converts various objects into a matrix format.
}
\section{ggalign attributes}{

\itemize{
\item \code{gene_summary}: gene summary informations. See
\code{maftools::getGeneSummary()} for details.
\item \code{sample_summary}: sample summary informations. See
\code{maftools::getSampleSummary()} for details.
\item \code{sample_anno}: sample clinical informations. See
\code{maftools::getClinicalData()} for details.
\item \code{n_genes}: Total of genes.
\item \code{n_samples}: Total of samples.
\item \code{titv}: A list of \code{data.frames} with Transitions and Transversions
summary. See \code{maftools::titv()} for details.
}
}

\seealso{
Other fortify_matrix methods: 
\code{\link{fortify_matrix.GISTIC}()},
\code{\link{fortify_matrix.default}()},
\code{\link{fortify_matrix.phylo}()}
}
\concept{fortify_matrix methods}
