% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix-matrix.R
\name{fortify_matrix.matrix}
\alias{fortify_matrix.matrix}
\title{Build a matrix}
\usage{
\method{fortify_matrix}{matrix}(data, ..., data_arg = NULL, call = NULL)
}
\arguments{
\item{data}{A matrix object.}

\item{...}{These dots are for future extensions and must be empty.}

\item{data_arg}{The argument name for \code{data}. Developers can use it to
improve messages. Not used by the user.}

\item{call}{The execution environment where \code{data} and other arguments for
the method are collected. Developers can use it to improve messages. Not
used by the user.}
}
\description{
Build a matrix
}
\section{shape}{

\itemize{
\item \code{upset}: \code{\link[=fortify_matrix.matrix_upset]{fortify_matrix.matrix_upset()}}
\item \code{oncoplot}: \code{\link[=fortify_matrix.matrix_oncoplot]{fortify_matrix.matrix_oncoplot()}}
}
}

\seealso{
Other fortify_matrix methods: 
\code{\link{fortify_matrix.GISTIC}()},
\code{\link{fortify_matrix.MAF}()},
\code{\link{fortify_matrix.default}()},
\code{\link{fortify_matrix.list_upset}()},
\code{\link{fortify_matrix.matrix_oncoplot}()},
\code{\link{fortify_matrix.matrix_upset}()},
\code{\link{fortify_matrix.phylo}()}
}
\concept{fortify_matrix methods}
