% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix.R
\name{fortify_matrix}
\alias{fortify_matrix}
\title{Build a Matrix}
\usage{
fortify_matrix(data, ...)
}
\arguments{
\item{data}{An object to be converted to a matrix.}

\item{...}{Arguments passed to methods.}
}
\value{
A matrix.
}
\description{
This function converts various objects into a matrix format.
}
\seealso{
\itemize{
\item \code{\link[=fortify_matrix.default]{fortify_matrix.default()}}
\item \code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}}
\item \code{\link[=fortify_matrix.GISTIC]{fortify_matrix.GISTIC()}}
}
}
