% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-free-align.R,
%   R/alignpatch-free-border.R, R/alignpatch-free-lab.R,
%   R/alignpatch-free-size.R
\name{free_align}
\alias{free_align}
\alias{free_border}
\alias{free_lab}
\alias{free_size}
\title{Free from alignment}
\usage{
free_align(plot, axes = c("t", "l", "b", "r"))

free_border(plot, borders = c("t", "l", "b", "r"))

free_lab(plot, labs = c("t", "l", "b", "r"))

free_size(plot, ...)
}
\arguments{
\item{plot}{A \link[ggplot2:ggplot]{ggplot} or \link[=plot_grid]{alignpatches} object.}

\item{axes}{Which axes shouldn't be aligned? Allowed values: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}

\item{borders}{Which border shouldn't be aligned? Allowed values: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}

\item{labs}{Which axis labs to be free?}

\item{...}{What ggplot2 elements to remove?}
}
\value{
\itemize{
\item \code{free_align}: A modified version of \code{plot} with a \code{free_align} class.
}

\itemize{
\item \code{free_border}: A modified version of \code{plot} with a \code{free_border} class.
}

\itemize{
\item \code{free_lab}: A modified version of \code{plot} with a \code{free_lab} class.
}

\itemize{
\item \code{free_size}: A modified version of \code{plot} with a \code{free_size} class.
}
}
\description{
\link{plot_grid} will try to align plots by their panels, and align every elements
of the plot, following functions romove this restrictions.
\itemize{
\item \code{free_align}: if we want to compose plots without alignment of some panel
axes (panel won't be aligned). we can wrap the plot with \code{free_align}.
\item \code{free_border}: If we want to compose plots without alignment of the panel
borders (but still align the panels themselves), we can wrap the plot with
\code{free_border}.
\item \code{free_lab}: If we want to compose plots without alignment of the axis
title, we can wrap the plot with \code{free_lab}.
\item \code{free_size}: Removing the ggplot element sizes when aligning
}
}
