% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/free-.R
\name{free_gg}
\alias{free_gg}
\alias{ggfree}
\title{Add ggplot to layout}
\usage{
free_gg(x = waiver(), ..., size = NULL, action = NULL, context = NULL)
}
\arguments{
\item{x}{A dataset used to initialize a \code{\link[ggplot2:ggplot]{ggplot}} object.
Alternatively, a pre-defined \code{\link[ggplot2:ggplot]{ggplot}} object can be
provided directly. By default, it will inherit from the parent layout if
applicable.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggplot]{ggplot()}}.}

\item{size}{Plot size, specified as a \code{\link[grid:unit]{unit}} object.}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used for the plot. By default,
it inherits from the parent layout.}

\item{context}{A \code{\link[=plot_context]{plot_context()}} object that defines the action when
added to a layout.}
}
\value{
A \code{free_gg} object.
}
\description{
\code{ggfree} is an alias for \code{free_gg}.
}
\examples{
ggheatmap(matrix(rnorm(56), nrow = 7)) +
    anno_top() +
    ggfree(mtcars, aes(wt, mpg)) +
    geom_point()
}
