% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign-package.R
\docType{package}
\name{ggalign-package}
\alias{ggalign-package}
\title{ggalign: Align Multiple 'ggplot' Objects}
\description{
This package is an extension of ggplot2 that aligns multiple ggplot objects along their axes in the same order. It is highly beneficial for organizing plots that involve changing the order of data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/ggalign}
  \item Report bugs at \url{https://github.com/Yunuuuu/ggalign/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
