% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign-package.R
\docType{package}
\name{ggalign-package}
\alias{ggalign-package}
\title{ggalign: Align Multiple 'ggplot' Objects}
\description{
This package extends ggplot2 and provides numerous benefits for organizing and arranging plots. It is specifically designed to align a specific axis of multiple ggplot objects in a consistent order. This functionality is particularly useful for plots that require manipulation of data order. A common plot combination that can be effectively organized using this package includes a dendrogram and a heatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/ggalign}
  \item Report bugs at \url{https://github.com/Yunuuuu/ggalign/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
