% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign-package.R
\docType{package}
\name{ggalign-package}
\alias{ggalign-package}
\title{ggalign: Align Multiple 'ggplot' Objects}
\description{
This package extends ggplot2 and offers various tools for organizing and arranging plots. It is designed to consistently align a specific axis across multiple ggplot objects, making it especially useful for plots requiring data order manipulation. A typical use case includes organizing combinations like a dendrogram and a heatmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/ggalign}
  \item \url{https://yunuuuu.github.io/ggalign/}
  \item Report bugs at \url{https://github.com/Yunuuuu/ggalign/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID}) [copyright holder]

}
\keyword{internal}
