% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggalign-package.R
\docType{package}
\name{ggalign-package}
\alias{ggalign-package}
\title{ggalign: A 'ggplot2' Extension for Consistent Axis Alignment}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2' extension providing an integrative framework for composable visualization, enabling the creation of complex multi-plot layouts such as insets, circular arrangements, and multi-panel compositions. Built on the grammar of graphics, it offers intuitive tools to align, stack, and nest plots, simplifying the construction of richly annotated figures for high-dimensional data contexts—such as genomics, transcriptomics, and microbiome studies—by making it easy to link related plots, overlay clustering results, or highlight shared patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/ggalign}
  \item \url{https://yunuuuu.github.io/ggalign/}
  \item Report bugs at \url{https://github.com/Yunuuuu/ggalign/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

Authors:
\itemize{
  \item Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})
}

Other contributors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
