% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_attr}
\alias{ggalign_attr}
\title{Get data from the \code{ggalign} attribute}
\usage{
ggalign_attr(x, field = NULL)
}
\arguments{
\item{x}{Input data for the function used to transform the layout data.}

\item{field}{A string specifying the particular data to retrieve from the
\code{ggalign} attribute. If \code{NULL}, the entire \code{ggalign} attribute will be
returned.  Commonly, this attribute list is attached by \code{\link[=fortify_matrix]{fortify_matrix()}}
or \code{\link[=fortify_data_frame]{fortify_data_frame()}} functions (refer to the \verb{ggalign attributes}
section in the documentation for details). For examples, see
\code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}}.}
}
\value{
The specified data from the \code{ggalign} attribute or \code{NULL} if it is
unavailable.
}
\description{
This function extracts data from the \code{ggalign} attribute retained in the data
when rendering \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object. The \code{ggalign} attribute holds supplementary
information for input data.
}
