% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_attr}
\alias{ggalign_attr}
\title{Get data from the attribute attached across the building process}
\usage{
ggalign_attr(x, field = NULL)
}
\arguments{
\item{x}{Input data for the function used to transform the layout data.}

\item{field}{A string specifying the particular data to retrieve from the
attached attribute. If \code{NULL}, the entire attached attribute list will be
returned.}
}
\value{
The specified data from the attached attribute or \code{NULL} if it is
unavailable.
}
\description{
This function extracts data from the attached attribute when rendering
\code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object. The attached attribute usually holds supplementary
information for the input data.
}
\details{
Commonly, this attribute list is attached by \code{\link[=fortify_matrix]{fortify_matrix()}}
or \code{\link[=fortify_data_frame]{fortify_data_frame()}} functions (refer to the \verb{ggalign attributes}
section in the documentation for details). For examples, see
\code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}}.
}
