% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_attr}
\alias{ggalign_attr}
\title{Get Data from the Attached Attribute in the Rendering Process}
\usage{
ggalign_attr(x, field = NULL)
}
\arguments{
\item{x}{Data used, typically inherited from the layout \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object.}

\item{field}{A string specifying the particular data to retrieve from the
attached attribute. If \code{NULL}, the entire attached attribute list will be
returned.}
}
\value{
The specified data from the attached attribute or \code{NULL} if it is
unavailable.
}
\description{
\code{ggalign_attr} provides access to supplementary information stored as
attributes during the layout rendering process in \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object. These
attributes, commonly attached during data transformation by functions like
\code{\link[=fortify_matrix]{fortify_matrix()}} or \code{\link[=fortify_data_frame]{fortify_data_frame()}}, can include essential
details such as filtered or supplementary data that inform downstream
operations.
}
\details{
Attributes attached to the data are particularly useful when input data is
transformed in ways that restrict access to the complete dataset. For
instance, \code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}} might filter mutation data while adding
attributes that retain essential context, such as the total number of
observations, for detailed or aggregated analyses.
}
