% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_attr_set}
\alias{ggalign_attr_set}
\alias{ggalign_attr_get}
\title{Set or get the Attached Attribute across the Rendering Process}
\usage{
ggalign_attr_set(x, values)

ggalign_attr_get(x)
}
\arguments{
\item{x}{Input data for the layout.}

\item{values}{A list to be attached.}
}
\description{
\itemize{
\item \code{ggalign_attr_set}: Attaches supplementary data to the input, facilitating
downstream use.
\item \code{ggalign_attr_get}: Extracts previously attached supplementary data during
the transformation process.
}
}
\details{
Attributes attached to the data are particularly useful when input data is
transformed in ways that restrict access to the complete dataset. For
instance, \code{\link[=fortify_matrix.MAF]{fortify_matrix.MAF()}} might filter mutation data while adding
attributes that retain essential context, such as the total number of
observations, for detailed or aggregated analyses.
}
\seealso{
\code{\link[=ggalign_attr]{ggalign_attr()}}
}
