% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalign_build}
\alias{ggalign_build}
\title{Build a Graphical Object for Rendering}
\usage{
ggalign_build(x)
}
\arguments{
\item{x}{The object to be processed and built into a renderable format.}
}
\value{
A processed object that can be used for rendering or further
manipulation.
}
\description{
This generic function takes an object (typically a plot) and performs all
necessary steps to transform it into a graphical object that can be rendered
in a grid-based layout.
}
