% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-methods.R
\name{ggalign_build}
\alias{ggalign_build}
\alias{ggalignGrob}
\title{Build ggplot for rendering.}
\usage{
ggalign_build(x)

ggalignGrob(x)
}
\arguments{
\item{x}{A \link[=layout_heatmap]{HeatmapLayout} or \link[=layout_stack]{StackLayout}
object.}
}
\description{
\itemize{
\item \code{ggalign_build}: Build layout into an \code{alignpatches} object
\item \code{ggalignGrob}: Generate a layout plot grob.
}
}
