% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{ggalign_stat}
\alias{ggalign_stat}
\alias{ggalign_stat.HeatmapLayout}
\alias{ggalign_stat.StackLayout}
\title{Get the statistics from the layout}
\usage{
ggalign_stat(x, ...)

\method{ggalign_stat}{HeatmapLayout}(x, ..., position, what)

\method{ggalign_stat}{StackLayout}(x, ..., what)
}
\arguments{
\item{x}{A \code{\link[=layout_heatmap]{layout_heatmap()}} or \code{\link[=layout_stack]{layout_stack()}} object.}

\item{...}{Not used currently.}

\item{position}{A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.}

\item{what}{A single number or string of the plot elements in the stack
layout.}
}
\value{
The statistics
}
\description{
Get the statistics from the layout
}
