% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R, R/alignpatch-patchwork.R,
%   R/alignpatch-plot.R
\name{ggalignGrob}
\alias{ggalignGrob}
\alias{ggalignGrob.default}
\alias{ggalignGrob.patchwork}
\alias{ggalignGrob.patch}
\alias{ggalignGrob.alignpatches}
\title{Generate a plot grob.}
\usage{
ggalignGrob(x)

\method{ggalignGrob}{default}(x)

\method{ggalignGrob}{patchwork}(x)

\method{ggalignGrob}{patch}(x)

\method{ggalignGrob}{alignpatches}(x)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}
}
\value{
A \link[grid:grid.grob]{grob()} object.
}
\description{
Generate a plot grob.
}
