% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ggalignGrob}
\alias{ggalignGrob}
\title{Generate a plot grob.}
\usage{
ggalignGrob(x)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}
}
\value{
A \code{\link[grid:grid.grob]{grob()}} object.
}
\description{
Generate a plot grob.
}
\examples{
ggalignGrob(ggplot())
}
\seealso{
\itemize{
\item \code{\link[=ggalign_build]{ggalign_build()}} for the process of building the object.
\item \code{\link[=ggalign_gtable]{ggalign_gtable()}} for converting the object into a table format.
}
}
