% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno-.R
\name{gganno}
\alias{gganno}
\alias{gganno_top}
\alias{gganno_bottom}
\alias{gganno_left}
\alias{gganno_right}
\title{Heatmap Annotation}
\usage{
gganno(
  mapping = aes(),
  data = NULL,
  position = NULL,
  size = unit(10, "mm"),
  active = TRUE,
  name = NULL,
  order = NULL
)

gganno_top(...)

gganno_bottom(...)

gganno_left(...)

gganno_right(...)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{position}{Possible values are \code{"top"}, \code{"left"}, \code{"bottom"}, and
\code{"right"}. If \code{NULL}, the active context of the ggheatmap will be used.}

\item{size}{Annotation size, must be a \link[grid:unit]{unit} object.}

\item{active}{A logical value of length \code{1} or \code{2}.}

\item{name}{A string of the annotation name.}

\item{order}{Annotation order, must be an single integer.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Heatmap Annotation
}
