% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno-.R
\name{gganno}
\alias{gganno}
\alias{gganno_top}
\alias{gganno_bottom}
\alias{gganno_left}
\alias{gganno_right}
\title{Heatmap annotation of \code{gganno}}
\usage{
gganno(
  mapping = aes(),
  data = NULL,
  position = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = TRUE,
  order = NULL,
  name = NULL
)

gganno_top(...)

gganno_bottom(...)

gganno_left(...)

gganno_right(...)
}
\arguments{
\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}. If \code{NULL}, the active context
of the \code{ggheatmap} will be used.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{labels}{Labels for axis parallelly with heatmap, the default will use
the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the heatmap
axis)
\item An expression vector (must be the same length as heatmap axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. If \code{waiver()},
will inherit from the heatmap. If \code{NULL}, no breaks, in this way labels will
be removed too.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{...}{Additional arguments passed to \code{gganno}.}
}
\value{
A \code{gganno} object.
}
\description{
Heatmap annotation of \code{gganno}
}
