% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganno.R
\name{gganno}
\alias{gganno}
\alias{gganno_top}
\alias{gganno_bottom}
\alias{gganno_left}
\alias{gganno_right}
\title{Heatmap Annotation}
\usage{
gganno(
  position,
  mapping = aes(),
  data = NULL,
  size = unit(10, "mm"),
  active = TRUE,
  name = NULL,
  order = NULL
)

gganno_top(...)

gganno_bottom(...)

gganno_left(...)

gganno_right(...)
}
\arguments{
\item{position}{Possible values are \code{"top"}, \code{"left"}, \code{"bottom"}, and
\code{"right"}.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Heatmap Annotation
}
