% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat-package.R
\docType{package}
\name{ggheat-package}
\alias{ggheat-package}
\title{ggheat: Create Complex Heatmap Using the Grammar of Graphics}
\description{
Creating Heatmap using ggplot2 syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yunuuuu/ggheat}
  \item Report bugs at \url{https://github.com/Yunuuuu/ggheat/issues}
}

}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunp96@163.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
