% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat-package.R
\docType{package}
\name{ggheat-package}
\alias{ggheat-package}
\title{ggheat: Create Complex Heatmap Using the Grammar of Graphics}
\description{
Creating Heatmap using ggplot2 syntax.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
