% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggheat.R
\name{ggheat}
\alias{ggheat}
\title{Heatmap with ggplot2}
\usage{
ggheat(data, mapping = aes(), ...)
}
\arguments{
\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{...}{Additional arguments.}
}
\description{
Heatmap with ggplot2
}
