% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R.R
\name{ggheatmap-add}
\alias{ggheatmap-add}
\alias{+.ggheatmap}
\title{Add components to \code{ggheatmap}}
\usage{
\S4method{+}{ggheatmap,ANY}(e1, e2)
}
\arguments{
\item{e1}{A \link[=ggheat]{ggheatmap} object.}

\item{e2}{An object to be added to the plot, including
\link[ggplot2:gg-add]{gg} elements, \link[=gganno]{ggannotation} object, or \link{htanno}
object.}
}
\value{
A modified \code{ggheatmap} object.
}
\description{
Add components to \code{ggheatmap}
}
