% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{hmanno}
\alias{hmanno}
\title{Determine the active context of heatmap layout}
\usage{
hmanno(
  position = NULL,
  size = NULL,
  action = NULL,
  width = NULL,
  height = NULL,
  free_guides = waiver(),
  what = waiver(),
  ...,
  guides = deprecated(),
  free_spaces = deprecated(),
  plot_data = deprecated(),
  theme = deprecated(),
  free_labs = deprecated()
)
}
\arguments{
\item{position}{Which heatmap annotation should get activated? Possible
values are follows:
\itemize{
\item A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.
\item \code{NULL}: means set the active context into the \code{heatmap} itself.
}}

\item{size}{An \link[grid:unit]{unit} object to set the total size of the heatmap
annotation. Only used if \code{position} is a string.
\itemize{
\item If \code{position} is \code{"top"} or \code{"bottom"}, \code{size} set the total height of the
annotation.
\item If \code{position} is \code{"left"} or \code{"right"}, \code{size} set the total width of the
annotation.
}}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used to define the default plot action
in the layout.}

\item{width, height}{Heatmap body width/height, can be a \link[grid:unit]{unit} object. Only used when \code{position} is
\code{NULL}.}

\item{free_guides}{Override the \code{guides} collection behavior specified in the
heatmap layout for the annotation stack layout. Only used when \code{position} is
a string.}

\item{what}{What should get activated for the anntoation stack? Only used
when \code{position} is a string. Options include::
\itemize{
\item A single number or string of the plot elements in the stack layout.
\item \code{NULL}: remove any active context.
}}

\item{...}{Additional arguments passed to
\link[ggplot2:geom_tile]{geom_tile()}/\link[ggplot2:geom_tile]{geom_raster()}. Only
used when \code{filling = TRUE}.}

\item{guides}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_spaces}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{plot_data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}

\item{theme}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action} argument
instead.}

\item{free_labs}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Please use \code{action}
argument instead.}
}
\value{
A \code{heatmap_active} object which can be added into \link{heatmap_layout}.
}
\description{
Determine the active context of heatmap layout
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    align_dendro()
}
