% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{hmanno}
\alias{hmanno}
\title{Determine the active context of heatmap layout}
\usage{
hmanno(
  position = NULL,
  size = NULL,
  what = waiver(),
  width = NULL,
  height = NULL,
  guides = NULL,
  align_axis_title = NULL,
  plot_data = NA
)
}
\arguments{
\item{position}{Which heatmap annotation should get activated? Possible
values are follows:
\itemize{
\item A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.
\item \code{NULL}: means set the active context into the \code{heatmap} itself.
}}

\item{size}{A \link[grid:unit]{unit} object to set the total size of the heatmap
annotation. This will only be used if \code{what} is a string of \code{"top"},
\code{"left"}, \code{"bottom"}, or \code{"right"}.
\itemize{
\item If position is \code{"top"} or \code{"bottom"}, \code{size} set the total height of the
annotation, otherwise.
\item If position is \code{"left"} or \code{"right"}, \code{size} set the total width of the
annotation.
}}

\item{what}{What should get activated for the stack layout? Possible values
are follows:
\itemize{
\item A single number or string of the plot elements in the stack layout.
\item \code{NULL}: means set the active context into the last \code{Align} object. In
this way, we can add other \code{Align} objects into the \code{StackLayout}.
}}

\item{width, height}{Heatmap body width/height, can be a \link[grid:unit]{unit}
object.}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{"collect"} will collect guides below to the given nesting level, removing
duplicates. \verb{"keep} will stop collection at this level and let guides be
placed alongside their plot. \code{"auto"} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not. If you modify
default guide \code{"position"} with \code{theme(legend.position=...)} while also
collecting guides you must apply that change to the overall layout.}

\item{align_axis_title}{A boolean value or a character of the axis position
(\code{"t"}, \code{"l"}, \code{"b"}, \code{"r"}) indicates how to align the axis title. By
default, all axis title won't be aligned.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is to not modify the data. Use \code{NULL}, if you don't want to modify
anything. Used to modify the data after layout preparation has been applied,
but before the data is handled of to the ggplot2 for rendering. Use this hook
if the you needs change the default data for all \code{geoms}.}
}
\value{
A \code{active} object which can be added into
\link[=layout_heatmap]{HeatmapLayout}.
}
\description{
Determine the active context of heatmap layout
}
