% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{hmanno}
\alias{hmanno}
\title{Determine the active context of heatmap layout}
\usage{
hmanno(
  position = NULL,
  size = NULL,
  guides = NA,
  free_spaces = NA,
  plot_data = NA,
  theme = NA,
  free_labs = NA,
  what = waiver(),
  width = NULL,
  height = NULL
)
}
\arguments{
\item{position}{Which heatmap annotation should get activated? Possible
values are follows:
\itemize{
\item A string of \code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.
\item \code{NULL}: means set the active context into the \code{heatmap} itself.
}}

\item{size}{A \link[grid:unit]{unit} object to set the total size of the heatmap
annotation. This will only be used if \code{position} is a string of
\code{"top"}, \code{"left"}, \code{"bottom"}, or \code{"right"}.
\itemize{
\item If position is \code{"top"} or \code{"bottom"}, \code{size} set the total height of the
annotation.
\item If position is \code{"left"} or \code{"right"}, \code{size} set the total width of the
annotation.
}}

\item{guides}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which guide should be collected. If \code{NULL}, no
guides will be collected. Default: \code{"tlbr"}.}

\item{free_spaces}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which border spaces should be removed. If
\code{NULL} (default), no space will be removed.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is \code{NULL}, which means we won't modify anything.

Used to modify the data after layout has been created, but before the data is
handled of to the ggplot2 for rendering. Use this hook if the you needs
change the default data for all \code{geoms}.}

\item{theme}{One of:
\itemize{
\item \link[ggplot2:waiver]{waiver()}: will inherit from the parent layout.
\item \code{NULL}: won't add any theme components.
\item \link[ggplot2:theme]{theme()}: will be added to all plots in the layout.
}}

\item{free_labs}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which axis title should be free from
alignment. If \code{NULL}, all axis title will be aligned. Default: \code{"tlbr"}.}

\item{what}{What should get activated for the anntoation stack? Only used
when position is not \code{NULL}. See \link{stack_active} for details.}

\item{width, height}{Heatmap body width/height, can be a \link[grid:unit]{unit}
object.}
}
\value{
A \code{heatmap_active} object which can be added into \link{heatmap_layout}.
}
\description{
Determine the active context of heatmap layout
}
\examples{
ggheatmap(matrix(rnorm(81), nrow = 9)) +
    hmanno("top") +
    align_dendro()
}
