% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-.R
\name{htanno}
\alias{htanno}
\title{Build \code{htanno} object}
\usage{
htanno(
  htanno,
  data = NULL,
  position = NULL,
  params = list(),
  size = NULL,
  active = NULL,
  name = NULL,
  order = NULL,
  check.param = TRUE
)
}
\arguments{
\item{htanno}{A \code{HtannoProto} object}

\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{ "bottom"}, and \code{"right"}.}

\item{params}{A list of parameters passed  to \code{htanno}.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{active}{A logical value of length \code{1} or \code{2}.}

\item{name}{A string of the annotation name.}

\item{order}{Annotation order, must be an single integer.}

\item{check.param}{A boolean value indicates whether to check the supplied
parameters and warn.}
}
\value{
A \code{htanno} object.
}
\description{
\code{htanno} is a special annotation, which can act with the main heatmap.
Especially control the order of the main heatmap or split the heatmap into
different slices.
}
