% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno.R
\name{htanno}
\alias{htanno}
\title{Determine the context of subsequent manipulations}
\usage{
htanno(position = NULL, what = NULL)
}
\arguments{
\item{position}{Possible values are "heatmap", "top", "left", "bottom", and
"right".}

\item{what}{A string of the name the annotation.}
}
\description{
Determine the context of subsequent manipulations
}
