% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-.R
\docType{data}
\name{htanno}
\alias{htanno}
\alias{HtannoProto}
\title{Heatmap annotation of \code{htanno}}
\usage{
htanno(
  data = NULL,
  params = list(),
  position = NULL,
  size = NULL,
  labels = NULL,
  labels_nudge = NULL,
  set_context = NULL,
  order = NULL,
  name = NULL,
  check.param = TRUE,
  htanno_class = HtannoProto
)
}
\arguments{
\item{data}{A matrix, if it is a simple vector, it will be converted to a
one-column matrix. Data.frame will also be coerced into matrix.}

\item{params}{A list of parameters passed  to \code{htanno}.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}.}

\item{size}{Annotation size, can be a \link[grid:unit]{unit} object.}

\item{labels}{Labels for axis parallelly with heatmap, the default will use
the rownames of the \code{data}. One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels
\item A character vector giving labels (must be same length as the heatmap
axis)
\item An expression vector (must be the same length as heatmap axis). See
\code{?plotmath} for details.
\item A function that takes the default labels as the input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{labels_nudge}{A single numeric or a numeric value of length
\code{nrow(data)}, to nudge each text label away from the center. If \code{waiver()},
will inherit from the heatmap. If \code{NULL}, no breaks, in this way labels will
be removed too.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{order}{Annotation order, must be an single integer.}

\item{name}{A string of the annotation name.}

\item{check.param}{A single boolean value indicates whether to check the
supplied parameters and warn.}

\item{htanno_class}{A \code{HtannoProto} object}
}
\value{
A \code{htanno} object.
}
\description{
\code{htanno} is a special annotation, which can act with the main heatmap.
Especially control the order of the main heatmap or split the heatmap into
different slices.
}
\section{HtannoProto}{

Each of the \verb{Htanno*} objects is just a \link[ggplot2:ggproto]{ggproto()} object,
descended from the top-level \code{HtannoProto}, and each implements various
methods and fields.

To create a new type of \verb{Htanno*} object, you typically will want to
override one or more of the following:
\itemize{
\item \code{setup_params}: Prepare parameter or check parameters used by this
annotation.
\item \code{add}: A method used to add other components into this annotation. Usually
\code{gg} components.
\item \code{compute}: A method used to compute statistics.
\item \code{layout}: A method used to group heamap rows/columns into panels or
reorder heamtap rows/columns.
\item \code{draw}: A method used to draw the plot.
}
}

\keyword{datasets}
