% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-group.R
\name{htanno_group}
\alias{htanno_group}
\title{Group Heatmap rows/columns}
\usage{
htanno_group(
  group,
  position = NULL,
  set_context = NULL,
  name = NULL,
  check.param = TRUE
)
}
\arguments{
\item{group}{A character define the heatmap groups, this will split the
heatmap into different panels.}

\item{position}{A string of the annotation position, Possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{name}{A string of the annotation name.}

\item{check.param}{A single boolean value indicates whether to check the
supplied parameters and warn.}
}
\value{
A \code{htanno} object.
}
\description{
Group Heatmap rows/columns
}
