% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htanno-kmeans.R
\name{htanno_kmeans}
\alias{htanno_kmeans}
\title{Group Heatmap rows/columns by kmeans}
\usage{
htanno_kmeans(
  centers,
  iter.max = 10,
  nstart = 1,
  algorithm = c("Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"),
  trace = FALSE,
  position = NULL,
  set_context = NULL,
  name = NULL,
  check.param = TRUE
)
}
\arguments{
\item{centers}{either the number of clusters, say \eqn{k}, or a set of
    initial (distinct) cluster centres.  If a number, a random set of
    (distinct) rows in \code{x} is chosen as the initial centres.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{if \code{centers} is a number, how many random sets
    should be chosen?}

\item{algorithm}{character: may be abbreviated.  Note that
    \code{"Lloyd"} and \code{"Forgy"} are alternative names for one
    algorithm.}

\item{trace}{logical or integer number, currently only used in the
    default method (\code{"Hartigan-Wong"}): if positive (or true),
    tracing information on the progress of the algorithm is
    produced.  Higher values may produce more tracing information.}

\item{position}{A string of the annotation position, possible values are
\code{"top"}, \code{"left"}, \code{"bottom"}, and \code{"right"}.}

\item{set_context}{A logical value of length \code{2} indicates whether to set the
active context to the \code{position} for the \link[=ggheat]{ggheatmap} and whether to
set the active context to current annotation for the annotation list in
\link[=ggheat]{ggheatmap} when added.}

\item{name}{A string of the annotation name.}

\item{check.param}{A single boolean value indicates whether to check the
supplied parameters and warn.}
}
\value{
A \code{htanno} object.
}
\description{
Group Heatmap rows/columns by kmeans
}
