% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{is.ggheatmap}
\alias{is.ggheatmap}
\title{Reports whether \code{x} is a \code{\link[=heatmap_layout]{heatmap_layout()}} object}
\usage{
is.ggheatmap(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A boolean value
}
\description{
Reports whether \code{x} is a \code{\link[=heatmap_layout]{heatmap_layout()}} object
}
\examples{
is.ggheatmap(ggheatmap(1:10))
}
