% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-operator.R
\name{layout-operator}
\alias{layout-operator}
\alias{+.QuadLayout}
\alias{&.QuadLayout}
\alias{-.QuadLayout}
\alias{+.HeatmapLayout}
\alias{&.HeatmapLayout}
\alias{-.HeatmapLayout}
\alias{+.ggheatmap}
\alias{&.ggheatmap}
\alias{-.ggheatmap}
\alias{+.ggside}
\alias{&.ggside}
\alias{-.ggside}
\alias{+.StackLayout}
\alias{&.StackLayout}
\alias{-.StackLayout}
\alias{+.CrossLayout}
\title{Layout operator}
\arguments{
\item{e1}{A \code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object.}

\item{e2}{An object to be added to the plot.}
}
\value{
A modified \code{Layout} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\itemize{
\item \code{+}: Adds elements to the active plot in the active layout.
\item \code{&}: Applies elements to all plots in the layout.
\item \code{-}: Adds elements to multiple plots in the layout.
}
}
\details{
The \code{+} operator is straightforward and should be used as needed.

In order to reduce code repetition \code{ggalign} provides two operators for
adding ggplot elements (geoms, themes, facets, etc.) to multiple/all plots in
\code{\link[=quad_layout]{quad_layout()}}/\code{\link[=ggheatmap]{ggheatmap()}} or \code{\link[=stack_layout]{stack_layout()}} object: \code{-} and \code{&}. See \code{vignette("operator")} for details.
}
\examples{
set.seed(123)
small_mat <- matrix(rnorm(56), nrow = 7)
ggheatmap(small_mat) +
    anno_top() +
    ggalign() +
    geom_point(aes(y = value))

# `&` operator apply it to all plots
ggheatmap(small_mat) +
    anno_top() +
    align_dendro() &
    theme(panel.border = element_rect(
        colour = "red", fill = NA, linewidth = unit(2, "mm")
    ))

# If the active layout is the annotation stack, the `-` operator will only
# add the elements to all plots in the active annotation stack:
ggheatmap(small_mat) +
    anno_left(size = 0.2) +
    align_dendro(aes(color = branch), k = 3L) +
    align_dendro(aes(color = branch), k = 3L) -
    # Modify the the color scales of all plots in the left annotation
    scale_color_brewer(palette = "Dark2")

# If the active layout is the `stack_layout()` itself, `-`
# applies the elements to all plots in the layout except the nested
# `ggheatmap()`/`quad_layout()`.
stack_alignv(small_mat) +
    align_dendro() +
    ggtitle("I'm from the parent stack") +
    ggheatmap() +
    # remove any active context
    stack_active() +
    align_dendro() +
    ggtitle("I'm from the parent stack") -
    # Modify the the color scales of all plots in the stack layout except the
    # heatmap layout
    scale_color_brewer(palette = "Dark2") -
    # set the background of all plots in the stack layout except the heatmap
    # layout
    theme(plot.background = element_rect(fill = "red"))

}
