% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-align_plots.R
\name{layout_annotation}
\alias{layout_annotation}
\title{Add layout annotation}
\usage{
layout_annotation(..., theme = waiver())
}
\arguments{
\item{...}{Currently unused. May accept a theme in the future.}

\item{theme}{A theme object. If not \code{waiver()}, an error will be raised.}
}
\value{
None. This function is used for input validation.
}
\description{
This function is a placeholder for future extensions.
If you're trying to apply a theme, use \code{\link[=layout_theme]{layout_theme()}} instead.
}
\keyword{internal}
