% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{layout_heatmap}
\alias{layout_heatmap}
\alias{ggheatmap}
\title{Arrange plots around a Heatmap}
\usage{
layout_heatmap(
  data,
  mapping = aes(),
  width = NULL,
  height = NULL,
  guides = waiver(),
  free_labs = waiver(),
  free_sizes = waiver(),
  plot_data = waiver(),
  filling = TRUE,
  ...,
  set_context = TRUE,
  order = NULL,
  name = NULL
)

ggheatmap(
  data,
  mapping = aes(),
  width = NULL,
  height = NULL,
  guides = waiver(),
  free_labs = waiver(),
  free_sizes = waiver(),
  plot_data = waiver(),
  filling = TRUE,
  ...,
  set_context = TRUE,
  order = NULL,
  name = NULL
)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, and any other data
which can be converted into a matrix. Simple vector will be converted into a
one column matrix.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If \code{NULL},
will using \code{aes(.data$.x, .data$.y)}.}

\item{width, height}{Heatmap body width/height, can be a \link[grid:unit]{unit}
object.}

\item{guides}{A single boolean value indicates whether to collect the guides,
or you can specify the string of the guide position to collect. Allowed
strings are: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}

\item{free_labs}{A boolean value or a character of the axis position (\code{"t"},
\code{"l"}, \code{"b"}, \code{"r"}) indicates which axis title should be free from
alignment. By default, all axis title won't be aligned.}

\item{free_sizes}{A character specifies the ggplot elements which won't
count space sizes when alignment.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is to not modify the data. Use \code{NULL}, if you don't want to modify
anything.

Used to modify the data after layout has been created, but before the data is
handled of to the ggplot2 for rendering. Use this hook if the you needs
change the default data for all \code{geoms}.}

\item{filling}{A boolean value indicates whether to fill the heatmap. If you
want to custom the filling style, you can set to \code{FALSE}.}

\item{...}{Additional arguments passed to \link[ggplot2:geom_tile]{geom_tile}.
Only used when \code{filling = TRUE}.}

\item{set_context}{A single boolean value indicates whether to set the active
context to current \code{Align} object. If \code{TRUE}, all subsequent ggplot elements
will be added into this \code{Align} object.}

\item{order}{An single integer for the layout order.}

\item{name}{A string of the object name.}
}
\value{
A \code{HeatmapLayout} object.
}
\description{
\code{ggheatmap} is an alias of \code{layout_heatmap}.
}
\section{ggplot2 specification}{

The data input in \code{ggheatmap} will be converted into the long formated data
frame when drawing. The default mapping will use \code{aes(.data$.x, .data$.y)},
you can use \code{mapping} argument to control it. The data in the underlying
\code{ggplot} object contains following columns:
\itemize{
\item \code{.xpanel} and \code{.ypanel}: the column and row panel
\item \code{.x} and \code{.y}: the \code{x} and \code{y} coordinates
\item \code{.row_names} and \code{.column_names}: the row and column names of the original
matrix (only applicable when names exist).
\item \code{.row_index} and \code{.column_index}: the row and column index of the original
matrix.
\item \code{value}: the actual matrix value.
}
}

\examples{
ggheatmap(1:10)
ggheatmap(letters)
}
