% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-heatmap-.R
\name{layout_heatmap}
\alias{layout_heatmap}
\alias{ggheatmap}
\alias{layout_heatmap.matrix}
\alias{layout_heatmap.data.frame}
\alias{layout_heatmap.numeric}
\alias{layout_heatmap.character}
\alias{layout_heatmap.default}
\alias{show,LayoutHeatmap-method}
\title{Heatmap layout}
\usage{
layout_heatmap(data = NULL, mapping = aes(), ...)

ggheatmap(data = NULL, mapping = aes(), ...)

\method{layout_heatmap}{matrix}(
  data,
  mapping = aes(),
  width = NULL,
  height = NULL,
  guides = "collect",
  filling = TRUE,
  ...
)

\method{layout_heatmap}{data.frame}(data, mapping = aes(), ...)

\method{layout_heatmap}{numeric}(data, mapping = aes(), ...)

\method{layout_heatmap}{character}(data, mapping = aes(), ...)

\method{layout_heatmap}{default}(data, mapping = aes(), ...)

\S4method{show}{LayoutHeatmap}(object)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, and any other data
which can be converted into a matrix. Simple vector will be converted into a
one column matrix.}

\item{mapping}{Default list of aesthetic mappings to use for plot. If \code{NULL},
will using \code{aes(.data$.x, .data$.y)}.}

\item{...}{Additional arguments passed to matrix method.}

\item{width, height}{Heatmap width/height, can be a \link[grid:unit]{unit} object.}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{'collect'} will collect guides below to the given nesting level, removing
duplicates. \code{'keep'} will stop collection at this level and let guides be
placed alongside their plot. \code{auto} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not.  If you modify
default guide "position" with \link[ggplot2:theme]{theme(legend.position=...)}
while also collecting guides you must apply that change to the overall
patchwork (see example).}

\item{filling}{A boolean value indicates whether filling the heatmap. If you
want to custom the filling style, you can set to \code{FALSE}.}

\item{object}{A \code{LayoutHeatmap} object.}
}
\value{
A \code{LayoutHeatmap} object.
}
\description{
\code{ggheatmap} is an alias of \code{layout_heatmap}.
}
\section{ggplot2 details}{

The data input in \code{ggheatmap} will be converted into the long formated data
frame when drawing. The default mapping will use \code{aes(.data$.x, .data$.y)},
you can use \code{mapping} argument to control it. The data in the underlying
\code{ggplot} object contains following columns:
\itemize{
\item \code{.row_panel}: the row panel
\item \code{.column_panel}: the column panel
\item \code{.row_names} and \code{.column_names}: the row and column names of the original
matrix (only applicable when names exist).
\item \code{.row_index} and \code{.column_index}: the row and column index of the original
matrix.
\item \code{.x} and \code{.y}: the \code{x} and \code{y} coordinates
\item \code{value}: the actual matrix value.
}
}

\examples{
ggheatmap(1:10)
ggheatmap(letters)
}
