% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{layout_stack}
\alias{layout_stack}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
layout_stack(data, direction = NULL, ..., environment = parent.frame())

ggstack(data, direction = NULL, ..., environment = parent.frame())
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{...}{Not used currently.}

\item{environment}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Used prior to tidy
evaluation.}
}
\value{
A \code{StackLayout} object.
}
\description{
\code{ggstack} is an alias of \code{layout_stack}.
}
\examples{
ggstack(matrix(rnorm(100L), nrow = 10L)) + align_dendro()
}
