% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{layout_stack}
\alias{layout_stack}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
layout_stack(
  data,
  direction = NULL,
  sizes = NULL,
  guides = waiver(),
  free_labs = waiver(),
  free_sizes = waiver(),
  plot_data = waiver()
)

ggstack(
  data,
  direction = NULL,
  sizes = NULL,
  guides = waiver(),
  free_labs = waiver(),
  free_sizes = waiver(),
  plot_data = waiver()
)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{sizes}{A numeric or \link[grid:unit]{unit} object of length \code{3} indicates
the relative widths (\code{direction = "horizontal"}) / heights (\code{direction = "vertical"}).}

\item{guides}{A single boolean value indicates whether to collect the guides,
or you can specify the string of the guide position to collect. Allowed
strings are: \code{"t"},\code{"l"},\code{"b"},and \code{"r"}.}

\item{free_labs}{A boolean value or a character of the axis position (\code{"t"},
\code{"l"}, \code{"b"}, \code{"r"}) indicates which axis title should be free from
alignment. By default, all axis title won't be aligned.}

\item{free_sizes}{A character specifies the ggplot elements which won't
count space sizes when alignment.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is to not modify the data. Use \code{NULL}, if you don't want to modify
anything.

Used to modify the data after layout has been created, but before the data is
handled of to the ggplot2 for rendering. Use this hook if the you needs
change the default data for all \code{geoms}.}
}
\value{
A \code{StackLayout} object.
}
\description{
\code{ggstack} is an alias of \code{layout_stack}.
}
