% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{layout_stack}
\alias{layout_stack}
\alias{ggstack}
\title{Stack layout}
\usage{
layout_stack(
  data,
  direction = NULL,
  labels = waiver(),
  labels_nudge = waiver(),
  rel_sizes = NULL,
  guides = NULL,
  align_axis_title = NULL
)

ggstack(
  data,
  direction = NULL,
  labels = waiver(),
  labels_nudge = waiver(),
  rel_sizes = NULL,
  guides = NULL,
  align_axis_title = NULL
)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, and any other data
which can be converted into a matrix. Simple vector will be converted into a
one column matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{labels, labels_nudge}{Default \code{labels}/\code{labels_nudge} for axis
parallelly with the layout.}

\item{rel_sizes}{A numeric or \link[grid:unit]{unit} object of length \code{3}
indicates the relative widths (\code{direction = "vertical"}) / heights
(\code{direction = "vertical"}).}

\item{guides}{A string specifying how guides should be treated in the layout.
\code{"collect"} will collect guides below to the given nesting level, removing
duplicates. \verb{"keep} will stop collection at this level and let guides be
placed alongside their plot. \code{"auto"} will allow guides to be collected if a
upper level tries, but place them alongside the plot if not. If you modify
default guide \code{"position"} with \code{theme(legend.position=...)} while also
collecting guides you must apply that change to the overall layout.}

\item{align_axis_title}{A boolean value or a character of the axis position
(\code{"t"}, \code{"l"}, \code{"b"}, \code{"r"}) indicates how to align the axis title. By
default, all axis title won't be aligned.}
}
\value{
A \code{LayoutStack} object.
}
\description{
\code{ggstack} is an alias of \code{layout_stack}.
}
