% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-align_plots.R
\name{layout_theme}
\alias{layout_theme}
\title{Modify components of the layout theme}
\usage{
layout_theme(...)
}
\arguments{
\item{...}{Additional arguments passed to \link[ggplot2:theme]{theme()}.}
}
\value{
A \code{layout_theme} object to be added to \code{\link[=heatmap_layout]{heatmap_layout()}} or \code{\link[=stack_layout]{stack_layout()}} object.
}
\description{
Used to render the \code{guides}, \code{title}, \code{subtitle}, \code{caption},
\code{margins} and \code{background}.
}
\section{Theme inheritance}{


Theme elements inherit properties from other theme elements hierarchically.
For example, \code{axis.title.x.bottom} inherits from \code{axis.title.x} which inherits
from \code{axis.title}, which in turn inherits from \code{text}. All text elements inherit
directly or indirectly from \code{text}; all lines inherit from
\code{line}, and all rectangular objects inherit from \code{rect}.
This means that you can modify the appearance of multiple elements by
setting a single high-level component.

Learn more about setting these aesthetics in \code{vignette("ggplot2-specs")}.

}

\examples{
layout_theme(plot.background = element_rect(fill = "green"))
}
