% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{LayoutProto}
\alias{LayoutProto}
\title{A \code{Layout} object}
\usage{
LayoutProto(
  data = NULL,
  schemes = Schemes(),
  titles = layout_title(),
  theme = NULL,
  name = NA_character_
)
}
\description{
A \code{Layout} object defines how to place the plots.
}
\keyword{internal}
