% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_draw}
\alias{mark_draw}
\title{Define the links to connect the marked observations}
\usage{
mark_draw(.draw, ..., .group1 = NULL, .group2 = NULL)
}
\arguments{
\item{.draw}{A function used to draw the links. The function must return a
\code{\link[grid:grid.grob]{grob()}} object. If the function does not return a valid
\code{grob}, nothing will be drawn. The input data for the function must contain
two arguments: a data frame for the panel side coordinates and a data frame
for the marked observation coordinates.}

\item{...}{A list of formulas, where each side of the formula should be an
\code{integer} or \code{character} index, or a \code{range_link()} object defining the
linked observations, use \code{NULL} to indicate no link in this side. If only
the left-hand side of the formula exists, you can input it directly. For
integer indices, wrap them with \code{\link[=I]{I()}} to indicate their order according
to the layout. You can also use \code{waiver()} to inherit values from the
opposite link argument.}

\item{.group1, .group2}{A single boolean value indicating whether to use the
panel group information from the layout as the links. By default, will
guess from the layout.}
}
\description{
This function allows users to define links between marked observations (e.g.,
for creating visual connections for related data), which could help explain
the observations.
}
\seealso{
\itemize{
\item \code{\link[=mark_line]{mark_line()}}
\item \code{\link[=mark_tetragon]{mark_tetragon()}}
\item \code{\link[=.mark_draw]{.mark_draw()}}
}
}
