% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_line}
\alias{mark_line}
\title{Link the observations and the panel with a line}
\usage{
mark_line(..., element = NULL)
}
\arguments{
\item{...}{Additional arguments passed on to \code{\link[=mark_draw]{mark_draw()}}.}

\item{element}{A \code{\link[ggplot2:element]{element_line()}} object.}
}
\description{
Link the observations and the panel with a line
}
