% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order2.R
\name{order2}
\alias{order2}
\alias{order2.default}
\alias{order2.hclust}
\alias{order2.dendrogram}
\alias{order2.ser_permutation_vector}
\alias{order2.ser_permutation}
\title{Ordering Permutation}
\usage{
order2(x)

\method{order2}{default}(x)

\method{order2}{hclust}(x)

\method{order2}{dendrogram}(x)

\method{order2}{ser_permutation_vector}(x)

\method{order2}{ser_permutation}(x)
}
\arguments{
\item{x}{Any objects can be extracting ordering.}
}
\value{
An integer vector unless any of the inputs has \code{2^31} or more
elements, when it is a double vector.
}
\description{
\code{order2} returns a permutation which rearranges its first argument into
ascending order. By default, \code{order2} will call \code{\link[=order]{order()}} directly.
}
\examples{
order2(sample(10L))
order2(letters)
}
