% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{pair_links}
\alias{pair_links}
\alias{range_link}
\title{Helper function to create a pair of links}
\usage{
pair_links(...)

range_link(point1, point2)
}
\arguments{
\item{...}{A list of formula, each side of the formula should be an integer
or character index, or a \code{range_link()} object to define the linked
observations. You can use \code{NULL} to indicate none. For integer indices, you
can wrap them with \code{\link[=I]{I()}} to indicate the order based on the layout. You
can also use \code{waiver()} to inherit the values from the opposite link
argument.}

\item{point1, point2}{A single integer or character index, defining the lower
and higher bounds of the range. For integer indices, wrap them with \code{\link[=I]{I()}}
to indicate the ordered index by the layout.}
}
\description{
\itemize{
\item \code{pair_link}: Helper function to create a pair of links.
\item \code{range_link}: Helper function to create a range of observations.
}
}
