% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair-links.R
\name{pair_links}
\alias{pair_links}
\alias{range_link}
\title{Helper function to create a pair of links}
\usage{
pair_links(...)

range_link(point1, point2)
}
\arguments{
\item{...}{<\link[rlang:dyn-dots]{dyn-dots}> A list of formulas, where each side
of the formula should be an \code{integer} or \code{character} index, or a
\code{range_link()} object defining the linked observations. Use \code{NULL} to
indicate no link on that side. You can also combine these by wrapping them
into a single \code{list()}. If only the left-hand side of the formula exists,
you can input it directly. For integer indices, wrap them with \code{\link[=I]{I()}} to
preserve their order according to the layout. You can also use \code{\link[=waiver]{waiver()}}
to inherit values from the opposite link argument.}

\item{point1, point2}{A single integer or character index, defining the lower
and higher bounds of the range. For integer indices, wrap them with \code{\link[=I]{I()}}
to indicate the ordered index by the layout.}
}
\description{
\itemize{
\item \code{pair_links}: Helper function to create pair of links.
\item \code{range_link}: Helper function to create a range of observations.
}
}
\examples{
x <- pair_links(
    # links in the left hand only
    1:2,
    c("a", "b"),
    range_link(1, 6),
    range_link("a", "b"),
    # links in the right hand only
    ~ 1:2,
    ~ c("a", "b"),
    ~ range_link(1, 6),
    # links in the both side
    range_link(1, 6) ~ c("a", "b"),
    # waiver() indicates the right hand is the same of the left hand
    range_link(1, 6) ~ waiver(),
    # the same for the left hand
    waiver() ~ 1:2,
    ~NULL # an empty link
)
x

# we can modify it as usual list
x[[1]] <- NULL # remove the first link
x$a <- ~LETTERS
x

# modify with a list of links
x[1:2] <- list(~ c("a", "b"), ~ range_link("a", "b"))
x
}
