% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{patch.alignpatches}
\alias{patch.alignpatches}
\title{Convert Object into a Grob}
\usage{
\method{patch}{alignpatches}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Not used currently.}
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} to convert objects into a
\link[grid:grid.grob]{grob}, enabling their alignment within \code{\link[=align_plots]{align_plots()}}.
}
\seealso{
\link[=align_plots]{alignpatches}
}
