% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch.formula}
\alias{patch.formula}
\alias{patch.function}
\title{Convert Object into a Grob}
\usage{
\method{patch}{formula}(x, ..., device = NULL, name = NULL)

\method{patch}{`function`}(x, ..., device = NULL, name = NULL)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Graphical Parameters passed on to \link[graphics:par]{par()}.}

\item{device}{
    A function that opens a graphics device for \code{grid.echo()} to
    work on.  By default this is an off-screen, in-memory device
    based on the \code{pdf} device.  This default device may not be
    satisfactory when using custom fonts.
  }

\item{name}{
A character identifier.
  }
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
The \code{patch()} function is used by \code{\link[=ggwrap]{ggwrap()}} and \code{\link[=inset]{inset()}} to convert
objects into a \code{\link[grid:grid.grob]{grob}}.
}
\seealso{
\link{plot}
}
