% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-patch.R
\name{patch.ggplot}
\alias{patch.ggplot}
\title{Convert Object into a Grob}
\usage{
\method{patch}{ggplot}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Not used currently.}
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
The \code{patch()} function is used by \code{\link[=ggwrap]{ggwrap()}} and \code{\link[=inset]{inset()}} to convert
objects into a \code{\link[grid:grid.grob]{grob}}.
}
\seealso{
\link[ggplot2:ggplot]{ggplot}

Other \code{\link[=patch]{patch()}} methods:
\code{\link{patch.Heatmap}()},
\code{\link{patch.formula}()},
\code{\link{patch.ggalign::AlignPatches}()},
\code{\link{patch.grob}()},
\code{\link{patch.patch}()},
\code{\link{patch.patch_ggplot}()},
\code{\link{patch.patchwork}()},
\code{\link{patch.pheatmap}()},
\code{\link{patch.recordedplot}()},
\code{\link{patch.trellis}()}
}
\concept{patch}
