% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch.Heatmap}
\alias{patch.Heatmap}
\alias{patch.HeatmapList}
\alias{patch.HeatmapAnnotation}
\title{Convert Object into a Grob}
\usage{
\method{patch}{Heatmap}(x, ..., device = NULL)

\method{patch}{HeatmapList}(x, ..., device = NULL)

\method{patch}{HeatmapAnnotation}(x, ..., device = NULL)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Additional arguments passed to \link[ComplexHeatmap:draw-dispatch]{draw()}.}

\item{device}{ A function that opens a graphics device for temporary
    rendering. By default this is an off-screen, in-memory device
    based on the \code{pdf} device, but this default device may not be
    satisfactory when using custom fonts.
  }
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} and \code{\link[=inset]{inset()}} to convert objects
into a \link[grid:grid.grob]{grob}.
}
\seealso{
\itemize{
\item \link[ComplexHeatmap:Heatmap]{Heatmap}
\item \link[ComplexHeatmap:HeatmapAnnotation]{HeatmapAnnotation}
}
}
