% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-inset.R
\name{patch}
\alias{patch}
\title{Convert Object into a Grob}
\usage{
patch(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \code{\link[grid:grid.grob]{grob}}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A \code{\link[grid:grid.grob]{grob}} object.
}
\description{
The \code{patch()} function is used by \code{\link[=ggwrap]{ggwrap()}} and \code{\link[=inset]{inset()}} to convert
objects into a \code{\link[grid:grid.grob]{grob}}.
}
\seealso{
\itemize{
\item \code{\link{patch.grob}} / \code{\link{patch.gList}}
\item \code{\link{patch.ggplot}}
\item \code{\link{patch.patch_ggplot}}
\item \code{\link{patch.patchwork}}
\item \code{\link{patch.patch}}
\item \code{\link{patch.trellis}}
\item \code{\link{patch.formula}} / \code{\link{patch.function}}
\item \code{\link{patch.recordedplot}}
\item \code{\link{patch.Heatmap}}
\item \code{\link{patch.HeatmapList}}
\item \code{\link{patch.HeatmapAnnotation}}
\item \code{\link{patch.pheatmap}}
}
}
\keyword{internal}
