% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{patch}
\alias{patch}
\title{Convert Object into a Grob}
\usage{
patch(x, ...)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} to convert objects into a
\link[grid:grid.grob]{grob}, enabling their alignment within \code{\link[=align_plots]{align_plots()}}.
}
\seealso{
\itemize{
\item \link{patch.grob}
\item \link{patch.ggplot}
\item \link{patch.patch_ggplot}
\item \link{patch.patchwork}
\item \link{patch.patch}
\item \link{patch.formula} / \link{patch.function}
\item \link{patch.recordedplot}
\item \link{patch.Heatmap}
\item \link{patch.HeatmapList}
\item \link{patch.HeatmapAnnotation}
\item \link{patch.pheatmap}
}
}
\keyword{internal}
