% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-wrap.R
\name{patch}
\alias{patch}
\alias{patch.grob}
\alias{patch.formula}
\alias{patch.Heatmap}
\alias{patch.HeatmapList}
\alias{patch.HeatmapAnnotation}
\title{Convert Object into a Grob for Wrapping}
\usage{
patch(x, ...)

\method{patch}{grob}(x, ...)

\method{patch}{formula}(x, ..., device = NULL)

\method{patch}{Heatmap}(x, ..., device = NULL)

\method{patch}{HeatmapList}(x, ..., device = NULL)

\method{patch}{HeatmapAnnotation}(x, ..., device = NULL)
}
\arguments{
\item{x}{An object to be converted into a \link[grid:grid.grob]{grob}.}

\item{...}{Additional arguments passed to specific methods.}

\item{device}{
    A function that opens a graphics device for \code{grid.echo()} to
    work on.  By default this is an off-screen, in-memory device
    based on the \code{pdf} device.  This default device may not be
    satisfactory when using custom fonts.
  }
}
\value{
A \link[grid:grid.grob]{grob} object.
}
\description{
The \code{patch()} function is used by \code{\link[=wrap]{wrap()}} to convert objects into a
\link[grid:grid.grob]{grob}, enabling their alignment within \code{\link[=align_plots]{align_plots()}}.
}
\keyword{internal}
