% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-alignpatches.R
\name{patch_options}
\alias{patch_options}
\title{Options passed to the Patch \code{gtable} method}
\usage{
patch_options(theme = NULL, guides = NULL, tag = NULL, metadata = list())
}
\description{
This class defines the options that can be passed to the \code{gtable} method of a
\code{Patch} object.  It includes:
}
\details{
\itemize{
\item \code{theme}: The theme to be applied, which can be either \code{NULL} or a ggplot2
\link[ggplot2:theme]{theme} object.
\item \code{guides}: The guides for the plot, which can be \code{NULL} or a character
vector.
\item \code{tag}: Either \code{NULL} (no tag), a single string, or a \code{LayoutTagger} object
that provides a \verb{$tag()} method to generate a tag string for tagging each
plot individually.
\item \code{metadata}: A list of additional metadata. This can store any extra
information relevant to the patch, such as custom parameters or auxiliary
data that do not fit into the other categories.
}
}
\keyword{internal}
