% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-.R
\name{patch_options}
\alias{patch_options}
\title{Options passed to the Patch \code{gtable} method}
\usage{
patch_options(theme = NULL, guides = NULL, tag = NULL)
}
\description{
This class defines the options that can be passed to the \code{gtable} method of a
\code{Patch} object.  It includes:
}
\details{
\itemize{
\item \code{theme}: The theme to be applied, which can be either \code{NULL} or a ggplot2
\link[ggplot2:theme]{theme} object.
\item \code{guides}: The guides for the plot, which can be \code{NULL} or a character
vector.
\item \code{tag}: Can be \code{NULL} (no tag), a single string, or a \code{LayoutTagger} object
that provides a \verb{$tag()} method to generate a tag string for each plot.
The \code{LayoutTagger} is used specifically by \code{\link[=alignpatches]{alignpatches()}}. For
individual plots, you typically call the \verb{$tag()} method of the
\code{LayoutTagger} object to return a string, which triggers the internal
\code{Patch$tag()} method to add a tag.
}
}
\keyword{internal}
