% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignpatch-title.R
\name{patch_titles}
\alias{patch_titles}
\title{Add patch titles around the plot}
\usage{
patch_titles(
  top = waiver(),
  left = waiver(),
  bottom = waiver(),
  right = waiver()
)
}
\arguments{
\item{top, left, bottom, right}{A string of title to be added into top, left,
bottom, or right.}
}
\description{
You can also use \link[ggplot2:labs]{labs()} to add patch titles in top, left,
bottom, and right.
}
