% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme-.R
\name{plot_add_scheme}
\alias{plot_add_scheme}
\title{Apply a Scheme to a plot}
\usage{
plot_add_scheme(x, object, ...)
}
\arguments{
\item{x}{The object whose properties are being updated.}

\item{object}{The object from which the properties will be applied to \code{x}.}
}
\description{
\code{plot_add_scheme()} is a generic used to apply a \code{\link{Scheme}} (or a \code{\link{Schemes}}
container) to a plot object. This allows schemes to update or modify
the plot's appearance or behavior according to their configuration.
}
\details{
By default When a \code{\link{Schemes}} object is passed, each individual \code{\link{Scheme}} is
applied in sequence via its respective \code{plot_add_scheme()} method.
}
\keyword{internal}
