% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-align.R
\name{plot_align}
\alias{plot_align}
\title{Align Specifications in the Layout}
\usage{
plot_align(guides = NA, free_spaces = NA, free_labs = NA)
}
\arguments{
\item{guides}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicating
which side of guide legends should be collected. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent layout, all guides will be collected. If \code{NULL}, no guides will be
collected.}

\item{free_spaces}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}
indicating which border spaces should be removed. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent, the default is \code{NULL}, meaning no spaces are removed.

Usually you want to apply this with the whole layout, instead of individual
plots.}

\item{free_labs}{A string with one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}
indicating which axis titles should be free from alignment. Defaults to
\code{\link[ggplot2:waiver]{waiver()}}, which inherits from the parent layout. If no
parent layout, no axis titles will be aligned. If \code{NULL}, all axis titles
will be aligned.}
}
\value{
A \code{plot_align} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} The \code{plot_align()} function defines the
behavior of plots within a layout.  It can be used in the \code{action} argument
of layout functions like \code{\link[=quad_switch]{quad_switch()}} or \code{\link[=stack_switch]{stack_switch()}} to set
global actions for all plots in the layout. Additionally, \code{plot_align()} can
be added directly to a plot.
}
\examples{
# used in the layout, define the default action for all plots in the layout
ggheatmap(matrix(rnorm(72), nrow = 8)) -
    theme(plot.background = element_rect(fill = "red"))

# You can also add it for a single plot
ggheatmap(matrix(rnorm(72), nrow = 8)) -
    theme(plot.background = element_rect(fill = "red")) +
    # here, we modify the plot action for the heatmap body
    theme(plot.background = element_rect(fill = "blue"))

}
