% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-.R
\name{show,Layout-method}
\alias{show,Layout-method}
\title{Print Layout object}
\usage{
\S4method{show}{Layout}(object)
}
\arguments{
\item{object}{A \code{\link[=layout_heatmap]{layout_heatmap()}} or \code{\link[=layout_stack]{layout_stack()}} object.}
}
\description{
Print Layout object
}
\keyword{internal}
