% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{stack_active}
\alias{stack_active}
\title{Determine the active context of stack layout}
\usage{
stack_active(
  guides = NA,
  free_spaces = NA,
  plot_data = NA,
  theme = NA,
  free_labs = NA,
  what = NULL,
  sizes = NULL
)
}
\arguments{
\item{guides}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which guide should be collected. If \code{NULL}, no
guides will be collected. Default: \code{"tlbr"}.}

\item{free_spaces}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which border spaces should be removed. If
\code{NULL} (default), no space will be removed.}

\item{plot_data}{A function used to transform the plot data before rendering.
By default, it'll inherit from the parent layout. If no parent layout, the
default is \code{NULL}, which means we won't modify anything.

Used to modify the data after layout has been created, which should be a data
frame, but before the data is handled of to the ggplot2 for rendering. Use
this hook if the you needs change the default data for all \code{geoms}.}

\item{theme}{Default layout theme. One of:
\itemize{
\item \link[ggplot2:waiver]{waiver()}: will inherit from the parent layout.
\item \code{NULL}: Use the default theme associated with the package.
\item \link[ggplot2:theme]{theme()}: will be added with the parent layout theme.
}}

\item{free_labs}{A boolean value or a string containing one or more of
\code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"} indicates which axis title should be free from
alignment. If \code{NULL}, all axis title will be aligned. Default: \code{"tlbr"}.}

\item{what}{What should get activated for the stack layout?
Possible values are follows:
\itemize{
\item A single number or string of the plot elements in the stack layout.
\item \code{NULL}: remove any active context, this is useful when the active context is a
\code{\link[=heatmap_layout]{heatmap_layout()}} object, where any \verb{align_*()} will be added into the
heatmap. By removing the active context, we can add \verb{align_*()} into the
\code{\link[=stack_layout]{stack_layout()}}.
}}

\item{sizes}{A numeric or \link[grid:unit]{unit} object of length \code{3} indicates
the relative widths (\code{direction = "horizontal"}) / heights (\code{direction = "vertical"}).}
}
\value{
A \code{stack_active} object which can be added into
\link[=stack_layout]{StackLayout}.
}
\description{
Determine the active context of stack layout
}
\examples{
ggstack(matrix(1:9, nrow = 3L)) +
    ggheatmap() +
    # ggheamtap will set the active context, directing following addition
    # into the heatmap plot area. To remove the heatmap active context,
    # we can use `stack_active()` which will direct subsequent addition into
    # the stack
    stack_active() +
    # here we add a dendrogram to the stack.
    align_dendro()
}
