% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_discrete}
\alias{stack_discrete}
\alias{stack_align}
\alias{stack_discretev}
\alias{stack_alignv}
\alias{stack_discreteh}
\alias{stack_alignh}
\title{Arrange Plots Horizontally or Vertically by aligning discrete axis}
\usage{
stack_discrete(direction, data = NULL, ..., theme = NULL, sizes = NA)

stack_align(direction, data = NULL, ..., theme = NULL, sizes = NA)

stack_discretev(data = NULL, ...)

stack_alignv(data = NULL, ...)

stack_discreteh(data = NULL, ...)

stack_alignh(data = NULL, ...)
}
\arguments{
\item{direction}{A string indicating the direction of the stack layout,
either \code{"h"}(\code{horizontal}) or \code{"v"}(\code{vertical}).}

\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to
convert data to a matrix.}

\item{...}{Additional arguments passed to \code{\link[=fortify_matrix]{fortify_matrix()}}.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} object used to customize various
elements of the layout, including \code{guides}, \code{title}, \code{subtitle}, \code{caption},
\code{margins}, \code{panel.border}, and \code{background}. By default, the theme will
inherit from the parent \code{layout}. It also controls the panel spacing for all
plots in the layout.}

\item{sizes}{A numeric value or a \code{\link[grid:unit]{unit}} object. When used for
the \code{\link[=quad_layout]{quad_layout()}} annotation, it must be of length \code{1}. When used in the
\code{\link[=stack_layout]{stack_layout()}} with a nested \code{\link[=quad_layout]{quad_layout()}}, it should be of length
\code{3}, specifying the relative heights (for \code{direction = "h"}) or widths (for
\code{direction = "v"}) to be applied to the layout.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{stack_discrete} function arranges plots by aligning discrete variables.
\code{stack_align} is an alias for \code{stack_discrete} for historical reasons.

Several aliases are provided for convenience:
\itemize{
\item \code{stack_discretev} and \code{stack_alignv}: A special case of \code{stack_discrete}
that sets \code{direction = "v"}.
\item \code{stack_discreteh} and \code{stack_alignh}: A special case of \code{stack_discrete}
that sets \code{direction = "h"}.
}
}
\examples{
set.seed(123)
stack_discrete("h", matrix(rnorm(56), nrow = 7L)) +
    align_dendro()
}
