% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_layout}
\alias{stack_layout}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
stack_layout(
  data,
  direction = NULL,
  ...,
  sizes = NA,
  action = NULL,
  theme = NULL
)

ggstack(data, direction = NULL, ..., sizes = NA, action = NULL, theme = NULL)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{...}{These dots are for future extensions and must be empty.}

\item{sizes}{A numeric or a \link[grid:unit]{unit} object of length \code{3} indicates
the relative widths (\code{direction = "horizontal"}) / heights (\code{direction = "vertical"}).}

\item{action}{A \code{\link[=plot_action]{plot_action()}} object used to define the default plot action
in the layout.}

\item{theme}{A \code{\link[ggplot2:theme]{theme()}} used to render the \code{guides},
\code{title}, \code{subtitle}, \code{caption}, \code{margins}, \code{patch.title}, \code{panel.border}, and
\code{background}. If \code{NULL} (default), will inherit from the parent \code{layout}.}
}
\value{
A \code{StackLayout} object.
}
\description{
\code{ggstack} is an alias of \code{stack_layout}.
}
\examples{
ggstack(matrix(rnorm(100L), nrow = 10L)) + align_dendro()
}
