% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_layout}
\alias{stack_layout}
\alias{ggstack}
\title{Put plots horizontally or vertically}
\usage{
stack_layout(data, direction = NULL, ...)

ggstack(data, direction = NULL, ...)
}
\arguments{
\item{data}{A numeric or character vector, a data frame, or a matrix.}

\item{direction}{A string of \code{"horizontal"} or \code{"vertical"}, indicates the
direction of the stack layout.}

\item{...}{Not used currently.}
}
\value{
A \code{StackLayout} object.
}
\description{
\code{ggstack} is an alias of \code{stack_layout}.
}
\examples{
ggstack(matrix(rnorm(100L), nrow = 10L)) + align_dendro()
}
