% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-stack-.R
\name{stack_layout}
\alias{stack_layout}
\alias{stack_horizontal}
\alias{stack_vertical}
\title{Arrange plots horizontally or vertically}
\usage{
stack_layout(direction, data = NULL, limits = waiver(), ...)

stack_horizontal(data = NULL, limits = waiver(), ...)

stack_vertical(data = NULL, limits = waiver(), ...)
}
\arguments{
\item{direction}{A string indicating the direction of the stack layout,
either \code{"h"}(\code{horizontal}) or \code{"v"}(\code{vertical}).}

\item{data}{Default dataset to use for the layout. If not specified, it must be supplied in each plot added to the layout:
\itemize{
\item If \code{limits} is not provided, \code{\link[=fortify_matrix]{fortify_matrix()}} will be used to get a
matrix.
\item If \code{limits} is specified, \code{\link[=fortify_data_frame]{fortify_data_frame()}} will be used to get a
data frame. Note that if the data is a \code{matrix}, it will be automatically
converted to a long-formatted data frame, which differs from \code{ggplot2}'s
behavior.
}}

\item{limits}{A \code{\link[=continuous_limits]{continuous_limits()}} object specifying the left/lower
limit and the right/upper limit of the scale. Used to align the continuous
axis.}

\item{...}{Additional arguments passed to \code{\link[=fortify_data_frame]{fortify_data_frame()}}.}
}
\value{
A \code{StackLayout} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function integrates the functionalities of \code{stack_discrete()} and
\code{stack_continuous()} into a single interface.
}
\examples{
set.seed(123)
small_mat <- matrix(rnorm(56), nrow = 7L)

stack_horizontal(small_mat) + align_dendro()

# this is the same with:
stack_discreteh(small_mat) + align_dendro()

# For vertical layout:
stack_vertical(small_mat) + align_dendro()

}
\seealso{
\itemize{
\item \code{\link[=stack_discrete]{stack_discrete()}}
\item \code{\link[=stack_continuous]{stack_continuous()}}
}
}
