% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-matrix-matrix.R
\name{tune.matrix}
\alias{tune.matrix}
\title{Convert the shape of a matrix for fortify method}
\usage{
\method{tune}{matrix}(data, shape)
}
\arguments{
\item{data}{A matrix.}

\item{shape}{A string of \code{"upset"} or \code{"oncoplot"}.}
}
\description{
Convert the shape of a matrix for fortify method
}
\seealso{
\itemize{
\item \code{\link[=fortify_matrix.matrix]{fortify_matrix.matrix()}}
\item \code{\link[=fortify_matrix.matrix_upset]{fortify_matrix.matrix_upset()}}
\item \code{\link[=fortify_matrix.matrix_oncoplot]{fortify_matrix.matrix_oncoplot()}}
}

Other tune methods: 
\code{\link{tune.MAF}()},
\code{\link{tune.list}()}
}
\concept{tune methods}
