% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-quad-operator.R
\name{with_position}
\alias{with_position}
\title{Modify \code{-} Operator Context in \code{quad_layout()}}
\usage{
with_position(x, position = waiver())
}
\arguments{
\item{x}{Objects to add to the layout using the \code{-} operator.}

\item{position}{A string containing one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}
indicating which annotation stack should be used as the context. If \code{NULL},
the context will default to the \code{\link[=quad_layout]{quad_layout()}} itself. By default,
\code{waiver()} is used, which triggers the following behavior: If the current
active context in \code{\link[=quad_layout]{quad_layout()}} is \code{top} or \code{bottom}, the operator will
also act on the corresponding \code{bottom} or \code{top} annotation. If the context is
\code{left} or \code{right}, the operator will also act on the \code{right} or \code{left}
annotation, respectively. If there is no active annotation stack, it defaults
to \code{NULL}.}
}
\value{
The input object with an additional attribute that specifies the
selected context.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} By default, the \code{-} operator acts on all
plots at the same position as the current active annotation in
\code{\link[=quad_layout]{quad_layout()}}. If there is no active annotation stack, it will act on all
plots in the \code{\link[=quad_layout]{quad_layout()}}. Wrapping objects with \code{with_position} allows
you to change the context of the subtraction \code{-} operator.
}
