% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-quad-operator.R
\name{with_quad}
\alias{with_quad}
\title{Modify \code{-} Operator Context in \code{quad_layout()}}
\usage{
with_quad(x, position = waiver(), main = NULL)
}
\arguments{
\item{x}{Objects to add to the layout using the \code{-} operator.}

\item{position}{A string containing one or more of \code{"t"}, \code{"l"}, \code{"b"}, and \code{"r"}
indicating which annotation stack should be used as the context. If \code{NULL},
the context will default to the \code{\link[=quad_layout]{quad_layout()}} itself. By default,
\code{waiver()} is used, which triggers the following behavior: If the current
active context in \code{\link[=quad_layout]{quad_layout()}} is \code{top} or \code{bottom}, the operator will
also act on the corresponding \code{bottom} or \code{top} annotation. If the context is
\code{left} or \code{right}, the operator will also act on the \code{right} or \code{left}
annotation, respectively. If there is no active annotation stack, it defaults
to \code{NULL}.}

\item{main}{A single boolean value indicates whether to apply \code{x} for the
main plot.}
}
\value{
The input object with an additional attribute that specifies the
selected context.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Wrapping objects with \code{with_quad}
allows you to change the context of the subtraction \code{-} operator in
\code{quad_layout()}.
}
